% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_completion_status.R
\name{add_completion_status}
\alias{add_completion_status}
\title{Determine completion status for every student}
\usage{
add_completion_status(dframe, midfield_degree = degree)
}
\arguments{
\item{dframe}{Working data frame of student-level records to which
completion-status columns are to be added. Required variables are
\code{mcid} and \code{timely_term}.}

\item{midfield_degree}{MIDFIELD \code{degree} data table or equivalent with
required variables \code{mcid} and \code{term_degree.}}
}
\value{
A data frame of the same type as \code{dframe}. The output has the
following properties:
\itemize{
\item Rows are not modified.
\item Columns are added, overwriting existing columns (if any) of the same name.
Other columns are not modified.
\item Groups are not preserved.
\item Data frame attributes are preserved for classes \code{data.frame}, \code{data.table},
or \code{tbl_df}.
}
}
\description{
Add columns to a data frame of student-level records that indicate whether a
student completed a degree, and if so, whether their completion was timely.
By "completion" we mean an undergraduate earning their first baccalaureate
degree (or degrees, for students earning more than one degree in the same
term). The term by which a student's completion would be considered timely
(the "timely completion term") is usually defined as 4-, 6-, or 8-years
after admission. Our default is 6 years.
}
\details{
The new columns are:
\itemize{
\item \code{term_degree} Character. Term in which the first degree(s) are
completed, encoded \code{YYYYT}. Joined from \code{midfield_degree}.
\item \code{completion_status} Character. Possible values are "timely","late",
and "NA". Completion status is "timely" for students completing a
degree no later than their timely completion terms; "late" for students
completing their program after their timely completion term; and "NA"
for non-completers.
}
}
\examples{
# Start with an excerpt from the student data set 
dframe <- toy_student[1:10, .(mcid)]

# Timely term column is required to add completion status column
dframe <- add_timely_term(dframe, toy_term)

# Add completion status column
add_completion_status(dframe, toy_degree)

# Existing completion_status column, if any, is overwritten
dframe[, completion_status := NA_character_][]
add_completion_status(dframe, toy_degree)
}
\seealso{
Other add_*:
\code{\link[=add_data_sufficiency]{add_data_sufficiency()}},
\code{\link[=add_timely_term]{add_timely_term()}}
}
\concept{add_*}
