% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_completion_timely.R
\name{add_completion_timely}
\alias{add_completion_timely}
\title{Add a column to evaluate timely completion}
\usage{
add_completion_timely(dframe, ..., record = NULL, details = NULL)
}
\arguments{
\item{dframe}{data frame}

\item{...}{not used, forces later arguments to be used by name}

\item{record}{data frame of degree attributes, default midfielddegrees}

\item{details}{logical scalar to add columns reporting information on
which the evaluation is based, default FALSE}
}
\value{
Data frame with the following properties:
\itemize{
    \item Rows are not modified
    \item Column \code{completion_timely} is added, columns
     \code{completion} and \code{term_degree} are added optionally
    \item Data frame attributes \code{tbl} or \code{data.table}
          are preserved
    \item Grouping structures are not preserved
}
}
\description{
A logical variable is added to a data frame indicating whether a student
has completed their program in a timely manner.
}
\details{
Program completion is typically considered timely if it occurs within a
given span of years after matriculation. In a persistence metric involving
graduation, students with timely completion should be grouped as graduates.
Students whose completion is not timely should be grouped with the
non-graduates.

The data frame argument must include the \code{term_timely} column
obtained using the \code{add_term_timely()} function. Completion is
considered timely if: 1) the student has completed a program; and 2) the
degree term is no later than the timely completion limit.

If \code{details} is TRUE, additional column(s) that support the finding
are returned as well. Here the extra columns are \code{completion}
indicating by TRUE/FALSE if the student completed their program and
\code{term_degree} from the degree attributes record giving the term in
which the first degree(s), if any, was earned.
}
\examples{
# TBD
}
