% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grad_column.R
\name{add_grad_column}
\alias{add_grad_column}
\title{Add column to classify graduation rate status}
\usage{
add_grad_column(
  starters,
  ...,
  span = NULL,
  cip_level = NULL,
  data_students = NULL,
  data_terms = NULL,
  data_degrees = NULL
)
}
\arguments{
\item{starters}{data frame of students starting in programs}

\item{...}{not used, force later arguments to be used by name.}

\item{span}{numeric scalar, number of years to define successful
graduation, default is 6 years}

\item{cip_level}{numeric scalar (2, 4, or 6), the CIP level used to
determine if a student completes the program in which they matriculate,
default is 2}

\item{data_students}{data frame of student attributes}

\item{data_terms}{data frame of term attributes}

\item{data_degrees}{data frame of degree attributes}
}
\value{
data frame with the following properties:
\itemize{
    \item Rows subset to satisfy IPEDS definition
    \item Columns \code{id}, \code{start}, and \code{grad_status}
    \item Grouping structures are not preserved
    \item Data frame extensions such as \code{tbl} or \code{data.table}
    are preserved
}
}
\description{
Add a TRUE/FALSE column to a data frame of student IDs and starting
programs. TRUE indicates that a student can be counted as a program graduate
when computing graduation rates.
}
\details{
For computing graduation rate, students are counted as graduates only if
they complete the same program in which they matriculate in 6 years or less.
Starting and ending programs are compared using the 2-digit CIP.

Optional arguments allow editing the completion span, for example, from 6
years to 4 years, and the CIP level defining "completing the same program",
for example, from the 2-digit CIP to a 4-digit CIP.

The \code{starters} input must have columns \code{id} and \code{cip6}.

The function accesses these midfielddata data sets:
\code{\link[midfielddata]{midfieldstudents}},
\code{\link[midfielddata]{midfieldterms}}, and
\code{\link[midfielddata]{midfielddegrees}} or their equivalents.
}
\examples{
# placeholder
}
\seealso{
Other functions: 
\code{\link{filter_text}()},
\code{\link{prepare_fye_mi}()},
\code{\link{prepare_multiway}()},
\code{\link{subset_feasible}()}
}
\concept{functions}
