% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_race_sex.R
\name{add_race_sex}
\alias{add_race_sex}
\title{Add columns for student race/ethnicity and sex}
\usage{
add_race_sex(dframe, midfield_student = student)
}
\arguments{
\item{dframe}{Data frame of student unit record (SUR) observations keyed 
by student ID. Required variable is \code{mcid}.}

\item{midfield_student}{Data frame of SUR student observations keyed 
by student ID. Default is \code{student}. Required variables are 
\code{mcid}, \code{race}, and \code{sex}.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
 \item Rows are not modified.
 \item Grouping structures are not preserved.
 \item Columns listed below are added. \strong{Caution!} An existing column 
 with the same name as one of the added columns is silently overwritten. 
 Other columns are not modified. 
}
Columns added:
\describe{
 \item{\code{race}}{Character, self-reported race/ethnicity, e.g., 
 Asian, Black, Hispanic/LatinX, etc.}
 \item{\code{sex}}{Character, self-reported sex, values are Female, Male, 
 and Unknown}
}
}
\description{
Add columns to a data frame of Student Unit Record (SUR) 
observations that labels each row with a student's race/ethnicity and sex. 
Requires a MIDFIELD \code{student} data frame in the environment.
}
\details{
MIDFIELD student data includes the variables \code{race} (race/ethnicity) 
and \code{sex} as self-reported to their institution.
}
\examples{
# Start with an excerpt from the degree data set 
dframe <- toy_degree[1:5, c("mcid", "cip6")]

# Add columns for race/ethnicity and sex
add_race_sex(dframe, midfield_student = toy_student)

# Repeat with an example from the term data set 
dframe <- toy_term[21:26, c("mcid", "institution", "level")]
add_race_sex(dframe, midfield_student = toy_student)

# Existing race and sex columns, if any, are overwritten
dframe[, c("race", "sex") := NA_character_]
add_race_sex(dframe, midfield_student = toy_student)
}
\seealso{
Other add_*: 
\code{\link{add_completion_status}()},
\code{\link{add_data_sufficiency}()},
\code{\link{add_timely_term}()}
}
\concept{add_*}
