% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_timely_completion.R
\name{add_timely_completion}
\alias{add_timely_completion}
\title{Determine program completion and timeliness for every student}
\usage{
add_timely_completion(dframe, midfield_degree = degree, ..., details = NULL)
}
\arguments{
\item{dframe}{Data frame of student unit record (SUR) observations keyed 
by student ID. Required variables are \code{mcid} and 
\code{timely_term}.}

\item{midfield_degree}{MIDFIELD degree data frame keyed by student ID.
Default is \code{degree}. Required variables are \code{mcid} 
and \code{term_degree}.}

\item{...}{Not used, forces later arguments to be used by name.}

\item{details}{TRUE returns the additional variables used in
determining the results. Default is FALSE.}
}
\value{
A \code{data.table} with the following properties:
\itemize{
    \item Rows are not modified.
    \item Columns \code{cip6_degree} and code{timely_completion} are
    added; additional columns are added via the \code{details} argument. 
    \item Grouping structures are not preserved.
}
}
\description{
Add a column of true/false values to a data frame of Student Unit Record 
(SUR) observations indicating whether a student completes their program 
in a timely manner. A column of CIP codes of their degree programs 
is returned as well.
}
\details{
Program completion is typically considered timely if it occurs within a
specified span of years after admission (default 6 years). In many studies, 
only students whose program completion is timely are counted as graduates; 
students whose program completion is untimely (taking longer than the 
specific span to complete) are counted as non-graduates.

Completion is considered timely if: 1) the student has completed a 
program; and 2) the degree term is no later than the estimated timely 
completion term.

The optional \code{details} argument returns the additional variables used 
in determining the results: whether a student completed their program and 
the term in which their first degree(s), if any, are earned.
}
\section{Caution}{

Existing columns with the same names as the added columns are silently 
overwritten.
}

\examples{
# Start with IDs and add institution and timely term
# dframe <- toy_student[1:10, .(mcid)]
# dframe <- add_institution(dframe, midfield_term = toy_term)
# dframe <- add_timely_term(dframe, midfield_term = toy_term)
# 
# 
# # Timely completion column without detail
# add_completion_timely(dframe, midfield_degree = toy_degree)
# 
# 
# # Timely completion column with detail
# add_completion_timely(dframe, midfield_degree = toy_degree, detail = TRUE)
# 
# 
# # If present, existing completion_timely column is overwritten
# DT1 <- add_completion_timely(dframe, midfield_degree = toy_degree)
# DT2 <- add_completion_timely(DT1, midfield_degree = toy_degree)
# all.equal(DT1, DT2)

}
\seealso{
Other add_*: 
\code{\link{add_data_sufficiency}()},
\code{\link{add_institution}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()}
}
\concept{add_*}
