% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_timely_limit.R
\name{add_timely_limit}
\alias{add_timely_limit}
\title{Add a column for the timely completion limit term}
\usage{
add_timely_limit(
  dframe,
  ...,
  span = NULL,
  record = NULL,
  details = NULL,
  heuristic = NULL
)
}
\arguments{
\item{dframe}{data frame}

\item{...}{not used, forces later arguments to be used by name}

\item{span}{numeric scalar, number of years to define timely completion,
default 6 years}

\item{record}{data frame of term attributes, default midfieldterms}

\item{details}{logical scalar to add columns reporting information on
which the timely completion limit is based, default FALSE}

\item{heuristic}{not used, placeholder for future alternative methods of
estimating them timely completion term}
}
\value{
Data frame with the following properties:
\itemize{
    \item Rows are not modified
    \item Column \code{timely_limit} is added, columns \code{term_i},
          \code{level_i}, and \code{adj_span} are added optionally
    \item Data frame attributes \code{tbl} or \code{data.table}
          are preserved
    \item Grouping structures are not preserved
}
}
\description{
Given information about a student's academic path, determine the latest
term for which program completion could be considered timely.
}
\details{
The current model for timely completion is simple, though future
heuristics may be added for more sophisticated estimates. The current
estimate starts with \code{span} number of years for each student
(default 6 years) and adjusts the span by subtracting a whole number of
years based on the level at which the student matriculates.

For example, a 4th-year or senior level matriculant is assumed to have
completed 3 years of a program, so their span is reduced by 3 years.
Similarly, a 3rd-year or junior-level matriculant is assumed to have
completed 2 years of a program, so their span is reduced by 2 years.

The adjusted span is added to their starting term; the result is the
limiting term for timely completion reported in a \code{timely_limit}
column added to the data frame.

If \code{details} is TRUE, additional column(s) that support the finding
are returned as well. Here the extra columns are the student's initial
(matriculation) term \code{term_i}, initial level \code{level_i}, and
the adjusted span \code{adj_span}.
}
\examples{
# TBD

}
\seealso{
Other functions: 
\code{\link{add_race_sex}()},
\code{\link{eval_fair_record}()},
\code{\link{fair_record_length}()},
\code{\link{filter_text}()},
\code{\link{prepare_fye_mi}()},
\code{\link{prepare_multiway}()},
\code{\link{subset_feasible}()},
\code{\link{subset_matriculant}()}
}
\concept{functions}
