% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio_sci}
\alias{bio_sci}
\title{CIP codes for programs in biological and biomedical science}
\format{A character vector.}
\source{
NSF STEM programs: \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}.
}
\usage{
bio_sci
}
\description{
All 6-digit codes from the 2010 CIP 26 series.
}
\examples{
\dontrun{
 bio_sci
}
}
\keyword{datasets}
