% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_blocs}
\alias{case_blocs}
\title{Case-study outcome by ID and grouping variables}
\format{
\code{data.table} with 8868 rows and 4 columns:
\describe{
\item{\code{mcid}}{Character. Anonymized student identifier
that connects the four data tables, e.g., "MCID3111142897."}
\item{\code{people}}{Character. Merged values of race/ethnicity and sex.}
\item{\code{program}}{Character. Abbreviation of the case-study program name.}
\item{\code{bloc}}{Character. Distinguish between program graduates and those
ever enrolled in the programs.}
}
}
\usage{
case_blocs
}
\description{
Data table of student IDs, program, race/ethnicity, sex, and bloc
(graduates or ever-enrolled)  of the case study (Civil, Electrical,
Industrial/Systems, and Mechanical Engineering) just prior to grouping and
summarizing to develop the stickiness metric.
}
\seealso{
Other case-study-data:
\code{\link{case_results}},
\code{\link{population_baseline}}
}
\concept{case-study-data}
\keyword{datasets}
