% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_fye}
\alias{case_fye}
\title{Imputed starting majors for FYE students}
\format{A data frame with columns:
\describe{
 \item{id}{Unique, anonymized MIDFIELD student identifier.}
 \item{start}{Imputed starting major (6-digit CIP code) of FYE students. }
}}
\usage{
case_fye
}
\description{
Some US institutions have first year engineering (FYE) programs---typically a common first year curriculum that is a prerequisite for declaring an engineering major. For some persistence metrics, we have to determine what we call a "starting major" for FYE students---the predicted engineering program such as Civil, Electrical, or Mechanical, the student would have declared had they not been required to enroll in FYE.
}
\examples{
case_stickiness
}
\keyword{datasets}
