% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_gradrate.R
\docType{data}
\name{case_gradrate}
\alias{case_gradrate}
\title{Graduation rate example results}
\format{A data frame (tibble) with 40 observations and 6 variables.

\describe{

  \item{ethnicity}{Asian, Black, Hispanic, or White}

  \item{sex}{Female or Male}

  \item{program}{Five engineering programs}

  \item{start}{The number of students starting in a program}

  \item{grad}{The number of starters graduating in the program}

  \item{rate}{The percentage of students starting in a program that graduate in the program}

}}
\source{
\url{https://engineering.purdue.edu/MIDFIELD}
}
\usage{
case_gradrate
}
\description{
A data frame of graduation rates disaggregated by race or ethnicity, sex, and academic program. Program graduation rate is the percentage of students starting in a program that graduate in the program.
}
\keyword{datasets}
