% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_program_group}
\alias{case_program_group}
\title{CIP case data for four programs}
\format{
A data frame (tibble) with 12 observations and 3 variables.
All variables are characters. An observation is a unique program.

\describe{
  \item{cip6}{An instructional program's 6-digit code, representing
  "specific instructional programs".}
  \item{cip6name}{Name of a program at the 6-digit level}
  \item{program}{A program name used for grouping, summarizing, and graphing.}
}
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata} package
}
\usage{
case_program_group
}
\description{
A subset of the [\code{cip}](cip.html) data frame for four
engineering programs.
}
\details{
The 6-digit CIP codes and names plus a user-selected program name for Civil,
Electrical, Industrial, and Mechanical Engineering. The "Select and label
programs" vignette describes how these data were assembled. The data frame is
used by several other vignettes as a starting point.
}
\keyword{datasets}
