% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_results}
\alias{case_results}
\title{Case-study results}
\format{
\code{data.table} with 43 rows and 4 columns:
\describe{
\item{\code{program}}{Character. Academic program label.}
\item{\code{people}}{Character. Race/ethnicity and sex as
self-reported by the student, e.g., "Asian Male", "Black Female", etc.}
\item{\code{ever}}{Numerical. The number of students
ever enrolled in a program.}
\item{\code{grad}}{Numerical. Number of students completing
a program.}
\item{\code{stick}}{Numerical. Program stickiness, the
ratio of the number of graduates to the number ever enrolled, in
percent.}
}
}
\usage{
case_results
}
\description{
Data table of longitudinal stickiness for the four programs of the case study
(Civil, Electrical, Industrial/Systems, and Mechanical Engineering) grouped
by program, race/ethnicity, and sex. Provided for the convenience of vignette
users.
}
\details{
Longitudinal stickiness is the ratio of the number of students graduating
from a program to the number of students ever enrolled in the program over
the time span of available data. Results are based on data that have been
filtered for data sufficiency, degree seeking, undergraduate terms, and
timely completion.
}
\seealso{
Other case-study-data:
\code{\link{case_blocs}},
\code{\link{population_baseline}}
}
\concept{case-study-data}
\keyword{datasets}
