% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_stickiness.R
\docType{data}
\name{case_stickiness}
\alias{case_stickiness}
\title{Program stickiness example results}
\format{A data frame (tibble) with 40 observations and 6 variables.

\describe{

  \item{ethnicity}{Asian, Black, Hispanic, or White}

  \item{sex}{Female or Male}

  \item{program}{Five engineering programs}

  \item{ever}{The number of students ever enrolled in a program}

  \item{grad}{The number of students graduating in the program}

  \item{stickiness}{The fraction of students ever enrolled in a program that  graduate in the program}

}}
\source{
\url{https://engineering.purdue.edu/MIDFIELD}
}
\usage{
case_stickiness
}
\description{
A data frame of graduation rates disaggregated by race or ethnicity, sex, and academic program. Program stickiness is the percentage of students ever enrolled in a program that graduate in the program.
}
\keyword{datasets}
