% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catch_error}
\alias{catch_error}
\title{Error handling}
\usage{
catch_error(f)
}
\arguments{
\item{f}{Function with arguments expecting an error}
}
\value{
Does not return anything. The side effect is to output to the terminal.
}
\description{
A wrapper on \code{base::tryCatch()} for previewing an error message, if any.
}
\examples{
# setup
selected_ids <- toy_student[14:18, (mcid)]
s <- toy_student[mcid \%chin\% selected_ids, .(mcid, sex)]
t <- toy_term[mcid \%chin\% selected_ids, .(mcid, term)]
d <- toy_degree[mcid \%chin\% selected_ids, .(mcid, term_degree)]

# No error
catch_error(pre_or_post_bacc(dframe = t, midf_table = d))

# Error, no term variable in dframe
catch_error(pre_or_post_bacc(dframe = s, midf_table = d))

# Error, missing dframe argument
catch_error(pre_or_post_bacc(midf_table = d))

# Error, missing degree value in environment
catch_error(pre_or_post_bacc(dframe = t))

}
