% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip6_select.R
\name{cip6_select}
\alias{cip6_select}
\title{Select 6-digit CIPs and add program labels}
\usage{
cip6_select(data = NULL, program = NULL)
}
\arguments{
\item{data}{Data frame or tibble with character variables for CIP codes
and program names, typically a filtered subset of \code{cip}.}

\item{program}{Scalar character, that is, of length 1. There are
three options:

If program is NULL, the default, the 4-digit CIP program names
fill the new column.

If program is "cip2name", "cip4name", or "cip6name", the
2-digit, 4-digit, or 6-digit CIP program name fills the new column.

If program is any other string, that string fills the new column.}
}
\value{
A data frame with 6-digit CIP codes and names and the
assigned program label.
}
\description{
From a CIP data frame, select the columns with the 6-digit codes and names
and add a new character variable named \code{program}.
}
\details{
The 2-digit and 4-digit codes and names are silently dropped.

Assigning a custom label to a program or a group of programs provides
the option of grouping, summarizing, and joining data in ways not afforded
by the default program names in CIP data. We can also create program names
that are less verbose than the default names.

The variable names and class in \code{data} must match the variable names
and class in the \code{cip} data set.
}
\examples{
ece <- cip_filter(cip, keep_any = "^1410")
cip6_select(ece, program = "ECE")
cip6_select(ece, program = "cip2name")
cip6_select(ece, program = "cip6name")
ece \%>\% cip6_select(program = "ECE")
ece \%>\% cip6_select()

}
\seealso{
\code{\link[midfieldr]{cip_filter}} for exploring CIP codes.

Other data_carpentry: 
\code{\link{cip_filter}()},
\code{\link{ever_filter}()},
\code{\link{grad_filter}()},
\code{\link{multiway_order}()},
\code{\link{race_sex_join}()},
\code{\link{start_filter}()},
\code{\link{term_summarize}()}
}
\concept{data_carpentry}
