% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip_after_fye.R
\name{cip_after_fye}
\alias{cip_after_fye}
\title{Determine the first post-FYE program by student ID}
\usage{
cip_after_fye(id = NULL, ..., data_students = NULL, data_terms = NULL)
}
\arguments{
\item{id}{character vector of student IDs}

\item{...}{not used for values. Forces optional arguments to be usable
by name only.}

\item{data_students}{data frame of student attributes at matriculation}

\item{data_terms}{data frame of term attributes}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows for which student ID were ever enrolled in an FYE program
  \item Columns \code{id} and \code{cip6}
  \item Grouping structures, if any, are not preserved
  \item Data frame extensions \code{tbl} or \code{data.table} are preserved
}
}
\description{
Subset a vector of student IDs to retain students ever enrolled in FYE and
return the CIP code of their first post-FYE program.
}
\details{
Computing some persistence metrics, e.g., graduation rate, is
complicated for students required to enroll in first-year engineering
(FYE) programs. In such cases, we must predict the students'
"starting majors." \code{cip_after_fye()} takes the first step in this
process by identifying the programs in which students enroll immediately
following their final FYE term.

\describe{
  \item{FYE}{First year engineering program, typically a common first
  year curriculum that is a prerequisite for declaring an
  engineering major}
  \item{starting major}{The degree-granting engineering program---for
  example, Civil Engineering, Electrical Engineering, Mechanical
  Engineering, etc.---that we predict the student would have declared had
  they not been required to enroll in FYE.}
}

The function accesses \code{data_students}
(\code{\link[midfielddata]{midfieldstudents}} or its equivalent) to
restrict the IDs to those enrolled in an FYE program at matriculation.
It also accesses \code{data_terms}
(\code{\link[midfielddata]{midfieldterms}} or its equivalent) to determine
the 6-digit CIP code of a student's first post-FYE program. For students
with no program after FYE, the CIP is NA.
}
\examples{
# placeholder

}
\seealso{
Other data_carpentry: 
\code{\link{completion_feasible}()},
\code{\link{order_multiway}()},
\code{\link{term_addition}()}
}
\concept{data_carpentry}
