% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip_label.R
\name{cip_label}
\alias{cip_label}
\title{Add a column of program labels to a CIP data frame}
\usage{
cip_label(.cip, program = NULL)
}
\arguments{
\item{.cip}{A subset of the midfieldr \code{cip} data frame.}

\item{program}{An optional character variable. There are four options:

If NULL, the default, the 6-digit CIP program names are assigned to the new column.

If NULL, but one of the named \code{cip_series} is used to create the \code{.cip} argument, then the series name is assigned to the new column.

If one of three strings ("cip2name", "cip4name", or "cip6name"), then the 2-digit, 4-digit, or 6-digit CIP program names are assigned to the new column.

Any other string is assigned to the new column.}
}
\value{
The input data frame with an added \code{program} variable.
}
\description{
Adds a new character variable named \code{program} to a CIP data frame.
}
\details{
Assigning a custom label to a program or a group of programs provides
the option of grouping and summarizing CIP data in ways not afforded by
the default program names in the 2010 CIP data.
}
\examples{
# Extract the Philosophy and Religion programs and label
x <- cip_filter(series = "^38")
y <- cip_label(x, program = "Phil-Relig")

# The function is pipe-compatible
y <- x \%>\% cip_label("test label")

# With no program argument, the 6-digit CIP names are assigned
y <- cip_label(x)

# Argument options include
y <- cip_label(x, program = "cip6name")
y <- cip_label(x, program = "cip4name")
y <- cip_label(x, program = "cip2name")
}
