% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cip_series}
\alias{cip_series}
\alias{cip_bio_sci}
\alias{cip_engr}
\alias{cip_math_stat}
\alias{cip_other_stem}
\alias{cip_phys_sci}
\alias{cip_stem}
\title{Named series of 6-digit CIP codes}
\format{A character vector of 6-digit CIP codes.
\describe{
\item{\code{cip_bio_sci}}{Biological and biomedical science programs,
  26 series, e.g., 260001.}
\item{\code{cip_engr}}{Engineering programs, series 14, e.g., 140201.}
\item{\code{cip_math_stat}}{Mathematics and statistics programs, series 27,
  e.g., 270103.}
\item{\code{cip_other_stem}}{Other STEM programs not in \code{cip_bio_sci},
  \code{cip_engr}, \code{cip_math_stat}, or \code{cip_phys_sci}.}
\item{\code{cip_phys_sci}}{Physical science programs, series 40, e.g.
  400202.}
\item{\code{cip_stem}}{All Science, Technology, Engineering, and Mathematics
  (STEM) programs.}
}}
\source{
NSF STEM programs, \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}

CIP 2010, US National Center for Education Statistics,
  \url{https://nces.ed.gov/ipeds/cipcode}.
}
\usage{
cip_bio_sci

cip_engr

cip_math_stat

cip_other_stem

cip_phys_sci

cip_stem
}
\description{
Collected Classification of Instructional Programs (CIP) 6-digit codes
for groups of programs.
}
\examples{
head(cip_math_stat, n = 10L)
head(cip_phys_sci, n = 10L)

cip_filter(series = cip_math_stat)
cip_filter(series = cip_phys_sci)
}
\keyword{datasets}
