% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion_feasible.R
\name{completion_feasible}
\alias{completion_feasible}
\title{Subset IDs for feasible completion}
\usage{
completion_feasible(
  id = NULL,
  ...,
  span = NULL,
  terms_transfer_max = NULL,
  data_students = NULL,
  data_terms = NULL,
  data_degrees = NULL
)
}
\arguments{
\item{id}{character vector of student IDs}

\item{...}{not used for values. Forces optional arguments to be usable
by name only.}

\item{span}{numeric scalar, number of years to define feasibility, default
is 6 years}

\item{terms_transfer_max}{numeric scalar, the term equivalent of the
maximum number of transfer credit hours, default is 4 terms}

\item{data_students}{data frame of student attributes}

\item{data_terms}{data frame of term attributes}

\item{data_degrees}{data frame of degree attributes}
}
\value{
character vector of student IDs, a subset of the input
}
\description{
Subset a vector of student IDs, retaining students for whom program
completion is feasible within a designated time limit.
}
\details{
To count a non-graduating student in a persistence metric, completing
their instructional program must have been feasible. Completion feasibility
depends on the range of data from the institution, the student's
matriculation term, and a designated time limit (typically 6 years).
The relevant terminology is:

\describe{
  \item{data range}{The span of terms for which student-record data
  are available. Varies by institution.}
  \item{completion feasibility}{A student may complete their program
  within the limits of the data range.}
  \item{matriculation limit}{The last term a student can matriculate
  and feasibly complete their program  within a designated time limit
  (typically 6 years). Transfer credits can advance the matriculation
  limit.}
}

The subsetting condition is that a student has graduated, or if not, that
their matriculation term is no later than the matriculation limit,
adjusted for transfer credits if any, up to a maximum of 4 terms equivalent.

The function accesses three of the midfielddata data sets:
\code{\link[midfielddata]{midfieldstudents}},
\code{\link[midfielddata]{midfieldterms}}, and
\code{\link[midfielddata]{midfielddegrees}} or their equivalents.
}
\examples{
# placeholder

}
\seealso{
Other data_carpentry: 
\code{\link{label_programs}()},
\code{\link{order_multiway}()},
\code{\link{term_addition}()}
}
\concept{data_carpentry}
