% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_fye.R
\name{condition_fye}
\alias{condition_fye}
\title{Condition FYE data for multiple imputation}
\usage{
condition_fye(dframe, midfield_term, ..., fye_codes = NULL)
}
\arguments{
\item{dframe}{Data frame of all degree-seeking engineering students in the
database, with required variables \code{mcid}, \code{race},
and \code{sex}.}

\item{midfield_term}{MIDFIELD \code{term} data table or equivalent
with required variables \code{mcid}, \code{institution},
\code{term}, and \code{cip6}.}

\item{...}{Not used, forces later arguments to be used by name.}

\item{fye_codes}{Optional character vector of 6-digit CIP codes to
identify FYE programs, default 140102. Codes must be 6-digit
strings of numbers; regular expressions are prohibited.
Non-engineering codes---those that do not start with
14"---are ignored.}
}
\value{
A \code{data.table} with the following properties:
\itemize{
    \item One row for every FYE student.
    \item Columns for ID, institution, race, sex, and CIP code, all
    except ID converted to factors. Additional columns in \code{dframe}
    are dropped.
    \item Grouping structures are not preserved.
}
}
\description{
Filter first-year-engineering (FYE) students and prepare variables for
predicting unknown starting majors. The prepared variables are institution,
race, and sex (predictors) and cip6 (with missing values to be imputed). The
function returns a data frame formatted for multiple imputation using
the mice package.
}
\details{
Some US institutions have first year engineering (FYE) programs, typically
a common first year curriculum that is a prerequisite for declaring
an engineering major. FYE programs are problematic for some persistence
metrics. For example, in the conventional graduation rate metric, the only
students who count as graduates are those who complete the same
program in which they were admitted. But students cannot graduate in FYE;
instead, upon completing FYE, they transition to a degree-granting
engineering program.

Therefore, to include FYE students in any persistence metric requiring a
degree-granting "starting" program, we have to predict the engineering
program the FYE student would have declared had they not been required to
enroll in FYE. \code{predict_fye()} sorts students into two categories:

\enumerate{
\item{Students who complete an FYE and declare an engineering major.
    This is the easy case--at the student's first opportunity, they
    enrolled in an engineering program of their choosing. We use that
    program as the predicted  starting program.}
\item{Students who, after FYE, do not declare an engineering major.
    This is the more complicated case---the data provide no information
    regarding what engineering program the student would have declared
    originally had the institution not required them to enroll in FYE.
    For these students, we treat the starting program as missing data be
    predicted using multiple imputation.}
}

The function extracts all terms for all FYE students from
\code{midfield_term}. In cases where students enter FYE, change programs,
and re-enter FYE, only the first group of FYE terms is considered. Any
programs before FYE are ignored. The first (if any) post-FYE program is
identified. If the program is in engineering, the CIP is retained as
their predicted starting major. If not, the CIP is replaced with NA to
be treated as missing data for the imputation.

Lastly, the predictor variables (institution, race, sex) and the imputed
variable (cip6) are converted to unordered factors. The resulting data
frame is ready for use as input for the mice package.
}
\examples{
# Using toy data
DT <- toy_student[, .(mcid, race, sex)]
condition_fye(dframe = DT, midfield_term = toy_term)


# Overwrites institution if present in dframe
DT <- toy_student[, .(mcid, institution, race, sex)]
condition_fye(dframe = DT, midfield_term = toy_term)


# Other columns, if any, are dropped
colnames(toy_student)
colnames(condition_fye(toy_student, toy_term))


# Optional argument permits multiple CIP codes for FYE
condition_fye(DT, toy_term, fye_codes = c("140101", "140102"))
\dontrun{
# Constraints on the \vode{fye_codes} argument
DT <- toy_student[, .(mcid, race, sex)]


# Must be an engineering CIP (starts with 14)
condition_fye(DT, toy_term, fye_codes = c("140101", "540102"))


# Must have 6-digits
condition_fye(DT, toy_term, fye_codes = "1401")


# Must contain numerals only (no regular expressions)
condition_fye(DT, toy_term, fye_codes = "^14010")
}

}
\seealso{
Other condition_*: 
\code{\link{condition_multiway}()}
}
\concept{condition_*}
