% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_fill.R
\name{count_and_fill}
\alias{count_and_fill}
\title{Group and summarize a count of students}
\usage{
count_and_fill(data, count_label = NULL)
}
\arguments{
\item{data}{A data frame of unique students}

\item{count_label}{An optional name for the count variable. Default is \code{n}.}
}
\value{
Data frame with a count of observations at the intersections of all variables in the \code{data} input data frame.
}
\description{
Counts the number of students grouped by all the variables in \code{data}. The count is assigned the variable name in \code{count_label}. If the variables \code{sex}, \code{race}, or \code{institution} are present, the function fills a zero for a missing count at every intersection of these variables.
}
