% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_traits}
\alias{data_traits}
\title{Obtain the traits of a data frame}
\usage{
data_traits(data)
}
\arguments{
\item{data}{data frame}
}
\description{
Returns a list with number of observations (rows), number of variables
(columns), and size of the file (bytes). The return object is a list named
components n_obs, n_var, n_bytes, n_inst, and year_limits.  Function is
exported to simplify use in vignettes.
}
\keyword{internal}
