% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec_midfieldr.R
\name{deprec_midfieldr}
\alias{deprec_midfieldr}
\title{Deprecated functions in midfieldr}
\description{
These functions were deprecated in midfieldr 1.0.4.
\describe{
\item{\code{select_required()}}{is deprecated in favor of \code{\link[=select_basic_cols]{select_basic_cols()}}.
The new functionality is similar but with exact matching to the default
column names plus preserving keys and data frame class.}
}
}
