% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{engr}
\alias{engr}
\title{CIP codes for engineering}
\format{A character vector.}
\source{
CIP 2010, US National Center for Education Statistics,
  \url{https://nces.ed.gov/ipeds/cipcode}.
}
\usage{
engr
}
\description{
Classification of Instructional Programs (CIP) 6-digit codes for
engineering programs, series 14.
}
\examples{
head(engr, n = 10L)

print(cip_filter(engr))
}
\keyword{datasets}
