% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fair_record.R
\name{eval_fair_record}
\alias{eval_fair_record}
\title{Add a column to evaluate record length}
\usage{
eval_fair_record(dframe, ..., record = NULL, details = NULL)
}
\arguments{
\item{dframe}{data frame}

\item{...}{not used, forces later arguments to be used by name}

\item{record}{data frame of term attributes}

\item{details}{logical scalar to add columns reporting information on
which the fair record evaluation is based, default FALSE}
}
\value{
Data frame with the following properties:
\itemize{
    \item Rows are not modified
    \item Column \code{fair_record} is added, column \code{inst_limit}
          is added optionally
    \item Data frame attributes \code{tbl} or \code{data.table}
          are preserved
    \item Grouping structures are not preserved
}
}
\description{
Given a student's estimated timely completion limit and the range of data
available from their institution, evaluate if the record is of
sufficient length to fairly assess timely completion.
}
\details{
The data frame argument must include the \code{timely_limit} column
obtained from the \code{add_timely_limit()} function.

If \code{details} is TRUE, additional column(s) that support the finding
are returned as well. Here the extra column is the latest term in the
record by institution, \code{inst_limit}.
}
\examples{
# TBD

}
\seealso{
Other functions: 
\code{\link{add_race_sex}()},
\code{\link{add_timely_limit}()},
\code{\link{fair_record_length}()},
\code{\link{filter_text}()},
\code{\link{prepare_fye_mi}()},
\code{\link{prepare_multiway}()},
\code{\link{subset_feasible}()},
\code{\link{subset_matriculant}()}
}
\concept{functions}
