% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_cip}
\alias{exa_cip}
\title{Example CIP codes and names}
\format{
A data frame (tibble) with 12 observations and 6 variables.
The variables are characters.
\describe{
  \item{cip6}{6-digit program code.}
  \item{cip6name}{6-digit program name.}
  \item{cip4}{4-digit program code.}
  \item{cip4name}{4-digit program name.}
  \item{cip2}{2-digit program code.}
  \item{cip2name}{2-digit program name.}
}
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}
 package
}
\usage{
exa_cip
}
\description{
Subset of the \code{cip} data set. CIP codes and names for Civil,
Electrical, Industrial, and Mechanical Engineering. Used in a case
study in the vignettes.
}
\examples{
exa_cip
label_programs(data = exa_cip, label = "cip4name")
}
\seealso{
Other example_data: 
\code{\link{exa_group}},
\code{\link{exa_id}},
\code{\link{exa_stickiness}}
}
\concept{example_data}
\keyword{datasets}
