% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_group}
\alias{exa_group}
\title{Example program group}
\format{
\code{data.table} with 12 rows and 2 columns keyed by the
6-digit CIP code. The variables are:
\describe{
  \item{cip6}{6-digit program code}
  \item{program}{program name assigned for grouping, summarizing,
  and joining}
}
}
\usage{
exa_group
}
\description{
Six-digit CIP codes and custom program names for a case study of Civil,
Electrical, Industrial, and Mechanical Engineering programs in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}.
}
\examples{
library(midfielddata)

# attributes of students matriculating in programs
cips_we_want <- exa_group$cip6
rows_we_want <- midfieldstudents$cip6 \%in\% cips_we_want
matriculants <- midfieldstudents[rows_we_want]
str(matriculants)

# attributes of students graduating in programs
cips_we_want <- exa_group$cip6
rows_we_want <- midfielddegrees$cip6 \%in\% cips_we_want
graduates <- midfielddegrees[rows_we_want]
str(graduates)

}
\seealso{
Other example_data: 
\code{\link{exa_cip}},
\code{\link{exa_id}},
\code{\link{exa_stickiness_mw}},
\code{\link{exa_stickiness}}
}
\concept{example_data}
\keyword{datasets}
