% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_id}
\alias{exa_id}
\alias{exa_ever}
\alias{exa_grad}
\alias{exa_ever_fc}
\title{Collections of student IDs}
\format{
character vector. The names of the available data vectors are:
\describe{
\item{exa_ever}{IDs of unique students ever enrolled in
  the example programs}
\item{exa_ever_fc}{subset of \code{exa_ever} for whom
  program completion is feasible in 6 years}
\item{exa_grad}{IDs of unique students graduating from
  the example programs}
}
}
\usage{
exa_ever

exa_grad

exa_ever_fc
}
\description{
Character vectors of student IDs at different points in a case study of
Civil, Electrical, Industrial, and Mechanical Engineering programs in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}.
}
\examples{
library(midfielddata)

# ever enrolled
rows_we_want <- midfieldstudents$id \%in\% exa_ever
enrollees <- midfieldstudents[rows_we_want]
head(enrollees)

# graduates
rows_we_want <- midfieldstudents$id \%in\% exa_grad
graduates <- midfieldstudents[rows_we_want]
head(graduates)

}
\seealso{
Other example_data: 
\code{\link{exa_cip}},
\code{\link{exa_group}},
\code{\link{exa_stickiness_mw}},
\code{\link{exa_stickiness}}
}
\concept{example_data}
\keyword{datasets}
