% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_program_group}
\alias{exa_program_group}
\title{Example program group}
\format{
A data frame (tibble) with 12 observations and 3 variables.
All variables are characters. An observation is a unique program.
\describe{
  \item{cip6}{6-digit program code.}
  \item{cip6name}{6-digit program name.}
  \item{program}{A short program name assigned for grouping,
  summarizing, and joining.}
}
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}
 package
}
\usage{
exa_program_group
}
\description{
Intermediate result in a case study.
}
\details{
The 6-digit CIP codes and names for Civil, Electrical, Industrial,
and Mechanical Engineering plus a column of custom program names
for grouping, summarizing, and joining. The data are an intermediate
result in a case study developed in the midfieldr vignettes.
}
\examples{
x <- unique(exa_program_group[["cip6"]])
y <- unique(exa_program_group[["program"]])
}
\seealso{
The "Gather program data" vignette.

Other example_data: 
\code{\link{exa_cip}},
\code{\link{exa_imputed_fye}},
\code{\link{exa_stickiness}}
}
\concept{example_data}
\keyword{datasets}
