% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_stickiness_mw}
\alias{exa_stickiness_mw}
\title{Example stickiness data in multiway form}
\format{
\code{data.table} with 32 rows and 3 columns keyed by
program, race/ethnicity, and sex. The variables are:
\describe{
 \item{program}{factor, instructional programs selected for study}
 \item{race_sex}{factor, student race/ethnicity and sex}
 \item{stick}{numerical, program stickiness, a persistence metric}
}
}
\usage{
exa_stickiness_mw
}
\description{
Stickiness metric results with two categorical variables as factors
with levels ordered by the median stickiness.
}
\examples{
# placeholder

}
\seealso{
Other example_data: 
\code{\link{exa_cip}},
\code{\link{exa_group}},
\code{\link{exa_id}},
\code{\link{exa_stickiness}}
}
\concept{example_data}
\keyword{datasets}
