% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_record_length.R
\name{fair_record_length}
\alias{fair_record_length}
\title{Subset ID vector for fair record length}
\usage{
fair_record_length(id, ..., span = NULL, record = NULL, heuristic = NULL)
}
\arguments{
\item{id}{character vector of student IDs}

\item{...}{not used, force later arguments to be used by name.}

\item{span}{numeric scalar, number of years to define fair assessment of
program completion, default is 6 years}

\item{record}{data frame of term attributes, default midfieldterms}

\item{heuristic}{not used. Argument to be used in the future for more
sophisticated algorithms to determine a student's timely completion term.}
}
\value{
character vector of student IDs, a subset of the input
}
\description{
Subset a vector of student IDs, omitting those who matriculate too near
the end of the data record for fairly assessing program completion.
}
\details{
Most persistence metrics are computed using program completion as a
criterion. To fairly assess whether or not a student has completed a
program, there must exist a sufficient number of terms in the data
record between matriculation and the institutional end-of-record.

The \code{span} argument (default 6 years) is the basis interval between
matriculation and the data end-of-record considered sufficient for
program completion assessment. The basis interval is used as-is for students
whose level in their first term is "01 Freshman" (first-year). For
students entering at higher levels, the span is reduced by one year per
level. For example, the span for an entering sophomore is reduced by one
year; entering junior, reduced by two years; entering senior, reduced by
three years.

The adjusted \code{span} defines the interval during which program
completion would be considered "timely" for that student. The last term of
this interval is the student's "timely completion term." The last term in
the term data overall, by institution, is the institution's
end-of-record term.

Program completion can be assessed fairly only for students whose timely
completion term is no later than their institution's end-of record term.
IDs of students satisfying this condition are returned by the function.
IDs not satisfying this condition are dropped.

The function accesses one of the midfielddata data sets:
\code{\link[midfielddata]{midfieldterms}}.
}
\examples{
#TBD

}
\seealso{
Other functions: 
\code{\link{add_race_sex}()},
\code{\link{add_timely_limit}()},
\code{\link{eval_fair_record}()},
\code{\link{eval_timely_completion}()},
\code{\link{filter_text}()},
\code{\link{prepare_fye_mi}()},
\code{\link{prepare_multiway}()},
\code{\link{subset_feasible}()},
\code{\link{subset_matriculant}()}
}
\concept{functions}
