% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_cip.R
\name{filter_by_cip}
\alias{filter_by_cip}
\title{Subset rows by program CIP}
\usage{
filter_by_cip(
  data,
  keep_cip,
  ...,
  keep_col = NULL,
  unique_row = NULL,
  first_degree = NULL
)
}
\arguments{
\item{data}{data frame}

\item{keep_cip}{character vector of 6-digit CIP codes}

\item{...}{not used, force later arguments to be used by name}

\item{keep_col}{character vector of column names, default all columns}

\item{unique_row}{logical, remove duplicate rows, default FALSE}

\item{first_degree}{logical, retain first degree term only, default FALSE}
}
\value{
data frame with the following properties:
\itemize{
    \item Rows with CIP matching an element of \code{keep_cip}
    \item All columns if \code{keep_col} NULL, otherwise columns named
    in \code{keep_col}
    \item Grouping structures are not preserved
    \item Data frame extensions such as \code{tbl} or \code{data.table}
    are preserved
}
}
\description{
Retain rows that match program CIPs. Default behavior is to retain all
columns. Use \code{keep_col} to subset columns. Use \code{unique_row} to
remove duplicate rows after subsetting.
}
\details{
The \code{data} argument---a midfielddata data set or equivalent---must
include a \code{cip6} column.

If applied to \code{midfielddegrees}, setting \code{first_degree} to TRUE
retains the first degree term only (by \code{id}), dropping rows with
degrees earned in subsequent terms. However, multiple degrees in the
first degree term are always retained.
}
\examples{
# placeholder
}
\seealso{
Other functions: 
\code{\link{filter_by_id}()},
\code{\link{filter_by_text}()},
\code{\link{prepare_fye_mi}()},
\code{\link{prepare_multiway}()},
\code{\link{subset_feasible}()},
\code{\link{subset_ipeds}()}
}
\concept{functions}
