% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_id.R
\name{filter_by_id}
\alias{filter_by_id}
\title{Subset rows of a data frame by student ID}
\usage{
filter_by_id(
  data,
  keep_id,
  ...,
  keep_col = NULL,
  unique_row = NULL,
  first_degree = NULL
)
}
\arguments{
\item{data}{data frame}

\item{keep_id}{character vector of student IDs}

\item{...}{not used, force later arguments to be used by name}

\item{keep_col}{character vector of column names, default all columns}

\item{unique_row}{logical, remove duplicate rows, default FALSE}

\item{first_degree}{logical, retain first degree term only, default FALSE}
}
\value{
data frame with the following properties:
\itemize{
    \item Rows with an ID matching an element of \code{keep_id}
    \item All columns if \code{keep_col} NULL, otherwise columns named
    in \code{keep_col}
    \item Grouping structures are not preserved
    \item Data frame extensions such as \code{tbl} or \code{data.table}
    are preserved
}
}
\description{
Retain rows that match student IDs. Default behavior is to retain all
columns. Use \code{keep_col} to subset columns. Use \code{unique_row} to
remove duplicate rows after subsetting.
}
\details{
The \code{data} argument---a midfielddata data set or equivalent---must
include an \code{id} column.

If applied to \code{midfielddegrees}, setting \code{first_degree} to TRUE
retains the first degree term only (by \code{id}), dropping rows with
degrees earned in subsequent terms. However, multiple degrees in the
first degree term are always retained.
}
\examples{
# placeholder
}
\seealso{
Other functions: 
\code{\link{feasible_subset}()},
\code{\link{filter_by_cip}()},
\code{\link{filter_by_text}()},
\code{\link{fye_mi_prep}()},
\code{\link{ipeds_subset}()},
\code{\link{order_multiway}()}
}
\concept{functions}
