% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_key.R
\name{filter_by_key}
\alias{filter_by_key}
\title{Subset rows by matching values in a common key column}
\usage{
filter_by_key(dframe, match_to, key_col, ..., select = NULL)
}
\arguments{
\item{dframe}{data frame to be subset and returned, column identified in
\code{key_col} required}

\item{match_to}{data frame, column identified in \code{key_col} required}

\item{key_col}{character, name of the key column}

\item{...}{not used, force later arguments to be used by name}

\item{select}{character vector of \code{dframe} column names to retain,
default all columns}
}
\value{
A \code{data.table} with the following properties:
\itemize{
    \item Unique rows with matching values
    \item All columns or those specified by \code{select}
    \item Grouping structures are not preserved
}
}
\description{
Subset rows of one data frame by values of a key variable that match values
of the same key variable in a second data frame.
}
\details{
Two data frames are input. The \code{match_to} data frame is subset to
retain its key column only.  The result is merged with \code{dframe} in an
inner-join, returning the rows of \code{dframe} with values matching the key
values in \code{match_to}.

All columns in \code{dframe} are returned unless a subset is given in the
\code{select} argument. Column subsetting occurs after the inner join, so the
key column does not have to be included in the vector of column names in
\code{select}. The final operation is to subset for unique rows.
}
\examples{
# TBD
}
\seealso{
Other functions: 
\code{\link{add_completion_timely}()},
\code{\link{add_data_sufficiency}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()},
\code{\link{condition_fye}()},
\code{\link{condition_multiway}()},
\code{\link{filter_text}()}
}
\concept{functions}
