% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_text.R
\name{filter_by_text}
\alias{filter_by_text}
\title{Subset rows of a data frame by text}
\usage{
filter_by_text(
  data,
  keep_any = NULL,
  ...,
  drop_any = NULL,
  keep_col = NULL,
  unique_row = NULL
)
}
\arguments{
\item{data}{data frame of character columns}

\item{keep_any}{character vector of search patterns for retaining rows}

\item{...}{not used, force later arguments to be used by name}

\item{drop_any}{character vector of search patterns for dropping rows}

\item{keep_col}{character vector of column names, default all columns}

\item{unique_row}{logical, remove duplicate rows, default FALSE}
}
\value{
data frame with the following properties:
\itemize{
  \item Rows are a subset of the input in the same order
  \item All columns if \code{keep_col} NULL, otherwise columns named
    in \code{keep_col}
  \item Grouping structures are not preserved
  \item Data frame extensions such as \code{tbl} or \code{data.table}
    are preserved
}
}
\description{
Retain rows that contain matching or partially matching character strings.
Not case-sensitive. Default behavior is to retain all columns.
Use \code{keep_col} to subset columns. Use \code{unique_row} to remove
duplicate rows after subsetting.
}
\details{
The \code{data} argument is typically the midfieldr \code{cip} data set,
though any data frame with all character-string columns can be searched
and subset.
}
\examples{
filter_by_text(cip, keep_any = c("^1407", "^1410"))
filter_by_text(cip, keep_any = "civil engineering", drop_any = "technology")
filter_by_text(cip, keep_any = "History") \%>\%
  filter_by_text(keep_any = "American")

}
\seealso{
Other data_query: 
\code{\link{filter_by_cip}()},
\code{\link{filter_by_id}()},
\code{\link{get_institution_limits}()}
}
\concept{data_query}
