% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_char_frame.R
\name{filter_char_frame}
\alias{filter_char_frame}
\title{Subset rows of character data frame}
\usage{
filter_char_frame(data, keep_any = NULL, ..., drop_any = NULL)
}
\arguments{
\item{data}{data frame of character variables}

\item{keep_any}{character vector of search patterns for retaining rows}

\item{...}{not used for values. Forces optional arguments to be usable
by name only.}

\item{drop_any}{(optional) character vector of search patterns for
dropping rows.}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows are a subset of the input in the same order
  \item Columns are not modified
  \item Grouping structures, if any, are not preserved
  \item Data frame extension attributes, e.g., tibble, are not preserved
}
}
\description{
Subset the rows of a data frame by matching character patterns. Rows are
subset by matching or partially matching the patterns. Used primarily
for subsetting operations by midfieldr functions such as \code{get_cip()}.
}
\details{
The \code{data} argument is any data frame comprised of character columns.
Rows are retained if they contain terms that match or partially match
patterns in \code{keep_any}; dropped if they match or partially match
patterns in \code{drop_any}.

Search patterns are not case-sensitive,  must
be supported by \code{\link[base:grep]{grepl}}, and can include
\code{\link[base:regex]{regular expressions}}.
}
\examples{
filter_char_frame(cip, keep_any = c("^1407", "^1410"))
filter_char_frame(cip,
        keep_any = "civil engineering",
        drop_any = "technology")
filter_char_frame(cip, keep_any = "History") \%>\%
  filter_char_frame(keep_any = "American")

}
\seealso{
Other data_carpentry: 
\code{\link{get_cip}()},
\code{\link{get_enrollees}()},
\code{\link{get_graduates}()},
\code{\link{get_race_sex}()},
\code{\link{label_programs}()},
\code{\link{order_multiway}()}
}
\concept{data_carpentry}
