% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cip.R
\name{filter_cip}
\alias{filter_cip}
\title{Filter CIP program data}
\usage{
filter_cip(type = NULL, series = NULL, data = NULL)
}
\arguments{
\item{type}{An integer (2, 4, or 6) representing the CIP taxonomy level}

\item{series}{The first few digits of a CIP code or vector of codes. A character vector, or coercible to one. Can also be one of the predefined vectors of CIP codes for Engineering, Humanities, Mathematics, Sciences, or the Social Sciences (\code{series_engr, series_hum, series_math, series_sci, series_socsci}).}

\item{data}{CIP data frame. The default is the midfieldr \code{cip} dataset with two variables, \code{CIP} and \code{PROGRAM}.}
}
\value{
A subset of the \code{cip} dataset filtered by type and series
}
\description{
Filter CIP program data by type and series. Type is the CIP taxonomy level: 2-digit, 4-digit, or 6-digit. Series is the CIP code.
}
\examples{
# Extract all 2-digit CIP data
filter_cip()

# Extract all 4-digit CIP data
filter_cip(type = 4)

# Extract all 4-digit CIP data in Engineering
filter_cip(type = 4, series = "14")

# Use the predefined Engineering series
filter_cip(type = 4, series = series_engr)
}
