% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cip_rows.R
\name{filter_cip_rows}
\alias{filter_cip_rows}
\title{Subset rows that include matches to search strings}
\usage{
filter_cip_rows(dframe, pattern, ..., negate = FALSE)
}
\arguments{
\item{dframe}{Data frame of CIP program codes to be searched, typically
\code{cip} that loads with midfieldr.}

\item{pattern}{Character vector of search strings for retaining rows,
not case-sensitive. Can include regular expressions.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{negate}{Logical. If true, searches for not-pattern. Default FALSE.}
}
\value{
A data frame subset of \code{dframe}. Output has the
following properties:
\itemize{
\item Rows are a subset of the input, but appear in the same order.
\item Columns are not modified.
\item Data frame attributes (except groups) are preserved.
}
}
\description{
Subset a CIP data frame, retaining rows that match or partially match a
vector of character strings.
}
\details{
Search terms can include regular expressions.
Uses \code{grepl()}, therefore non-character columns (if any) that can be
coerced to character are also searched for matches.
}
\examples{
# Subset using keywords
filter_cip_rows(cip, pattern = "history")

# Subset using codes
filter_cip_rows(cip, pattern = "^54")

# Multiple passes to narrow the results
first_pass <- filter_cip_rows(cip, "math")[, .(cip6name, cip6)]
first_pass

second_pass <- filter_cip_rows(first_pass, c("bio", "educ"), negate = TRUE)
second_pass

third_pass <- filter_cip_rows(second_pass, c("^27", "^30"))
third_pass
 
}
\concept{filter_*}
