% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_degree_seeking.R
\name{filter_degree_seeking}
\alias{filter_degree_seeking}
\title{Subset a data frame for matriculants}
\usage{
filter_degree_seeking(dframe, ..., dbase = NULL)
}
\arguments{
\item{dframe}{data frame with an ID column}

\item{...}{not used, forces later arguments to be used by name}

\item{dbase}{data frame of student attributes, default midfieldstudents}
}
\value{
Data frame with the following properties:
\itemize{
    \item Rows with matching IDs in \code{dbase}
    \item Columns are not modified
    \item Data frame attributes \code{tbl} or \code{data.table}
          are preserved
    \item Grouping structures are not preserved
}
}
\description{
Subset a data frame by student ID, retaining students who are degree-seeking.
}
\details{
An inner join between \code{dframe} and the ID column of \code{dbase}
retains all rows of \code{dframe} with matching IDs.
}
\examples{
# TBD
}
\seealso{
\code{add_race_sex()}, \code{add_term_timely()},
         \code{add_data_sufficiency()}, \code{add_completion_timely()}
}
