% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_undergrad.R
\name{filter_undergrad}
\alias{filter_undergrad}
\title{Choose rows of undergraduate terms only}
\usage{
filter_undergrad(dframe, midf_table = degree, ..., add_bacc_term = NULL)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \code{{mcid}} and one of the
following: \verb{\{term, term_course, term_degree\}.}}

\item{midf_table}{\code{degree} data frame with required variables
\verb{\{mcid, term_degree\}.}}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{add_bacc_term}{Logical, default false. If true, a column for the first
degree term is added and post-baccalaureate rows are not removed.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Post-baccalaureate rows are removed (default) unless \code{add_bacc_term} is true.
In all cases, duplicated rows are removed.
\item Columns are not modified (default). When \code{add_bacc_term} true, one new
column is added unless it is redundant (see Details). The new variable is:
\itemize{
\item \code{bacc_term}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA.}
Joined from the \code{term_degree} variable in \code{midf_table.}
}
}
}
\description{
Distinguish post-baccalaureate terms from undergraduate terms for each
student in a data frame and retain the undergraduate terms. Applied
to a data table having an academic term variable, e.g., the \verb{term, course,}
and \code{degree} tables.
}
\details{
\emph{Redundant columns.} To prevent overwriting, a variable such as
\code{bacc_term}  is not added to the data frame if it duplicates
an existing variable. The test for redundancy is managed internally
by calling \code{rm_redundant_cols()} before the final data frame is
returned. For documentation, see \code{?rm_redundant_cols}.
}
\examples{
# For illustration only, choose a minimum set of columns
term <- toy_term[, .(mcid, term)]
course <- toy_course[, .(mcid, term_course)]
degree <- toy_degree[, .(mcid, term_degree)]

# Example result
term

# Filter to retain undergrad terms only
x <- filter_undergrad(term, midf_table = degree)
x

# `add_bacc_term` argument adds the first degree term and rows are not removed
y <- filter_undergrad(term, midf_table = degree, add_bacc_term = TRUE)
y[order(bacc_term)]

# Filtering for undergraduate terms can be done manually
y <- y[bacc_term >= term | is.na(bacc_term)]
y[, bacc_term := NULL]
y

# Verify result
check_equiv_frames(x, y)


# Function is applied to all tables containing a term-value
term <- filter_undergrad(term, midf_table = degree)
course <- filter_undergrad(course, midf_table = degree)
degree <- filter_undergrad(degree, midf_table = degree)
}
