% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fye_mi_prep.R
\name{fye_mi_prep}
\alias{fye_mi_prep}
\title{Prepare FYE data for multiple imputation}
\usage{
fye_mi_prep(data_students = NULL, data_terms = NULL)
}
\arguments{
\item{data_students}{data frame of student attributes}

\item{data_terms}{data frame of term attributes}
}
\value{
data frame with the following properties:
\itemize{
    \item One row for every FYE student in the data, keyed by ID
    \item Columns for ID, institution, race, sex, and CIP code
    \item Data frame extensions such as \code{tbl} or \code{data.table}
    are preserved
}
}
\description{
Extract first-year-engineering (FYE) students from data and prepare
a data frame for input to the mice package for multiple imputation to
predict starting programs of FYE students.
}
\details{
Some US institutions have first year engineering (FYE) programs, typically
a common first year curriculum that is a prerequisite for declaring
an engineering major. FYE is problematic when computing a persistence
metric such as graduation rate that counts a graduation successful only
for students who matriculate and graduate in the same program.

We have to predict, therefore, the degree-granting engineering program the
FYE student would have declared had they not been required to enroll in
FYE. Predicting this "starting program" takes one of two forms.

\enumerate{
\item{Students who complete an FYE and declare an engineering major.
    This is the easy case--at the student's first opportunity, they
    enrolled in an engineering program of their choosing. We use that
    program as our predicted  starting program. }
\item{Students who, after FYE, do not declare an engineering major.
    This is the more complicated case---the data provide no information
    regarding what engineering program the student would have declared
    originally had the institution not required them to enroll in FYE.
    For these students, we treat the starting program as missing data and
    impute a predicted value using multiple imputation as implemented
    in the mice package.}
}

The function extracts all FYE students from \code{midfieldstudents} (or
equivalent) and from \code{midfieldterms} (or equivalent) determines
students' post-FYE programs. If the program is in engineering, the CIP
is retained. If not, the CIP is replaced with NA to represent missing
data for the imputation.

Finally, the predictor variables (institution, race, sex) are converted
to factors. The output is ready for use as an input to the mice package.
}
\examples{
# placeholder

}
\seealso{
Other functions: 
\code{\link{feasible_subset}()},
\code{\link{filter_by_cip}()},
\code{\link{filter_by_id}()},
\code{\link{filter_by_text}()},
\code{\link{ipeds_subset}()},
\code{\link{order_multiway}()}
}
\concept{functions}
