% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fye_predicted_start}
\alias{fye_predicted_start}
\title{Predicted starting programs for FYE students}
\format{
\code{data.table} with 5033 rows and 2 columns keyed by student
ID. The variables are:
\describe{

  \item{mcid}{Character, de-identified student ID.}

  \item{start}{Character, 6-digit CIP code of the predicted starting
  program.}

 }
}
\usage{
fye_predicted_start
}
\description{
Predicted starting programs for all First-Year Engineering (FYE) students 
in the midfielddata data tables. Keyed by student ID. Contains 
6-digit CIP codes that are substituted for the First-Year-Engineering (FYE) 
CIP code when an analysis requires starting programs that are 
degree-granting. Suitable for use with midfielddata practice data only.
}
\details{
Starting programs in \code{fye_predicted_start} are the engineering 
programs we predict that students would have declared had they not been 
required to enroll in FYE. The prediction is based on their first post-FYE 
program and a multivariate imputation suitable for multiple categorical 
variables using the mice package. The predictor variables are institution, 
race, and sex. The predicted variable is the 6-digit CIP code of a 
degree-granting engineering program at their institution.

\code{fye_predicted_start} holds only for the practice data in
midfielddata---these values cannot be commingled with the research
database available to MIDFIELD partners.
}
\seealso{
Other cip-data: 
\code{\link{cip}}
}
\concept{cip-data}
\keyword{datasets}
