% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fye_start}
\alias{fye_start}
\title{Predicted starting CIP for FYE students}
\format{
\code{data.table} with 5992 rows and 2 columns keyed by student
ID. The variables are:
\describe{
   \item{id}{student ID}
   \item{start}{starting program 6-digit CIP code}
   }
}
\usage{
fye_start
}
\description{
Predicted starting programs for all FYE students in midfielddata.
Includes degree-granting programs that substitute for
first-year-engineering (FYE) programs when needed for a persistence
metric, e.g., graduation rate.
}
\details{
Starting programs for FYE students are the engineering programs we
predict they would have declared had they not been required to enroll in
FYE. The prediction is based on their first post-FYE program and an
imputation suitable for multiple categorical variables using the mice
package. The predictor variables are institution, race, and sex. The
predicted variable is the 6-digit CIP code.
}
\seealso{
Other cip data: 
\code{\link{cip}}
}
\concept{cip data}
\keyword{datasets}
