% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_grad.R
\name{gather_grad}
\alias{gather_grad}
\title{Gather graduates}
\usage{
gather_grad(cip)
}
\arguments{
\item{cip}{A data frame created using \code{filter_cip()}.}
}
\value{
A data frame of students graduating in the specified programs.
}
\description{
Gather graduates from the \code{degree} dataset using specified CIP program codes. If in the same term as earning their first bachelor's degree a student also earns a second or third bachelor's degree, the additional degrees appear as additional observations (rows) in the result. The student MID appears once for each degree.
}
