% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cip.R
\name{get_cip}
\alias{get_cip}
\title{Subset CIP data frame}
\usage{
get_cip(data = NULL, keep_any = NULL, ..., drop_any = NULL)
}
\arguments{
\item{data}{data frame of CIP codes and names}

\item{keep_any}{character vector of search patterns for retaining rows}

\item{...}{not used for values. Forces optional arguments to be usable
by name only.}

\item{drop_any}{(optional) character vector of search patterns for
dropping rows}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows are a subset of the input in the same order
  \item Columns are not modified
  \item Grouping structures, if any, are not preserved
  \item Data frame extension attributes, e.g., tibble, are not preserved
}
}
\description{
Subset the rows of a data frame by matching character patterns. The data
are CIP attributes with 2-, 4-, and 6-digit codes and names in character
columns (\code{\link[midfieldr]{cip}} or equivalent). Rows are subset by
matching or partially matching patterns. The search is not case-sensitive.
}
\details{
The default \code{data} argument is \code{cip}, accessed by name or
by leaving the \code{data} argument vacant. Any other \code{data}
argument should be a subset of \code{cip} or equivalent.
}
\examples{
get_cip(cip, keep_any = c("^1407", "^1410"))
get_cip(cip, keep_any = "civil engineering", drop_any = "technology")
get_cip(cip, keep_any = "History") \%>\%
  get_cip(keep_any = "American")
}
\seealso{
Other data_query: 
\code{\link{get_enrollees}()},
\code{\link{get_graduates}()},
\code{\link{get_institution_hours_term}()},
\code{\link{get_institution_limits}()},
\code{\link{get_race_sex}()},
\code{\link{get_status_degrees}()},
\code{\link{get_status_transfers}()}
}
\concept{data_query}
