% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrollees.R
\name{get_enrollees}
\alias{get_enrollees}
\title{Get IDs of students ever enrolled in programs}
\usage{
get_enrollees(data = NULL, codes = NULL)
}
\arguments{
\item{data}{data frame of term attributes}

\item{codes}{character vector of 6-digit CIP codes}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows for which CIP is in \code{codes} and student-program
  pairings are unique
  \item Columns \code{id} and \code{cip6}
  \item Grouping structures, if any, are not preserved
  \item Data frame extensions \code{tbl} or \code{data.table} are preserved
}
}
\description{
Subset the rows of a data frame by CIP codes. The data are term attributes
keyed by student ID and term  (\code{\link[midfielddata]{midfieldterms}}
or equivalent). Rows are subset by 6-digit CIP codes and unique pairings of
CIP code and student ID. Results are the ID and CIP columns for
students ever enrolled in the programs.
}
\details{
The default \code{data} argument is \code{midfieldterms}, accessed by
name or by leaving the \code{data} argument vacant. Any other
\code{data} argument should be a subset of \code{midfieldterms}
or equivalent. Character columns \code{id} and \code{cip6} are required
by name.
}
\examples{
music_programs <- get_cip(cip, "^5009")
music_codes <- music_programs$cip6
x <- get_enrollees(codes = music_codes)
str(x)

music_codes <- c("500903", "500913")
x <- get_enrollees(codes = music_codes)
str(x)

}
\seealso{
Other data_query: 
\code{\link{get_cip}()},
\code{\link{get_graduates}()},
\code{\link{get_institution_hours_term}()},
\code{\link{get_institution_limits}()},
\code{\link{get_race_sex}()},
\code{\link{get_status_degrees}()},
\code{\link{get_status_transfers}()}
}
\concept{data_query}
