% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graduates.R
\name{get_graduates}
\alias{get_graduates}
\title{Get IDs of students graduating from programs}
\usage{
get_graduates(data = NULL, codes = NULL)
}
\arguments{
\item{data}{data frame of degree attributes}

\item{codes}{character vector of 6-digit CIP codes}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows for which CIP is in \code{codes} and graduation is
  observed, i.e., the value of \code{term_degree} is not NA
  \item Columns \code{id} and \code{cip6}
  \item Grouping structures, if any, are not preserved
  \item Data frame extension attributes, e.g., tibble, are not preserved
}
}
\description{
Subset the rows of a data frame by CIP codes. The data are degree
attributes keyed by student ID (\code{\link[midfielddata]{midfielddegrees}}
or equivalent). Rows are subset by program CIP codes and the first term
in which a student earns a degree. Results are the ID and CIP columns for
students graduating in the programs.
}
\details{
The default \code{data} argument is \code{midfielddegrees}, accessed by
name or by leaving the \code{data} argument vacant. Any other \code{data}
argument should be a subset of \code{midfielddegrees} or equivalent.
Character columns \code{id} and \code{cip6} and numerical column
\code{term_degree} are required by name.

Degrees earned after the first degree term are dropped. Multiple degrees,
if earned in the first degree term, are retained.
}
\examples{
music_programs <- get_cip(cip, "^5009")
music_codes <- music_programs$cip6
x <- get_graduates(codes = music_codes)
str(x)

music_codes <- c("500901", "500903", "500913")
x <- get_graduates(codes = music_codes)
str(x)

}
\seealso{
Other data_carpentry: 
\code{\link{filter_char_frame}()},
\code{\link{get_cip}()},
\code{\link{get_enrollees}()},
\code{\link{get_race_sex}()},
\code{\link{label_programs}()},
\code{\link{order_multiway}()}
}
\concept{data_carpentry}
