% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_institution_hours_term.R
\name{get_institution_hours_term}
\alias{get_institution_hours_term}
\title{Get median credit hours per term by institution}
\usage{
get_institution_hours_term(data = NULL, keep_id = NULL)
}
\arguments{
\item{data}{data frame of term attributes}

\item{keep_id}{character vector of student IDs}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item One row per institution
  \item Columns \code{institution} and \code{median_hr_per_term}
  \item Grouping structures, if any, are not preserved
  \item Data frame extension attributes, e.g., tibble, are not preserved
}
}
\description{
Determine the median number of credit hours per term by institution
for a given group of students. The data are term attributes keyed by
ID, institution, and term   (\code{\link[midfielddata]{midfieldterms}}
or equivalent). Results are the institutions and median hours per term.
}
\details{
The default \code{data} argument is \code{midfieldterms}, accessed by
name or by leaving the \code{data} argument vacant. Any other
\code{data} argument should be a subset of \code{midfieldterms}
or equivalent. Character columns \code{id}, \code{institution} and
\code{hours_term} are required by name.

The default `keep_id` argument is the set of IDs in `data`.
}
\examples{
# placeholder
}
\seealso{
Other data_query: 
\code{\link{get_cip}()},
\code{\link{get_enrollees}()},
\code{\link{get_graduates}()},
\code{\link{get_institution_limits}()},
\code{\link{get_race_sex}()},
\code{\link{get_status_degrees}()},
\code{\link{get_status_transfers}()}
}
\concept{data_query}
