% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_institution_limits.R
\name{get_institution_limits}
\alias{get_institution_limits}
\title{Get limits on feasible completion by institution}
\usage{
get_institution_limits(data = NULL, span = NULL)
}
\arguments{
\item{data}{data frame of term attributes}

\item{span}{number of years for feasible completion}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item One row per institution
  \item Columns \code{institution}, \code{matric_limit},
  and \code{data_limit}
  \item Grouping structures, if any, are not preserved
  \item Data frame extension attributes, e.g., tibble, are not preserved
}
}
\description{
Determine the data limit and matriculation limit for program completion
feasibility by institution. The data are term attributes keyed by
institution and term (\code{\link[midfielddata]{midfieldterms}}
or equivalent). Rows are subset by the latest term in the data set
(the data limit) by institution. Results are the institutions,
matriculation limits, and data limits.
}
\details{
The default \code{data} argument is \code{midfieldterms}, accessed by
name or by leaving the \code{data} argument vacant. Any other
\code{data} argument should be a subset of \code{midfieldterms}
or equivalent. Character columns \code{institution} and \code{term} are
required by name.
}
\examples{
# placeholder
}
\seealso{
Other data_query: 
\code{\link{get_cip}()},
\code{\link{get_enrollees}()},
\code{\link{get_graduates}()},
\code{\link{get_institution_hours_term}()},
\code{\link{get_race_sex}()},
\code{\link{get_status_degrees}()},
\code{\link{get_status_transfers}()}
}
\concept{data_query}
