% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_race_sex.R
\name{get_race_sex}
\alias{get_race_sex}
\title{Get race/ethnicity and sex of students by ID}
\usage{
get_race_sex(data = NULL, keep_id = NULL)
}
\arguments{
\item{data}{data frame of student attributes}

\item{keep_id}{character vector of student IDs}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows for which student ID is in \code{keep_id}
  \item Columns \code{id}, \code{race}, and \code{sex}
  \item Grouping structures, if any, are not preserved
  \item Data frame extension attributes, e.g., tibble, are not preserved
}
}
\description{
Subset the rows of a data frame by student ID. The data are student
matriculation attributes keyed by student ID
(\code{\link[midfielddata]{midfieldstudents}} or equivalent). Results are
student ID, sex, and race/ethnicity.
}
\details{
The default \code{data} argument is \code{midfieldstudents}, accessed by
name or by leaving the \code{data} argument vacant. Any other
\code{data} argument should be a subset of \code{midfieldstudents}
or equivalent. Character columns \code{id}, \code{race}, and \code{sex}
are required by name.
}
\examples{
music_codes <- c("500901", "500903", "500913")
music_grads <- get_graduates(codes = music_codes)
grads_id <- music_grads$id
grads <- get_race_sex(keep_id = grads_id)
str(grads)
}
\seealso{
Other data_carpentry: 
\code{\link{get_cip}()},
\code{\link{get_enrollees}()},
\code{\link{get_graduates}()},
\code{\link{label_programs}()},
\code{\link{order_multiway}()}
}
\concept{data_carpentry}
