% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status_degrees.R
\name{get_status_degrees}
\alias{get_status_degrees}
\title{Get degree status of students by ID}
\usage{
get_status_degrees(data = NULL, keep_id = NULL)
}
\arguments{
\item{data}{data frame of degree attributes}

\item{keep_id}{character vector of student IDs}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows for which student ID is in \code{keep_id}
  \item Columns \code{id}, \code{institution}, and \code{degree}
  \item Grouping structures, if any, are not preserved
  \item Data frame extensions \code{tbl} or \code{data.table} are preserved
}
}
\description{
Subset the rows of a data frame by student ID. The data are student
degree attributes keyed by student ID
(\code{\link[midfielddata]{midfielddegrees}} or equivalent). Results are
student ID, institution, and degree.
}
\details{
The default \code{data} argument is \code{midfielddegrees}, accessed by
name or by leaving the \code{data} argument vacant. Any other
\code{data} argument should be a subset of \code{midfielddegrees}
or equivalent. Character columns \code{id}, \code{institution}, and
\code{degree} are required by name.

If the \code{keep_id} argument is NULL, all IDs in \code{data} are used
by default.
}
\examples{
music_codes <- c("500901", "500903", "500913")
students <- get_enrollees(codes = music_codes)
students <- get_status_degrees(keep_id = students$id)
str(students)

}
\seealso{
Other data_query: 
\code{\link{get_cip}()},
\code{\link{get_enrollees}()},
\code{\link{get_graduates}()},
\code{\link{get_institution_hours_term}()},
\code{\link{get_institution_limits}()},
\code{\link{get_race_sex}()},
\code{\link{get_status_transfers}()}
}
\concept{data_query}
