% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_filter.R
\name{grad_filter}
\alias{grad_filter}
\title{Filter degree data for graduating students}
\usage{
grad_filter(
  data = NULL,
  codes = NULL,
  ...,
  id = "id",
  cip6 = "cip6",
  term = "term_degree"
)
}
\arguments{
\item{data}{Data frame of student IDs, academic terms, and CIP codes,
default \code{midfielddegrees}.}

\item{codes}{Atomic character vector of 6-digit CIP codes specifying
the programs to filter by.}

\item{...}{Not used for values. Forces the subsequent optional arguments
to only be usable by name.}

\item{id}{Column name in quotes of the student ID
variable in \code{data}. Default is "id". Optional argument.}

\item{cip6}{Column name in quotes of the 6-digit CIP code
variable in \code{data}. Default is "cip6". Optional argument.}

\item{term}{Column name in quotes of the degree term variable in
\code{data}. Default is "term_degree". Optional argument.}
}
\value{
Data frame with character variables for student ID and program CIP
}
\description{
Filter academic degree data to find all students graduating from
specified programs.
}
\details{
The \code{midfielddata} package must be installed to provide the
 \code{midfielddegrees} data frame as the default value for the
 \code{data} argument. Optionally, a data frame having the same structure
 as \code{midfielddegrees} may be used as the value for the \code{data}
  argument.

The \code{codes} argument is an atomic character vector of 6-digit CIP
codes used to filter \code{data} and extract student IDs and terms.

The function returns a subset of \code{data} with every unique combination
of student ID and CIP of the program(s) in which they earned their
first degree(s). Only the student ID and program CIP are returned;
other variables in \code{data} are quietly dropped.

The default values of the optional arguments are the column names
\code{"id"}, \code{"cip6"}, and \code{"term_degree"}.
If using a data frame with different column names, you can rename
your variables to match the defaults or use the optional arguments to
pass your variable name(s) to the function.
}
\examples{
library("midfielddata")
(grad_filter(codes = "540104"))
}
\seealso{
\code{\link[midfieldr]{cip6_select}} for selecting 6-digit CIP
codes and naming programs.

Other data_carpentry: 
\code{\link{cip6_select}()},
\code{\link{cip_filter}()},
\code{\link{ever_filter}()},
\code{\link{multiway_order}()},
\code{\link{race_sex_join}()},
\code{\link{start_filter}()},
\code{\link{term_summarize}()}
}
\concept{data_carpentry}
