% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grade_scale}
\alias{grade_scale}
\title{Grade scale}
\format{
\code{data.table} with 12 rows and 2 columns:
\describe{
\item{\code{letter_grade}}{Character, letter grades using the conventional US
scale from A to F.}
\item{\code{points}}{Numerical, 4.0 scale of points assigned to letter grades.}
}
}
\usage{
grade_scale
}
\description{
Data frame of letter grades and conventional point assignments used for
computing grade point averages.
}
\seealso{
Other scales:
\code{\link{act_sat_scale}}
}
\concept{scales}
\keyword{datasets}
