% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\docType{data}
\name{graph_functions}
\alias{expon_scale_x_log10}
\alias{midfield_theme}
\alias{darkBr}
\alias{medBr}
\alias{lightBr}
\alias{paleBr}
\alias{minBr}
\alias{minBG}
\alias{paleBG}
\alias{lightBG}
\alias{medBG}
\alias{darkBG}
\alias{darkPR}
\alias{medPR}
\alias{lightPR}
\alias{palePR}
\alias{minPR}
\alias{minGn}
\alias{paleGn}
\alias{lightGn}
\alias{medGn}
\alias{darkGn}
\alias{gray1}
\alias{gray2}
\alias{gray3}
\alias{gray4}
\alias{gray5}
\alias{gray6}
\alias{gray7}
\alias{gray8}
\alias{gray9}
\title{Graph functions}
\format{An object of class \code{character} of length 1.}
\usage{
expon_scale_x_log10()

midfield_theme()

darkBr

medBr

lightBr

paleBr

minBr

minBG

paleBG

lightBG

medBG

darkBG

darkPR

medPR

lightPR

palePR

minPR

minGn

paleGn

lightGn

medGn

darkGn

gray1

gray2

gray3

gray4

gray5

gray6

gray7

gray8

gray9
}
\value{
Modifies an existing graph
}
\description{
Provides a set of customized layers and arguments graphs created with for \code{ggplot2}.
}
\details{
\describe{

\item{\code{expon_scale_x_log10()}}{Applies the usual \code{scale_x_log10()} function with the scale marked powers of ten the exponent.}

\item{\code{midfield_theme()}}{Applies \code{theme_minimal()} with some additional edits.}

\item{\code{darkBr} ... \code{minBr}}{Brown color palette. Five hues named \code{darkBr, medBr, lightBr, paleBr, minBr} from the RColorBrewer "BrBG" palette.}

\item{\code{darkBG} ... \code{minBG}}{Blue-green color palette. Five hues named \code{darkBG, medBG, lightBG, paleBG, minBG} from the RColorBrewer "BrBG" palette.}

\item{\code{darkPR} ... \code{minPR}}{Purple color palette. Five hues named \code{darkPR, medPR, lightPR, palePR, minPR} from the RColorBrewer "PRGn" palette.}

\item{\code{darkGn} ... \code{minBr}}{Green color palette. Five hues named  \code{darkGn, medGn, lightGn, paleGn, minGn} from the RColorBrewer "PRGn" palette.}

\item{\code{gray9} ... \code{gray1}}{Gray color palette. Nine hues named   \code{gray9, gray8, ... , gray2, gray1} from the RColorBrewer "Greys" palette.}

}
}
\keyword{datasets}
