% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_filter}
\alias{id_filter}
\title{Filter by ID}
\usage{
id_filter(data = NULL, input_id = NULL, ..., id = "id")
}
\arguments{
\item{data}{One of the midfielddata data sets, e.g., \code{midfieldstudents}
or  \code{midfieldterms}. Must have a student ID variable.}

\item{input_id}{Character vector of student IDs used to filter
\code{data}.}

\item{...}{Not used for values. Forces the subsequent optional arguments
to only be usable by name.}

\item{id}{Column name in quotes of the student ID variable in
\code{data}. Default is "id". Optional argument.}
}
\description{
Internal function for filtering a midfielddata data set by student iD
}
\keyword{internal}
