% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_fye_proxies.R
\name{initialize_fye_proxies}
\alias{initialize_fye_proxies}
\title{Initialize FYE proxies for imputing missing data}
\usage{
initialize_fye_proxies(m_student, m_term, fye_cip = NULL, ..., alt_fye = NULL)
}
\arguments{
\item{m_student}{Data frame or data frame extension (e.g., data.table or tibble) with required character variables
\verb{\{mcid, race, sex\}.} Typically the original, unfiltered
\code{student} source data.}

\item{m_term}{Data frame or data frame extension (e.g., data.table or tibble) with required character variables
\verb{\{mcid, term, cip6, institution\}.} Typically the original,
unfiltered \code{term} source data.}

\item{fye_cip}{Character, one 6-digit CIP code used for FYE programs. Default
"140102", applied to all institutions except those (if any)
optionally defined by user in \code{alt_fye.}}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{alt_fye}{Data frame or data frame extension (e.g., data.table or tibble) with character variables
\verb{\{institution, alt_cip\}.} For users with institutions that use
a 6-digit CIP code other than the value in \code{fye_cip} for their FYE
programs. One FYE code only per institution.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Rows: One row for every degree-seeking FYE student. Rows in \code{m_student}
or \code{m_term} with NA values in any of the required variables are removed.
\item Columns: Conditioned for later use as an input to the mice R
package for multiple imputation as follows:
\itemize{
\item \code{mcid}   Character, anonymized student identifier.
\item \code{race}   Factor, race/ethnicity from the \code{student} input data
frame. An imputation predictor variable.
\item \code{sex}   Factor, sex from the \code{student} input data
frame. An imputation predictor variable.
\item \code{institution}   Factor, anonymized institution name from the
\code{term} data frame. An imputation predictor variable.
\item \code{proxy}   Factor, 6-digit CIP code of a student's known,
first degree-granting engineering program or NA representing missing
values to be imputed.
}
}
}
\description{
Conditions data for imputing starting
majors of First-Year Engineering (FYE) students. The result is suitably formatted for
input to the R mice package for multiple imputation.
}
\details{
Assembles a data frame of students ever enrolled in First-Year Engineering
(FYE) programs. Where practicable, a 6-digit CIP code is added to the data
frame as a proxy for the student's preferred engineering major. If
indeterminate, the proxy is NA and treated as missing data.

At some US institutions, engineering students are required to complete a First-Year Engineering (FYE) curriculum before they can be admitted to a degree-granting major such as Civil, Electrical, or Mechanical Engineering. This poses a problem when trying to count the number of students starting in one of these programs: the students don't start in Civil, Electrical, or Mechanical Engineering; they start in FYE.

For some metrics---graduation rate for example---correctly identifying starters is imperative. The problem posed by FYE programs is that we don't know the engineering starting majors these students would have selected had FYE not been required. We address the problem by constructing an \emph{FYE proxy} variable.

The FYE proxy has one of two values:
\enumerate{
\item{If the record of an FYE student includes a degree-granting
engineering program, the 6-digit CIP code of the first such program is
returned as the student's FYE proxy.}
\item{If not, the proxy is NA and is treated as a missing value to be
imputed later using the R mice package.}
}
This function does not perform the imputation. It produces a data
frame suitably formatted for input to the R mice package for multiple
imputation.
}
\examples{
library(data.table)

# Subset student and term data using selected IDs
IDs <- c("MCID3112319668", "MCID3112214437", "MCID3112328548", 
         "MCID3111447797", "MCID3111566004", "MCID3111697452", 
         "MCID3112268500", "MCID3112320295")
student <- select_basic_cols(toy_student[mcid \%chin\% IDs])
term <- select_basic_cols(toy_term[mcid \%chin\% IDs])

# Obtain results
proxy <- initialize_fye_proxies(student, term)
proxy

# ---------- Examine details
# Note: the CIP code and name for FYE is 140102 Pre-Engineering

# Join program names to term data for display
term_seq <- cip[term, .(mcid, term, cip6, cip6name), on = "cip6", nomatch = NULL]

# Function to display results for individual students
f <- function(IDs, i) {
    cat(paste("Student", i, "record\n"))
    print(term_seq[mcid == IDs[i]])
    cat("\ninitialize_fye_proxies() results\n")
    print(proxy[mcid == IDs[i]])
}

# Example 1: Non-Engineering -> FYE -> Engineering
# 400501 (Chemistry) -> FYE -> 140701 (Chemical Engng)
# FYE proxy is 140701
f(IDs, 1)

# Example 2: FYE -> Engineering -> Non-Engineering
# FYE -> 140901 (Computer Engng) -> 450601 (Economics)
# FYE proxy is 140901
f(IDs, 2)

# Example 3: FYE -> Engineering
# FYE -> 141001 (Electrical Engng)
# FYE proxy is 141001
f(IDs, 3)

# Example 4: FYE -> Engineering -> Engineering
# FYE -> 141901 (Mechanical Engng) -> 143501 (Industrial Engng)
# FYE proxy is 141901 
f(IDs, 4)

# Example 5: Non-Engineering -> FYE -> Leaves the database
# 240102 (General Studies) -> FYE
# FYE proxy is NA 
f(IDs, 5)

# Example 6: FYE -> Leaves the database
# FYE proxy is NA 
f(IDs, 6)

# Example 7: Non-Engineering -> FYE -> Non-Engineering
# 240102 (General Studies) -> FYE -> 110101 (Computer Science)
# FYE proxy is NA 
f(IDs, 7)

# Example 8: FYE -> Non-Engineering
# FYE -> 230101 (English Literature)
# FYE proxy is NA 
f(IDs, 8)
}
