% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_subset.R
\name{ipeds_subset}
\alias{ipeds_subset}
\title{Subset rows for IPEDS constraints}
\usage{
ipeds_subset(
  starters,
  ...,
  span = NULL,
  data_students = NULL,
  data_terms = NULL,
  data_degrees = NULL
)
}
\arguments{
\item{starters}{data frame of students starting in programs}

\item{...}{not used, force later arguments to be used by name.}

\item{span}{numeric scalar, number of years to define successful
graduation, default is 6 years}

\item{data_students}{data frame of student attributes}

\item{data_terms}{data frame of term attributes}

\item{data_degrees}{data frame of degree attributes}
}
\value{
data frame with the following properties:
\itemize{
    \item Rows subset to satisfy IPEDS definition
    \item Columns \code{id}, \code{start}, and \code{ipeds_grad}
    \item Grouping structures are not preserved
    \item Data frame extensions such as \code{tbl} or \code{data.table}
    are preserved
}
}
\description{
Subset a data frame of student IDs and starting programs, retaining
students satisfying the constraints of the IPEDS definition for
successful graduation. The results adds a yes/no column \code{ipeds_grad} to
indicate whether the student's graduation satisfied the IPREDS constraints.
}
\details{
The \code{starters} input must have columns \code{id} and \code{start}.

The function accesses these midfielddata data sets:
\code{\link[midfielddata]{midfieldstudents}},
\code{\link[midfielddata]{midfieldterms}}, and
\code{\link[midfielddata]{midfielddegrees}} or their equivalents.
}
\examples{
# placeholder
}
\seealso{
Other functions: 
\code{\link{feasible_subset}()},
\code{\link{filter_by_cip}()},
\code{\link{filter_by_id}()},
\code{\link{filter_by_text}()},
\code{\link{fye_mi_prep}()},
\code{\link{order_multiway}()}
}
\concept{functions}
