% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_atomic_character}
\alias{is_atomic_character}
\title{Vector type predicate}
\usage{
is_atomic_character(x)
}
\arguments{
\item{x}{Object to be tested}
}
\description{
Check for a given type and whether it is atomic
}
\keyword{internal}
