% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_undergrad.R
\name{is_undergrad}
\alias{is_undergrad}
\title{Categorize qualification level by term}
\usage{
is_undergrad(dframe, midf_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \code{{mcid}} and one of
\verb{\{term, term_course, term_degree\}.}}

\item{midf_table}{\code{degree} data frame with required variables
\verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added unless they duplicate existing
variables---redundant columns, whether new or existing, are dropped
(see Details). The new variables are:
\itemize{
\item \code{bacc_term}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA}.
Joined from the \code{term_degree} variable in \code{midf_table.}
\item \code{term_focus}    Character. Indicating a term contributes to study
before or after a student's first baccalaureate.
Possible values are "undergrad" and "post-bacc."
}
}
}
\description{
Categorize the qualification level towards which a student is working in
each term. Two levels are used: “undergrad” for terms before a student's
first degree and “post-bacc” (post-baccalaureate) for terms after the first
degree. Added columns support the findings. Post-baccalaureate terms are
typically excluded from the \verb{term, course,} and \code{degree} data tables.
}
\details{
\emph{Redundant columns.} To prevent overwriting, the name of an added
variable such as  \code{bacc_term}  that matches that of an existing
variable is made unique by adding a suffix, e.g., \code{bacc_term.1.}
An added variable that otherwise duplicates the existing variable is
redundant and dropped. If not, the presence of the suffixed variable
indicates a potential error. The variables added by
midfieldr functions depend on fixed quantities, e.g., a student's
admission term or an institution's data range, and do not change
during an analysis.
}
\examples{
# Select min required columns
term <- toy_term[, .(mcid, term)]
course <- toy_course[, .(mcid, term_course)]
degree <- toy_degree[, .(mcid, term_degree)]

# Example of starting data frame
term

# Labeling terms by group: undergrad & grad
term <- is_undergrad(term, midf_table = degree)
course <- is_undergrad(course, midf_table = degree)
degree <- is_undergrad(degree, midf_table = degree)

# Example result
term[order(-term_focus)]

# No change if added columns duplicate existing
x <- copy(term)
y <- is_undergrad(x, midf_table = degree)
check_equiv_frames(x, y)

# Filter to retain "undergraduate" rows only
term <- term[term_focus == "undergrad"]
course <- course[term_focus == "undergrad"]
degree <- degree[term_focus == "undergrad"]

# Example result
term
}
