% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_demographics.R
\name{join_demographics}
\alias{join_demographics}
\title{Join student demographics to term, course, or degree data}
\usage{
join_demographics(x)
}
\arguments{
\item{x}{A data frame of student attributes for term, course, or degree.}
}
\value{
The input data frame with added demographic variables joined by MID and Institution.
}
\description{
Gather student demographic information from the \code{student} dataset and use `dplyr::left_join()` to join to a data frame.
}
