% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kable2html}
\alias{kable2html}
\title{Create table}
\usage{
kable2html(x)
}
\arguments{
\item{x}{An object to be printed}
}
\description{
Internal function for vignettes to print to html using knitr::kable
and set font size with kableExtra::kable_styling.
}
\keyword{internal}
