% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_programs.R
\name{label_programs}
\alias{label_programs}
\title{Label programs and extract 6-digit CIP columns}
\usage{
label_programs(data = NULL, label = NULL)
}
\arguments{
\item{data}{data frame of CIP codes and names}

\item{label}{scalar character string}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows are not modified
  \item The 6-digit code and name columns are not modified. Other
   columns are dropped and a \code{program} column is added.
  \item Grouping structures, if any, are not preserved
  \item Data frame extension attributes, e.g., tibble, are not preserved
}
}
\description{
Extract two columns of a data frame and add a column of program
labels. The data are a subset of midfieldr \code{\link[midfieldr]{cip}} or
equivalent. The result is a 3-column data frame with 6-digit CIP codes,
6-digit CIP names, and user-assigned program labels.
}
\details{
The typical \code{data} argument is an output of \code{\link{get_cip}()}
with six columns: 2-digit codes and names, 4-digit codes and names, and
6-digit codes and names. \code{label_programs()} retains the two 6-digit
columns and adds a \code{program} column for the labels. Equivalent to

\code{data$program <- "my_chosen_label"}\cr
\code{data[, c("cip6", "cip6name", "program")]}

\code{label_programs()} facilitates selecting labels suitable for
grouping, summarizing, and joining operations. One can assign a custom
label or use the 2-, 4-, or 6-digit program names in CIP. The
\code{label} argument has three options:

\enumerate{
  \item A string (in quotes), for example \code{label = "Electrical
  engineering"}, places the string in every row of the new
  \code{program} column
  \item If \code{label = "cip2name"}, the labels are the 2-digit CIP
  program names. Argument "cip6name" produces similar results.
  \item If \code{label = NULL} or "cip4name", the labels are the
  4-digit CIP program names
}
}
\examples{
ece_cip <- get_cip(cip, keep_any = "^1410")
label_programs(ece_cip, label = "ECE")
label_programs(ece_cip, label = "cip2name")
label_programs(ece_cip, label = "cip6name")

}
\seealso{
Other data_carpentry: 
\code{\link{filter_char_frame}()},
\code{\link{get_cip}()},
\code{\link{get_enrollees}()},
\code{\link{get_graduates}()},
\code{\link{get_race_sex}()},
\code{\link{order_multiway}()}
}
\concept{data_carpentry}
