% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{look_at}
\alias{look_at}
\title{Display structure}
\usage{
look_at(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Does not return anything. The side effect is to output to the terminal.
}
\description{
A wrapper on \code{base::str()} with arguments set to not show attributes,
to not show length, and to cut the width.
}
\examples{
# data frames
look_at(cip)
look_at(toy_degree)

# character vectors
x <- sort(unique(toy_degree$institution))
look_at(x)
}
