% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{math_stat}
\alias{math_stat}
\title{CIP codes for programs in mathematics and statistics}
\format{A character vector.}
\source{
NSF STEM programs: \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}.
}
\usage{
math_stat
}
\description{
All 6-digit codes from the 2010 CIP 27 series.
}
\examples{
\dontrun{
 math_stat
}
}
\keyword{datasets}
