% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiway_order.R
\name{multiway_order}
\alias{multiway_order}
\title{Multiway order}
\usage{
multiway_order(x)
}
\arguments{
\item{x}{A data frame with one numerical variable and two categorical
variables.}
}
\value{
A data frame with its two categorical variables converted
to factors and ordered by increasing medians of the numerical variable.
}
\description{
Prepare multiway data for graphing.
}
\details{
In multiway data, there is a single quantitative value (or response) for
every combination of levels of two categorical variables.
\code{multiway_order()} converts the categorical variables to factors and
orders the factors by increasing median values of the levels of the
categorical variables.
}
