% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_session}
\alias{my_session}
\title{Session information}
\usage{
my_session(pkg_names)
}
\arguments{
\item{pkg_names}{A character vector of package names}
}
\description{
Internal function to print a less verbose session information output for
vignettes using sessioninfo::session_info
}
\keyword{internal}
