% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{other_stem}
\alias{other_stem}
\title{CIP codes for other programs in STEM}
\format{A character vector.}
\source{
NSF STEM programs: \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}.
}
\usage{
other_stem
}
\description{
All 6-digit codes for STEM programs excluding \code{bio_sci}, \code{engr},
\code{math_stat}, and \code{phys_sci}.
}
\examples{
\dontrun{
 other_stem
}
}
\keyword{datasets}
