% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phys_sci}
\alias{phys_sci}
\title{CIP codes for physical sciences}
\format{A character vector.}
\source{
CIP 2010, US National Center for Education Statistics,
  \url{https://nces.ed.gov/ipeds/cipcode}.
}
\usage{
phys_sci
}
\description{
Classification of Instructional Programs (CIP) 6-digit codes for physical
science programs, series 40.
}
\examples{
head(phys_sci, n = 10L)

print(cip_filter(series = phys_sci))
}
\keyword{datasets}
