% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\source{
https://github.com/rstudio/ggvis/blob/master/R/pipe.R
}
\arguments{
\item{lhs, rhs}{An object and a function to apply to it.}
}
\description{
Re-exports the magrittr pipe operator, \code{\%>\%}.
}
\details{
The goal of re-exporting this function is to make it accessible to the
examples in the midfieldr package.
}
\examples{
cip_filter(series = "History") \%>\% cip_filter(series = "American")
cip_filter(series = "^14") \%>\% cip_filter(series = "Electrical")
}
