% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_multiway.R
\name{prepare_multiway}
\alias{prepare_multiway}
\title{Prepare multiway data for graphing}
\usage{
prepare_multiway(dframe, ..., details = NULL)
}
\arguments{
\item{dframe}{data frame of multiway data}

\item{...}{not used, forces later arguments to be used by name}

\item{details}{logical scalar to add columns reporting the medians on which
the order of the levels is based.}
}
\value{
\code{data.frame} with the following properties:
\itemize{
  \item Rows are not modified
  \item Quantitative column is not modified
  \item Categorical columns are factors with levels ordered by
median quantitative values
  \item Grouping structures, if any, are not preserved
  \item Data frame extensions \code{tbl} or \code{data.table}
  are preserved
}
}
\description{
Transform a data frame such that two categorical variables are factors
with levels ordered by medians of the quantitative variable. Rows and
columns of the data frame are otherwise unaffected.
}
\details{
In multiway data, there is a single quantitative value (or response)
for every combination of levels of two categorical variables.
\code{order_multiway()} converts the columns of categorical variables
to factors and orders the factor levels by increasing medians of the
quantitative response variable.

The \code{dframe} argument, therefore, must have three columns only. One
column is a quantitative variable (type numeric) and two columns are
categorical variables (type character or factor).
}
\examples{
catg1 <- rep(c("urban", "rural", "suburb", "village"), each = 2)
catg2 <- rep(c("men", "women"), times = 4)
value <- c(0.22, 0.14, 0.43, 0.58, 0.81, 0.46, 0.15, 0.20)

# structure before
dframe <- data.frame(catg1, catg2, value)
str(dframe)

# structure after
mw <- prepare_multiway(dframe)
str(mw)
}
