% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_sex_join.R
\name{race_sex_join}
\alias{race_sex_join}
\title{Join student demographics to a data frame}
\usage{
race_sex_join(df)
}
\arguments{
\item{df}{A data frame of student attributes that includes the variable \code{ID}.}
}
\value{
The incoming data frame plus two new columns \code{sex} and \code{race} joined by student ID. Other columns are unaffected.
}
\description{
Add variables \code{race} and \code{sex} from the \code{midfieldstudents} dataset to a data frame. `dplyr::left_join()` is the joining function and \code{ID} is the join-by variable.
}
