% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcb.R
\name{rcb}
\alias{rcb}
\title{Hex color codes from RColorBrewer}
\usage{
rcb(pattern)
}
\arguments{
\item{pattern}{A character variable with custom color names of the form "level_hue", for example, "dark_Br", "mid_BG", "light_Gn", and "pale_Gray."}
}
\value{
A hexadecimal color code in character form "#nnnnnn"
}
\description{
Accesses the \code{midfieldr::rcb_colors} dataset and returns a hex color code.
}
\examples{
rcb("dark_Br")

}
