% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcb_colors}
\alias{rcb_colors}
\title{Named colors from the ColorBrewer palettes}
\format{A data frame with columns:
\describe{
  \item{rcb_name}{Character variable of names.}
  \item{rcb_code}{Character variable of hex color codes.}
}}
\source{
Cynthia Brewer (\url{http://colorbrewer2.org}) and RColorBrewer  (\url{https://cran.r-project.org/package=RColorBrewer}).
}
\usage{
rcb_colors
}
\description{
Shorthand color names assigned to selected colors from the ColorBrewer palettes.
}
\details{
This dataset provides convenient access to selected ColorBrewer palettes: diverging brown-bluegreen (BrBG) with 8 levels; diverging purple-green (PRGn) with 8 levels; and sequential gray (Greys) with the middle four of 6 levels.

The color names have the form "level_hue" with 4 saturation levels (dark, mid, light, pale) and 5 hues (Br, BG, PR, Gn, Gray or Grey).
}
\examples{
rcb_colors
rcb("dark_Br")
rcb("light_Gn")
}
\keyword{datasets}
