% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_bracket.R
\name{record_bracket}
\alias{record_bracket}
\title{Identify terms after the first completion term}
\usage{
record_bracket(dframe, midf_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \code{{mcid}} and one of
\verb{\{term, term_course, term_degree\}.}}

\item{midf_table}{\code{degree} data frame with required variables
\verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added or replace existing columns of the same name (if
any). Other columns are not modified. The following variables are added:
\itemize{
\item \code{term_1st_degree}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA}.
Joined from the \code{term_degree} variable in \code{midf_table.}
\item \code{bracket}    Character, indicating that a term belongs
to one of two groups: "undergrad" terms are those leading up to
and including the term in which a student completes their first
degree; and "post-bacc" (post-baccalaureate) for all terms after the
first degree.
}
\item Groups and keys are not preserved.
}
}
\description{
For each student's term in a data frame, determine its relationship to the
student's first degree term (pre-degree, first-degree, or post-first-degree)
and add columns that support the findings. Post-first-baccalaureate terms
are typically excluded from the \verb{term, course,} and \code{degree} data tables.
}
\details{
In a typical analysis, one is interested in a student's progress up to
and including the term in which they earn their first degree or degrees.
Any terms later than the first baccalaureate can usually be excluded from
study.
}
\examples{
# select min required columns
term <- toy_term[, .(mcid, term)]
course <- toy_course[, .(mcid, term_course)]
degree <- toy_degree[, .(mcid, term_degree)]

# labeling terms by group: undergrad & grad
term <- record_bracket(term, midf_table = degree)
course <- record_bracket(course, midf_table = degree)
degree <- record_bracket(degree, midf_table = degree)

# results
term[order(-bracket)]
term[, .N, by = "bracket"][order(-N)]

course[order(-bracket)]
course[, .N, by = "bracket"][order(-N)]

degree[order(-bracket)]
degree[, .N, by = "bracket"][order(-N)]
}
