% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export-wrapr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{stop_if_dot_args}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{wrapr}{\code{\link[wrapr]{stop_if_dot_args}}}
}}

