% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_cip}
\alias{rep_cip}
\title{Case CIP codes}
\format{
\code{data.table} with 12 rows and 6 columns keyed by the
6-digit CIP code. The variables are:
\describe{
  \item{cip6}{character 6-digit code representing "specific
  instructional programs" (US National Center for Education Statistics)}
  \item{cip6name}{character program name at the 6-digit level}
  \item{cip4}{character 4-digit code (the first 4 digits of \code{cip6})
  representing "intermediate groupings of programs that have
  comparable content and objectives"}
  \item{cip4name}{character program name at the 4-digit level}
  \item{cip2}{character 2-digit code (the first 2 digits of \code{cip6})
  representing "the most general groupings of related programs"}
  \item{cip2name}{character program name at the 2-digit level}
}
}
\usage{
rep_cip
}
\description{
Representative case study. A subset of the \code{cip} data set containing
CIP codes and names for Civil, Electrical, Industrial,
and Mechanical Engineering programs in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}.
}
\seealso{
Other case-study-data: 
\code{\link{rep_ever}},
\code{\link{rep_group}},
\code{\link{rep_start}},
\code{\link{rep_stickiness_mw}},
\code{\link{rep_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
