% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_ever}
\alias{rep_ever}
\title{Case student ID}
\format{
character vector
}
\usage{
rep_ever
}
\description{
Representative case study. Character vector of student IDs ever enrolled
in Civil, Electrical, Industrial, and Mechanical Engineering programs in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}.
}
\examples{
library(midfielddata)
rows_we_want <- midfieldstudents$id \%in\% rep_ever
enrollees <- midfieldstudents[rows_we_want]
head(enrollees)
}
\seealso{
Other example_data: 
\code{\link{rep_cip}},
\code{\link{rep_group}},
\code{\link{rep_stickiness_mw}},
\code{\link{rep_stickiness}}
}
\concept{example_data}
\keyword{datasets}
