% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_grad_rate}
\alias{rep_grad_rate}
\title{Graduation rate results}
\format{
\code{data.table} with 62 rows and 6 columns keyed by
program, race/ethnicity, and sex. The variables are:
\describe{
 \item{program}{character, instructional programs selected for study}
 \item{race}{character, student race/ethnicity}
 \item{sex}{character, student sex}
 \item{start}{numerical, number of students starting in a program
     and completion is feasible}
 \item{grad}{numerical, number of students graduating from the program}
 \item{rate}{numerical, program graduation rate, a persistence metric}
}
}
\usage{
rep_grad_rate
}
\description{
Representative case study. Graduation rate metric results for Civil,
Electrical, Industrial, and Mechanical Engineering programs in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}. Results
are grouped by program, race/ethnicity, and sex.
}
\seealso{
Other example_data: 
\code{\link{rep_cip}},
\code{\link{rep_grad_rate_mw}},
\code{\link{rep_group}},
\code{\link{rep_id}},
\code{\link{rep_start}},
\code{\link{rep_stickiness_mw}},
\code{\link{rep_stickiness}}
}
\concept{example_data}
\keyword{datasets}
