% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_grad_rate_mw}
\alias{rep_grad_rate_mw}
\title{Graduation rate results in multiway form}
\format{
\code{data.table} with 32 rows and 3 columns keyed by
program, race/ethnicity, and sex. The variables are:
\describe{
 \item{program}{factor, instructional programs selected for study}
 \item{race_sex}{factor, student race/ethnicity and sex}
 \item{rate}{numerical, program graduation rate, a persistence metric}
}
}
\usage{
rep_grad_rate_mw
}
\description{
Representative case study. Graduation rate metric results with two
categorical variables as factors with levels ordered by the median
graduation rate.
}
\seealso{
Other example_data: 
\code{\link{rep_cip}},
\code{\link{rep_grad_rate}},
\code{\link{rep_group}},
\code{\link{rep_id}},
\code{\link{rep_start}},
\code{\link{rep_stickiness_mw}},
\code{\link{rep_stickiness}}
}
\concept{example_data}
\keyword{datasets}
