% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_id}
\alias{rep_id}
\alias{rep_ever}
\alias{rep_grad}
\alias{rep_ever_fc}
\title{Student ID data sets}
\format{
character vector. The names of the available data vectors are:
\describe{
\item{rep_ever}{IDs of unique students ever enrolled in
  the example programs}
\item{rep_ever_fc}{subset of \code{rep_ever} for whom
  program completion is feasible in 6 years}
\item{rep_grad}{IDs of unique students graduating from
  the example programs}
}
}
\usage{
rep_ever

rep_grad

rep_ever_fc
}
\description{
Representative case study. Character vectors of student IDs at different
points in a case study of Civil, Electrical, Industrial, and Mechanical
Engineering programs in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}.
}
\examples{
library(midfielddata)

# ever enrolled
rows_we_want <- midfieldstudents$id \%in\% rep_ever
enrollees <- midfieldstudents[rows_we_want]
head(enrollees)

# graduates
rows_we_want <- midfieldstudents$id \%in\% rep_grad
graduates <- midfieldstudents[rows_we_want]
head(graduates)
}
\seealso{
Other example_data: 
\code{\link{rep_cip}},
\code{\link{rep_group}},
\code{\link{rep_start}},
\code{\link{rep_stickiness_mw}},
\code{\link{rep_stickiness}}
}
\concept{example_data}
\keyword{datasets}
