% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_stickiness_mw}
\alias{rep_stickiness_mw}
\title{Case stickiness multiway results}
\format{
\code{data.table} with 32 rows and 3 columns keyed by
program, race/ethnicity, and sex. The variables are:
\describe{
 \item{program}{factor, instructional programs selected for study}
 \item{race_sex}{factor, student race/ethnicity and sex}
 \item{stick}{numerical, program stickiness, a persistence metric}
}
}
\usage{
rep_stickiness_mw
}
\description{
Representative case study. Stickiness metric results with two categorical
variables as factors with levels ordered by the median stickiness.
}
\seealso{
Other example data: 
\code{\link{rep_cip}},
\code{\link{rep_ever}},
\code{\link{rep_group}},
\code{\link{rep_stickiness}}
}
\concept{example data}
\keyword{datasets}
