% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_redundant_cols.R
\name{rm_redundant_cols}
\alias{rm_redundant_cols}
\title{Remove redundant columns}
\usage{
rm_redundant_cols(dframe)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble)}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Rows are preserved.
\item Unique columns preserved. Redundant columns dropped.
}
}
\description{
Subset a data frame to remove redundant columns. The goal
of this function is to prevent overwriting existing columns and to avoid
duplicating columns. Primarily used internally in midfieldr functions that
add columns to a data frame.
}
\details{
Several midfieldr functions add columns to a working data frame. If the name
of a new variable matches that of an existing  variable,
the new variable name acquires a suffix ".1", ".2", etc., provided by
\verb{base::make.unique().} Before the final data frame is returned, all variable names
(including existing variables) are temporarily split from their dot suffixes, if any.
Resulting columns are dropped if they duplicate a previous column (where
'previous' means to the left of the duplicate in the data frame).
}
\examples{
# Construct a sample data frame
mcid <- c("mc_01", "mc_02", "mc_03")
term <- c("19911", "19912", "19913")
x <- data.frame(mcid, term)

# In the following, column names "var" and "var.i" 
# represent identical names except for the suffix ".i"
 
# No effect if columns are unique
x
x <- rm_redundant_cols(x)
x

# Desired effect: drop new "var.i" if it duplicates existing "var"
x$term.1 <- x$term
x
x <- rm_redundant_cols(x)
x

# Retain "var.i" when "var" does not exist
x <- data.frame(mcid, term.1 = term)
x
x <- rm_redundant_cols(x)
x

# The leftmost of the redundant columns is retained
x$term <- term
x
x <- rm_redundant_cols(x)
x

# Redundancy is checked for the dot-separator only
x <- data.frame(mcid, term, term_1 = term, term.1 = term)
x
x <- rm_redundant_cols(x)
x

# When a midfieldr function adds a variable, e.g., `bacc_term`, and
# a subsequent operation adds it again, `bacc_term.1`, the second 
# addition is usually expected to be redundant and dropped.
x$term_1 <- NULL
x$bacc_term <- c("19951", "19951", "19951")
x$bacc_term.1 <- c("19951", "19951", "19951")
x
x <- rm_redundant_cols(x)
x

# However, if the suffixed column remains, then the two columns 
# have different values---likely indicating an error in the 
# operations leading up to this point. 
x$bacc_term.1 <- c("19953", "19951", "19951")
x
x <- rm_redundant_cols(x)
x
}
