% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_content.R
\name{same_content-deprecated}
\alias{same_content-deprecated}
\title{Test for equal content between two data tables}
\usage{
same_content(x, y)
}
\arguments{
\item{x}{Data frame to be compared. Same as \code{target} argument in
\code{all.equal()}}

\item{y}{Data frame to be compared. Same as \code{current} argument in
\code{all.equal()}}
}
\value{
Either TRUE or a description of the differences between \code{x} and \code{y}.
}
\description{
This function is deprecated in favor of \code{\link[wrapr:check_equiv_frames]{wrapr::check_equiv_frames()}} imported from
the wrapr package, accessible by loading 'midfieldr'.
}
\details{
Test of data equality between data.table objects. Convenience function used
in 'midfieldr' articles.

Wrapper around \code{all.equal()} for class data.table that ignores row order,
column order, and data.table keys. Both inputs must be date frames.
Equivalent to:

\code{all.equal(target, current, ignore.row.order = TRUE, ignore.col.order = TRUE)}
}
\seealso{
\code{\link{midfieldr-deprecated}}
}
\keyword{internal}
