% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_record_cols.R
\name{select_record_cols}
\alias{select_record_cols}
\title{Choose columns of student records}
\usage{
select_record_cols(dframe, type = NULL, ..., col_pattern = NULL)
}
\arguments{
\item{dframe}{Data frame of student records from which columns are selected.
Expected choices are \code{student}, \code{term}, \code{course}, \code{degree} or their
equivalent.}

\item{type}{Character identifying the record type. Possible values are "s",
"t", "c", "d", or "a". See Details.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{col_pattern}{Character vector containing strings or regular
expressions to be matched or partially matched to the column
names of \code{dframe}.}
}
\value{
A data frame of the same type as \code{dframe}. The output has the
following properties:
\itemize{
\item Rows are not modified.
\item Columns are a subset of the input, but appear in the same order.
\item Groups are not necessarily preserved.
\item Data frame attributes are preserved with the exception of grouped tibbles.
}
}
\description{
Subset a data frame, selecting columns by matching a vector of character
strings. A convenience function to reduce the dimensions of a MIDFIELD
data table by selecting only those columns required by other midfieldr
functions or that are required to form a composite key. Particularly
useful in interactive sessions when viewing the data tables at various
stages of an analysis.
}
\details{
Several midfieldr functions require input data frames containing
specific variables (column names) such as \code{mcid} or \code{cip6}. In addition,
the MIDFIELD data tables have specific variables that act as keys
or composite keys to the information in that table. The \code{type} argument
determines which columns are returned, if those columns exist in \code{dframe}:
\itemize{
\item \code{type = "s"} (student table) returns columns \verb{mcid, race, sex}
\item \code{type = "t"} (term table) returns columns \verb{mcid, term, cip6, institution, level}
\item \code{type = "c"} (course table) returns columns \verb{mcid, term_course, abbrev, number}
\item \code{type = "d"} (degree table) returns columns \verb{mcid, term_degree, cip6}
\item \code{type = "a"} (default) returns all the above
}

Additional column names can be included by using the \code{col_pattern}
argument. In all cases, unmatched search strings are silently ignored.
}
\examples{
# Basic usage
select_record_cols(toy_student[1:5])
select_record_cols(toy_term[1:5])
select_record_cols(toy_course[1:5])
select_record_cols(toy_degree[1:5])

# Return columns by record type
select_record_cols(toy_student[1:5], type = "s")
select_record_cols(toy_term[1:5], "t")
select_record_cols(toy_course[1:5], "c")
select_record_cols(toy_degree[1:5], "d")

# With col_patterns for additional columns
DT <- toy_student[141:146]
select_record_cols(DT, "t", col_pattern = c("transfer", "hours_tranfer"))

# Using regular expressions
these_IDs <- DT$mcid
DT <- toy_term[mcid \%chin\% these_IDs]
select_record_cols(DT, "t", col_pattern = c("^gpa"))
}
