% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_records.R
\name{select_records}
\alias{select_records}
\title{Choose columns of student records}
\usage{
select_records(dframe, type = NULL, ..., col_pattern = NULL)
}
\arguments{
\item{dframe}{Data frame of student records from which columns are selected.
Expected choices are \code{student}, \code{term}, \code{course}, \code{degree} or their
equivalent.}

\item{type}{Character identifying the record type. Possible values are "s",
"t", "c", "d", "a", or NULL (default). See Details.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{col_pattern}{Character vector containing strings or regular
expressions to be matched or partially matched to the column
names of \code{dframe}.}
}
\value{
A data frame of the same type as \code{dframe}. The output has the
following properties:
\itemize{
\item Rows are not modified.
\item Columns are a subset of the input, but appear in the same order.
\item Groups are not necessarily preserved.
\item Data frame attributes are preserved with the exception of grouped tibbles.
}
}
\description{
Subset any of the four MIDFIELD data tables \verb{\{student, term, course, degree\}}
by selecting from that data frame the columns required by other midfieldr
functions.
}
\details{
A convenience function to reduce the dimensions of a MIDFIELD
data table by selecting only those columns required by other midfieldr
functions or that are required to form a composite key. Particularly
useful in interactive sessions when viewing the data tables at various
stages of an analysis.

Several midfieldr functions require input data frames containing
specific variables (column names) such as \code{mcid} or \code{cip6}. In addition,
the MIDFIELD data tables have specific variables that act as keys
or composite keys to the information in that table. If the \code{type} argument
is NULL (default), one of the following codes is assigned to return the
column names indicated (if present):
\itemize{
\item \code{type = "s"} (student) looks for \verb{\{mcid, race, sex\}}
\item \code{type = "t"} (term) looks for \verb{\{mcid, term, cip6, institution, level\}}
\item \code{type = "c"} (course) looks for  \verb{\{mcid, term_course, abbrev, number\}}
\item \code{type = "d"} (degree) looks for  \verb{\{mcid, term_degree, cip6\}}
\item \code{type = "a"} looks for all the above columns
}

Specifying the type manually \verb{\{s, t, c, d, a\}} in the argument overrides
the automatic selection. Additional column names can be included by using
the \code{col_pattern} argument. In all cases, unmatched search strings are
silently ignored.
}
\examples{
# Basic usage
select_records(toy_student[1:5])
select_records(toy_term[1:5])
select_records(toy_course[1:5])
select_records(toy_degree[1:5])

# With col_patterns for additional columns
DT <- toy_student[141:146]
select_records(DT, col_pattern = c("transfer", "hours_tranfer"))

# Using regular expressions
these_IDs <- DT$mcid
DT <- toy_term[mcid \%chin\% these_IDs]
select_records(DT, col_pattern = c("^gpa"))
}
