% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_cip.R
\docType{data}
\name{series_cip}
\alias{series_cip}
\alias{series_socsci}
\alias{series_hum}
\alias{series_perform}
\alias{series_hum_perform}
\alias{series_ag_sci}
\alias{series_chem}
\alias{series_comp_sci}
\alias{series_envir_sci}
\alias{series_geo_sci}
\alias{series_life_sci}
\alias{series_phy}
\alias{series_sci}
\alias{series_techn}
\alias{series_engr}
\alias{series_math}
\alias{series_stem}
\title{CIP codes of predefined groups of programs}
\format{A character vector}
\usage{
series_socsci

series_hum

series_perform

series_hum_perform

series_ag_sci

series_chem

series_comp_sci

series_envir_sci

series_geo_sci

series_life_sci

series_phy

series_sci

series_techn

series_engr

series_math

series_stem
}
\description{
For several groups of programs, CIP codes have been collected in predefined character vectors that can be used as the series argument in \code{filter_cip()}.
}
\examples{
length(series_engr)
series_engr

length(series_hum)
series_hum

length(series_math)
series_math

length(series_sci)
series_sci

length(series_socsci)
series_socsci

}
\keyword{datasets}
