% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_filter.R
\name{start_filter}
\alias{start_filter}
\title{Filter student data for starting students}
\usage{
start_filter(data = NULL, codes = NULL, ..., id = "id", cip6 = "cip6")
}
\arguments{
\item{data}{Data frame of student IDs and CIP codes, default
\code{midfieldterms}.}

\item{codes}{Atomic character vector of 6-digit CIP codes specifying the
programs to filter by.}

\item{...}{Not used for values. Forces the subsequent optional arguments
to only be usable by name.}

\item{id}{Column name in quotes of the student ID variable in
\code{data}. Default is "id". Optional argument.}

\item{cip6}{Column name in quotes of the 6-digit CIP code variable in
\code{data}. Default is "cip6". Optional argument.}
}
\value{
Data frame with character variables for student ID and CIP of
the starting program.
}
\description{
Filter entering student data to find all students starting in specified
programs.
}
\details{
The \code{midfielddata} package must be installed to provide the
 \code{midfieldstudents} data frame as the default value for the
 \code{data} argument. Optionally, a data frame having the same structure
 as \code{midfieldstudents} may be used as the value for the \code{data}
  argument.

The \code{codes} argument is an atomic character vector of 6-digit CIP
codes used to filter \code{data} and extract student IDs.

The function returns a subset of \code{data} with the student ID and
starting major CIP. Only the student ID and program CIP are returned;
other variables in \code{data} are quietly dropped.

The default values of the optional arguments are the column names
\code{"id"} and \code{"cip6"}.
If using a data frame with different column names, you can rename
your variables to match the defaults or use the optional arguments to
pass your variable name(s) to the function.
}
\examples{
placeholder <- 2 + 3
}
\seealso{
\code{\link[midfieldr]{cip6_select}} for selecting 6-digit CIP
codes and naming programs.

Other data_carpentry: 
\code{\link{cip6_select}()},
\code{\link{cip_filter}()},
\code{\link{ever_filter}()},
\code{\link{grad_filter}()},
\code{\link{multiway_order}()},
\code{\link{race_sex_join}()},
\code{\link{term_summarize}()}
}
\concept{data_carpentry}
