% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stem}
\alias{stem}
\title{CIP codes for all programs in STEM}
\format{A character vector.}
\source{
NSF STEM programs: \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}.
}
\usage{
stem
}
\description{
All 6-digit codes for STEM programs. Concatenates all named series:
\code{bio_sci}, \code{engr}, \code{math_stat}, \code{other_stem},
and \code{phys_sci}.
}
\examples{
\dontrun{
 stem
}
}
\keyword{datasets}
