% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_enrollees}
\alias{study_enrollees}
\title{Case-study student IDs and programs}
\format{
\code{data.table} with 12,409 rows and 2 columns keyed by
student ID. The variables are:
\describe{
  \item{id}{student ID}
  \item{cip6}{6-digit program code}
}
}
\usage{
study_enrollees
}
\description{
Data frame of student IDs and program codes of students ever enrolled
in the Civil, Electrical, Industrial, and Mechanical Engineering programs
from midfielddata, including predicted starting programs for FYE students.
Used in the case study developed in the vignettes.
}
\details{
The data have not been subset for timely completion, fair record length, or for
having a matriculation record.
}
\seealso{
Other case-study-data: 
\code{\link{study_programs}},
\code{\link{study_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
