% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_ever}
\alias{study_ever}
\title{Case-study ever enrolled}
\format{
\code{data.table} with 5653 rows and 2 columns.

\describe{

\item{mcid}{Character, anonymized student identifier}

\item{program}{Character, abbreviated labels for four engineering programs.
Values are "CE" (Civil Engineering), "EE" (Electrical Engineering), "ISE"
(Industrial/Systems Engineering), and  "ME" (Mechanical Engineering).}

}
}
\usage{
study_ever
}
\description{
Data table of IDs of students ever enrolled in the four programs of the study
(Civil, Electrical, Industrial/Systems, and Mechanical Engineering) after
filtering for data sufficiency, degree seeking, and program. Provided for
the convenience of vignette users.
}
\seealso{
Other case-study-data: 
\code{\link{study_observations}},
\code{\link{study_programs}},
\code{\link{study_results}},
\code{\link{study_starters}}
}
\concept{case-study-data}
\keyword{datasets}
