% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_grad_rate}
\alias{study_grad_rate}
\title{Case-study graduation rate results}
\format{
\code{data.table} with 32 rows and 3 columns keyed by
program, race/ethnicity, and sex.
\describe{
 \item{program}{Character, instructional programs selected for study}
 \item{race_sex}{Character, student race/ethnicity and sex}
 \item{grad_rate}{Numerical, program graduation rate, a persistence metric}
}
}
\usage{
study_grad_rate
}
\description{
Graduation rate metric results for Civil, Electrical,
Industrial, and Mechanical Engineering programs from midfielddata. Results
are grouped by program, race/ethnicity, and sex. Used in the case study
developed in the vignettes.
}
\details{
Graduation rate of a program is fraction of students starting in a
program who graduate in that program.

Data have also been subset to remove ambiguous levels of race/ethnicity.
}
\seealso{
Other case-study-data: 
\code{\link{study_programs}},
\code{\link{study_stickiness}},
\code{\link{study_students}}
}
\concept{case-study-data}
\keyword{datasets}
