% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_mcid}
\alias{study_mcid}
\title{Case-study IDs}
\format{
\code{data.table} with 7133 rows and 1 column.

\describe{

 \item{mcid}{Character, anonymized student identifier}

}
}
\usage{
study_mcid
}
\description{
An intermediate result provided for the convenience of vignette users. 
The IDs of students ever enrolled in the four programs of the study 
(Civil, Electrical, Industrial/Systems, and Mechanical Engineering) after 
filtering for data sufficiency, degree seeking, program, and timely 
completion.
}
\seealso{
Other case-study-data: 
\code{\link{study_observations}},
\code{\link{study_pool}},
\code{\link{study_program_cips}},
\code{\link{study_results}}
}
\concept{case-study-data}
\keyword{datasets}
