% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_pool}
\alias{study_pool}
\title{Case-study starting pool}
\format{
\code{data.table} with 161,696 rows and 2 columns.

\describe{

 \item{mcid}{Character, anonymized student identifier}

 \item{cip6}{Character, 6-digit CIP program codes.}

}
}
\usage{
study_pool
}
\description{
An intermediate result provided for the convenience of vignette users. 
The starting pool of students ever enrolled in the four programs of the 
case study. Keyed by student ID.
}
\details{
Starting with the midfielddata \code{term} practice data set, we extracted 
every observation with CIP codes of the four engineering programs in the 
case study: Civil, Electrical, Industrial/Systems, and Mechanical 
Engineering.
}
\seealso{
Other case-study-data: 
\code{\link{study_observations}},
\code{\link{study_results}}
}
\concept{case-study-data}
\keyword{datasets}
