% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_program}
\alias{study_program}
\title{Case-study programs}
\format{
\code{data.table} with 12 rows and 2 columns keyed by the
6-digit CIP code. The variables are:
\describe{

  \item{cip6}{Character, 6-digit code of a program.}

  \item{program}{Character, program name assigned for grouping and
  summarizing and for display in graphs and tables.}

}
}
\usage{
study_program
}
\description{
Data frame of 6-digit CIP codes and custom program names for Civil,
Electrical, Industrial, and Mechanical Engineering programs
from midfielddata. Used in the case study developed in the vignettes.
}
\details{
Program names have been shortened for convenience when creating
graphs and tables.
}
\seealso{
Other case-study-data: 
\code{\link{study_grad_rate}},
\code{\link{study_reduced_ever}},
\code{\link{study_reduced_grad}},
\code{\link{study_stickiness}},
\code{\link{study_student}}
}
\concept{case-study-data}
\keyword{datasets}
