% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_program_cips}
\alias{study_program_cips}
\title{Case-study program labels and CIP codes}
\format{
\code{data.table} with 15 rows and 2 columns.

\describe{

 \item{cip6}{Character, 6-digit CIP program codes.}
 
 \item{program}{Character, abbreviated labels for four engineering programs. 
 Values are "CE" (Civil Engineering), "EE" (Electrical Engineering), "ISE" 
 (Industrial/Systems Engineering), and  "ME" (Mechanical Engineering).}

}
}
\usage{
study_program_cips
}
\description{
An intermediate result provided for the convenience of vignette users. 
Program CIP codes and labels of the four programs of the case study. 
Keyed by 6-digit CIPs.
}
\details{
Starting with the midfieldr \code{cip} data set, we extracted the CIPs 
of the four programs of the case study and assigned them a custom label 
to be used for grouping and summarizing.
}
\seealso{
Other case-study-data: 
\code{\link{study_mcid}},
\code{\link{study_observations}},
\code{\link{study_pool}},
\code{\link{study_results}}
}
\concept{case-study-data}
\keyword{datasets}
