% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_reduced_data}
\alias{study_reduced_data}
\title{Case-study reduced data}
\format{
\code{data.table} with 7800 rows and 5 columns keyed by ID.

\describe{

 \item{mcid}{Character, anonymized student identifier}

 \item{ever}{Character, abbreviation indicating a program in 
 which the student was ever enrolled}
 
 \item{grad}{Character, abbreviation indicating the program 
 from which the student graduated in a timely fashion}

 \item{race}{Character, race/ethnicity as self-reported by the student, 
 e.g., Asian, Black, Hispanic/LatinX, etc.}
 
 \item{sex}{Character, sex as self-reported by the student, 
 values are Female, Male, and Unknown}

}
}
\usage{
study_reduced_data
}
\description{
An intermediate result provided for the convenience of vignette users. 
Data from the midfielddata package is reduced to the observations needed 
to determine stickiness for the case study programs of  Civil, 
Electrical, Industrial/Systems, and Mechanical Engineering.
}
\seealso{
Other case-study-data: 
\code{\link{study_grad_rate}},
\code{\link{study_program}},
\code{\link{study_stickiness}},
\code{\link{study_student}}
}
\concept{case-study-data}
\keyword{datasets}
