% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_reduced_ever}
\alias{study_reduced_ever}
\title{Case-study ever-enrolled data}
\format{
\code{data.table} with 7388 rows and 4 columns keyed by ID.

\describe{

 \item{mcid}{Character, anonymized student identifier}

 \item{program}{Character, abbreviation indicating a program in 
 which the student was ever enrolled}

 \item{race}{Character, race/ethnicity as self-reported by the student, 
 e.g., Asian, Black, Hispanic/LatinX, etc.}
 
 \item{sex}{Character, sex as self-reported by the student, 
 values are Female, Male, and Unknown}

}
}
\usage{
study_reduced_ever
}
\description{
An intermediate result provided for the convenience of vignette users. 
Data from the midfielddata package is reduced to observations of
all students ever enrolled in the case study programs of  Civil, 
Electrical, Industrial/Systems, and Mechanical Engineering.
}
\seealso{
Other case-study-data: 
\code{\link{study_grad_rate}},
\code{\link{study_program}},
\code{\link{study_reduced_grad}},
\code{\link{study_stickiness}},
\code{\link{study_student}}
}
\concept{case-study-data}
\keyword{datasets}
