% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_reduced_grad}
\alias{study_reduced_grad}
\title{Case-study graduation data}
\format{
\code{data.table} with 3435 rows and 4 columns keyed by ID.

\describe{

 \item{mcid}{Character, anonymized student identifier}

 \item{program}{Character, abbreviation indicating a program the 
 student completed}

 \item{race}{Character, race/ethnicity as self-reported by the student, 
 e.g., Asian, Black, Hispanic/LatinX, etc.}
 
 \item{sex}{Character, sex as self-reported by the student, 
 values are Female, Male, and Unknown}

}
}
\usage{
study_reduced_grad
}
\description{
An intermediate result provided for the convenience of vignette users. 
Data from the midfielddata package is reduced to observations of
all students graduating in timely fashion from the case study programs of  
Civil, Electrical, Industrial/Systems, and Mechanical Engineering.
}
\seealso{
Other case-study-data: 
\code{\link{study_grad_rate}},
\code{\link{study_program}},
\code{\link{study_reduced_ever}},
\code{\link{study_stickiness}},
\code{\link{study_student}}
}
\concept{case-study-data}
\keyword{datasets}
