% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_starting_pool}
\alias{study_starting_pool}
\title{Case-study observations starting pool}
\format{
\code{data.table} with 161,696 rows and 2 columns keyed by ID.

\describe{

 \item{mcid}{Character, anonymized student identifier}

 \item{cip6}{Character, 6-digit CIP codes of programs in which students
 are ever enrolled.}

}
}
\usage{
study_starting_pool
}
\description{
An intermediate result provided for the convenience of vignette users. 
Data from the midfielddata package is processed to obtain observations 
of students ever enrolled in the case study programs of Civil, Electrical, 
Industrial/Systems, and Mechanical Engineering without filtering for 
data sufficiency or timely completion. Keyed by student ID.
}
\seealso{
Other case-study-data: 
\code{\link{study_stickiness_observ}},
\code{\link{study_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
