% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_stickiness}
\alias{study_stickiness}
\title{Case-study stickiness results}
\format{
\code{data.table} with 31 rows and 3 columns keyed by
program, race/ethnicity, and sex.
\describe{
 \item{program}{character, instructional programs selected for study}
 \item{race_sex}{character, student race/ethnicity ans sex}
 \item{stick}{numerical, program stickiness, a persistence metric}
}
}
\usage{
study_stickiness
}
\description{
Stickiness metric results for Civil, Electrical,
Industrial, and Mechanical Engineering programs from midfielddata. Results
are grouped by program, race/ethnicity, and sex. Used in the case study
developed in the vignettes.
}
\details{
Stickiness of a program is the ratio of the number of students graduating
from a program to the number of students ever enrolled in the program.

With a quantitative response value for each combination of levels of two
categorical variables, these are multiway data. However, some responses
are NA where observations didn't exist or were removed because of small
population size.

Data have also been subset to remove ambiguous levels of race/ethnicity.
}
\seealso{
Other case-study-data: 
\code{\link{study_programs}},
\code{\link{study_students}}
}
\concept{case-study-data}
\keyword{datasets}
