% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_stickiness_observ}
\alias{study_stickiness_observ}
\title{Case-study stickiness observations}
\format{
\code{data.table} with 11212 rows and 5 columns keyed by ID.

\describe{

 \item{mcid}{Character, anonymized student identifier}

 \item{program}{Character, abbreviation indicating a program the 
 student completed}
 
 \item{group}{Character, indicating the grouping (\code{ever} or 
 \code{grad}) to which the student belongs. Graduate IDs are a subset of 
 ever enrolled IDs.} 

 \item{race}{Character, race/ethnicity as self-reported by the student, 
 e.g., Asian, Black, Hispanic/LatinX, etc.}
 
 \item{sex}{Character, sex as self-reported by the student, 
 values are Female, Male, and Unknown}

}
}
\usage{
study_stickiness_observ
}
\description{
An intermediate result provided for the convenience of vignette users. 
Data from the midfielddata package is processed to obtain observations of
students ever enrolled in, and those graduating in timely fashion from,
the case study programs of Civil, Electrical, Industrial/Systems, and 
Mechanical Engineering. Keyed by student ID.
}
\seealso{
Other case-study-data: 
\code{\link{study_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
