% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_student}
\alias{study_student}
\title{Case-study students}
\format{
\code{data.table} with 8219 rows and 8 columns keyed by
student ID. The variables are:
\describe{

  \item{mcid}{Character, de-identified student ID.}

  \item{institution}{Character, anonymized institution name, e.g.,
  Institution A, Institution B, etc.}

  \item{cip6}{Character, 6-digit code of a program in which a student
  is enrolled.}

  \item{program}{Character, program name assigned for grouping and
  summarizing and for display in graphs and tables.}

  \item{race}{Character, race/ethnicity as self-reported by the student,
  e.g., Asian, Black, Hispanic/LatinX, etc.}

  \item{sex}{Character, sex as self-reported by the student, values are
  Female, Male, and Unknown.}

  \item{timely_term}{Character, the last academic term in which program
  completion would be considered timely for a given student,
  format YYYYT.}

  \item{data_sufficiency}{Logical, indicating whether the available data
  include a sufficient range of years to justify including a student in
  an analysis. Because the data sufficiency criterion has already been
  evaluated, the values should all be TRUE.}

}
}
\usage{
study_student
}
\description{
Data frame of student attributes in the Civil, Electrical, Industrial,
and Mechanical Engineering programs from midfielddata. These data are
subset for data sufficiency and can be used as the starting point for
different persistence metrics such as stickiness or graduation rate.
}
\seealso{
Other case-study-data: 
\code{\link{study_grad_rate}},
\code{\link{study_program}},
\code{\link{study_stickiness_observ}},
\code{\link{study_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
