% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_matriculant.R
\name{subset_matriculant}
\alias{subset_matriculant}
\title{Subset a data frame for matriculants}
\usage{
subset_matriculant(dframe, ..., record = NULL)
}
\arguments{
\item{dframe}{data frame with an ID column}

\item{...}{not used, forces later arguments to be used by name}

\item{record}{data frame of student attributes}
}
\value{
Data frame with the following properties:
\itemize{
    \item Rows with matching IDs in \code{record}
    \item Columns are not modified
    \item Data frame attributes \code{tbl} or \code{data.table}
          are preserved
    \item Grouping structures are not preserved
}
}
\description{
Subset a data frame by student ID, retaining students who are present in
the matriculation record.
}
\details{
An inner join between \code{dframe} and the ID column of \code{record}
retains all rows of \code{dframe} with matching IDs.
}
\examples{
# TBD
}
\seealso{
Other functions: 
\code{\link{add_race_sex}()},
\code{\link{add_timely_limit}()},
\code{\link{eval_fair_record}()},
\code{\link{fair_record_length}()},
\code{\link{filter_text}()},
\code{\link{prepare_fye_mi}()},
\code{\link{prepare_multiway}()},
\code{\link{subset_feasible}()}
}
\concept{functions}
