% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_stickiness.R
\name{tally_stickiness}
\alias{tally_stickiness}
\title{Computes stickiness}
\usage{
tally_stickiness(ever, grad)
}
\arguments{
\item{ever}{A data frame}

\item{grad}{A data frame}
}
\value{
A data frame

The two input data frames are joined by the set of variables they have in common.
}
\description{
Given the numbers of students graduating from a program and the number of students ever enrolled in a program, the ratio of the two numbers is stickiness.
}
