% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_addition.R
\name{term_addition}
\alias{term_addition}
\title{Add terms}
\usage{
term_addition(data = NULL, ..., term_col = NULL, add_col = NULL)
}
\arguments{
\item{data}{data frame}

\item{...}{not used for values. Forces optional arguments to be usable
by name only.}

\item{term_col}{character scalar, name of a term column, numeric format YYYYT}

\item{add_col}{character scalar, name of a column of terms to add, numeric
format T}
}
\value{
\code{data.frame}
}
\description{
Term arithmetic
}
\examples{
# placeholder

}
\seealso{
Other data_carpentry: 
\code{\link{feasible_completion}()},
\code{\link{label_programs}()},
\code{\link{order_multiway}()}
}
\concept{data_carpentry}
