% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_qual_focus.R
\name{term_qual_focus}
\alias{term_qual_focus}
\title{Categorize the qualification focus of a term}
\usage{
term_qual_focus(dframe, midf_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \code{{mcid}} and one of
\verb{\{term, term_course, term_degree\}.}}

\item{midf_table}{\code{degree} data frame with required variables
\verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added unless they are redundant (see Details). The new variables are:
\itemize{
\item \code{bacc_term}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA}.
Joined from the \code{term_degree} variable in \code{midf_table.}
\item \code{term_focus}    Character. Indicating a term contributes to study
before or after a student's first baccalaureate.
Possible values are "undergrad" and "post-bacc."
}
}
}
\description{
Determine the qualification focus (undergraduate or post-baccalaureate) of every term for each student in a data frame and add columns to the data frame to support the finding.
}
\details{
Every term in a student's record can be categorized as "undergraduate" if
the term predates their first degree and "post-baccalaureate" if it postdates
the degree. Post-baccalaureate terms are
typically excluded from the \verb{term, course,} and \code{degree} data tables.

\emph{Redundant columns.} To prevent overwriting, a variable such as
\code{bacc_term}  is not added to the data frame if it duplicates
an existing variable. The test for redundancy is managed internally
by calling \code{rm_redundant_cols()} before the final data frame is
returned. For documentation, see \code{?rm_redundant_cols}.
}
\examples{
# Select min required columns
term <- toy_term[, .(mcid, term)]
course <- toy_course[, .(mcid, term_course)]
degree <- toy_degree[, .(mcid, term_degree)]

# Example of starting data frame
term

# Labeling terms by group: undergrad & grad
term <- term_qual_focus(term, midf_table = degree)
course <- term_qual_focus(course, midf_table = degree)
degree <- term_qual_focus(degree, midf_table = degree)

# Example result
term[order(-term_focus)]

# No change if added columns duplicate existing
x <- term_qual_focus(term, midf_table = degree)
check_equiv_frames(term, x)

# Filter to retain "undergraduate" rows only
term <- term[term_focus == "undergrad"]
course <- course[term_focus == "undergrad"]
degree <- degree[term_focus == "undergrad"]

# Example result
term

}
