% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{term_separate}
\alias{term_separate}
\title{Separate YYYYT into year and term}
\usage{
term_separate(data = NULL, col = NULL)
}
\arguments{
\item{data}{Data frame having a YYYYT term variable.}

\item{col}{Column name in quotes of the term variable in \code{data}.}
}
\description{
Internal function for separating a MIDFIELD term variable (encoding YYYYT)
into a YYYY column and a T column.
}
\details{
The original columns in \code{data} remain. Two columns are added, one for
the year YYYY and the second for the term T.
}
\keyword{internal}
