% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_summarize.R
\name{term_summarize}
\alias{term_summarize}
\alias{term_summarise}
\title{Determine range of terms}
\usage{
term_summarize(data = NULL, ..., institution = "institution", term = "term")

term_summarise(data = NULL, ..., institution = "institution", term = "term")
}
\arguments{
\item{data}{Data frame. Default is \code{midfieldterms}.}

\item{...}{Not used for values. Forces the subsequent optional arguments
to only be usable by name.}

\item{institution}{Column name in quotes of the institution variable in
\code{data}. Default is "institution". Optional argument.}

\item{term}{Column name in quotes of the term variable in
\code{data}. Default is "term". Optional argument.}
}
\value{
Data frame with minimum and maximum range of terms by institution.
}
\description{
Determine first and last academic terms for each institution in a data frame.
}
\details{
The default values of the optional arguments are the column names
\code{"institution"} and \code{"term"}.
If using a data frame with different column names, you can rename
your variables to match the defaults or use the optional arguments to
pass your variable name(s) to the function.

\code{term_summarise()} is synonymous with \code{term_summarize()}.
}
\examples{
library(midfielddata)
term_summarize(midfieldterms)
}
\seealso{
Other data_carpentry: 
\code{\link{cip6_select}()},
\code{\link{cip_filter}()},
\code{\link{ever_filter}()},
\code{\link{grad_filter}()},
\code{\link{multiway_order}()},
\code{\link{race_sex_join}()},
\code{\link{start_filter}()}
}
\concept{data_carpentry}
