% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timely_term.R
\name{timely_term}
\alias{timely_term}
\title{Determine timely completion terms}
\usage{
timely_term(dframe, midf_table = term, ..., span = NULL, sched_span = NULL)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variable \verb{\{mcid\}.}}

\item{midf_table}{\code{term} data frame with required variables
\verb{\{mcid, term, level\}.}}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{span}{Integer scalar (default 6), number of years to define timely
completion, typically 4, 6, or 8 years (100\%, 150\%, 200\% respectively
of \code{sched_span}).#'}

\item{sched_span}{Integer scalar (default 4), the number of years an
institution officially schedules for completing a program.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added unless they are redundant (see Details). The new variables are:
\itemize{
\item \code{entry_term}   Character. Initial term of a student's longitudinal
record, encoded \code{YYYYT}. Extracted from \code{midf_table.}
\item \code{entry_level}   Character. Student level (01 Freshman, 02 Sophomore,
etc.) in their initial term. Extracted from \code{midf_table.}
\item \code{adj_span}   Numeric. Integer span of years for timely
completion adjusted for a student's initial level.
\item \code{timely_term}   Character. Latest term by which program completion
would be considered timely. Encoded \code{YYYYT.}
}
}
}
\description{
Determines the latest term by which program
completion would be considered timely. Adds columns to the data frame that support the
finding.
}
\details{
Completing an academic program in a "timely" manner means that a student
completes the requirements for a degree within a set time span, typically
4, 6, or 8 years after admission depending on the definition adopted in a
particular study. The final term of that span is the
\emph{timely completion term.}

Our heuristic assigns a time span of 6 academic years for timely completion
(other values can be assigned via the \code{span} argument). For students
admitted at second-year level or higher, the span value is reduced by
one academic year for each full year the student is assumed to have
completed. The adjusted span is added to their initial term at an
institution to create the \code{timely_term} value for each observation.

\emph{Redundant columns.} To prevent overwriting, a variable such as
\code{timely_term}  is not added to the data frame if it duplicates
an existing variable. The test for redundancy is managed internally
by calling \code{rm_redundant_cols()} before the final data frame is
returned.
}
\examples{
# Assign toy data sets
student <- toy_student
term <- toy_term

# Start with a selected population
x <- student[c(9:11, 21:30, 344:345), .(mcid)]
x

# Add timely term columns
x <- timely_term(x, midf_table = term)
x

# No change if added columns duplicate existing
y <- timely_term(x, midf_table = term)
check_equiv_frames(x, y)
}
