% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_course}
\alias{toy_course}
\title{Small 'course' dataset for examples}
\format{
Data frame with 8950 rows and 12 columns (\code{data.table} class).
Composite key: \verb{\{mcid, term_course, abbrev, number\}.}
\describe{
\item{\code{mcid}}{Character. Anonymized student identifier
that connects the four data tables, e.g., "MCID3111142897."}
\item{\code{term_course}}{Character. Academic year and term,
encoded \code{YYYYT.}}
\item{\code{abbrev}}{Character. Course alphabetical identifier,
e.g. "ENGR", "MATH", "ENGL."}
\item{\code{number}}{Character. Course numeric identifier, e.g.
"101", "3429."}
\item{\code{institution}}{Character. The anonymized name of
the institution the student attended in a given term, e.g.,
"Institution A", "Institution B", etc.}
\item{\code{course}}{Character. Course name, e.g., "Astrophysics
III", "Calculus For Social Science And Business", "Corp Financial
Rprtng 1", "Environmental Sanitation II", "Fitness and Wellness",
"Introductory Astronomy 2", "Our Changing Environment", etc.}
\item{\code{section}}{Character. Course section identifier, from
one to four characters, e.g., "1", "2", "01", "14", "001", "040", "785",
"H02", "R01", "300E", "888R", etc.}
\item{\code{type}}{Character. Predominant delivery method for this
section, e.g., "Blended", "Distance Education", "Face-to-Face",
"Online", etc.}
\item{\code{faculty_rank}}{Character. The academic rank of the
person teaching the course, e.g., "Assistant Professor", "Associate
Professor", "Graduate Assistant", "Visiting Faculty", etc.}
\item{\code{hours_course}}{Numeric. Number of credit-hours
for successful course completion.}
\item{\code{grade}}{Character. Course grade, e.g., "A+", "A", "A-",
"B+", "I", "NG", etc.}
\item{\code{discipline_midfield}}{Character. A variable
for grouping courses by academic discipline assigned by the pre-2023
MIDFIELD data curator, e.g., "Anthropology", "Business", "Computer
Science", "Engineering", "Language and Literature", "Mathematics",
"Visual and Performing Arts", etc.}
}
}
\usage{
toy_course
}
\description{
A subset of rows from the midfielddata \code{course} table matching the IDs in
\code{toy_student.} A small dataset for use in examples.
}
\seealso{
Other toy-data:
\code{\link{toy_degree}},
\code{\link{toy_student}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
