% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_student}
\alias{toy_student}
\title{Small 'student' dataset for examples}
\format{
Data frame with 351 rows and 13 columns (\code{data.table} class).
Key: \verb{\{mcid\}.}
\describe{
\item{\code{mcid}}{Character. Anonymized student identifier
that connects the four data tables, e.g., "MCID3111142897."}
\item{\code{race}}{Character. Race/ethnicity as self-reported
by the student, e.g., "Asian", "Black", "Hispanic", etc.}
\item{\code{sex}}{Character. Sex as self-reported by the student,
possible values are "Female", "Male", and "Unknown."}
\item{\code{institution}}{Character. The anonymized name of
the institution the student attended in a given term, e.g.,
"Institution A", "Institution B", etc.}
\item{\code{transfer}}{Character. Transfer status, possible
values are "First-Time in College", "First-Time Transfer."}
\item{\code{hours_transfer}}{Numeric. Number of credit
hours transferred (or NA).}
\item{\code{age_desc}}{Character. Age group, possible values
are "25 and Older", "Under 25."}
\item{\code{us_citizen}}{Character. US citizenship, possible
values are "No", "Yes."}
\item{\code{home_zip}}{Character. Home ZIP code (or NA),
e.g., "02056", "20170", "51301", "80129", etc.}
\item{\code{high_school}}{Character. Code for the last high
school attended before admission (or NA), e.g., "060075",
"210512", "431800", "502195", etc.}
\item{\code{sat_math}}{Numeric. SAT mathematics test score
(or NA).}
\item{\code{sat_verbal}}{Numeric. SAT reading test score
(or NA).}
\item{\code{act_comp}}{Numeric. ACT composite test score
(or NA).}
}
}
\usage{
toy_student
}
\description{
A subset of rows from the midfielddata \code{student} table. A small dataset
for use in examples.
}
\seealso{
Other toy-data:
\code{\link{toy_course}},
\code{\link{toy_degree}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
