% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_term}
\alias{toy_term}
\title{Small 'term' dataset for examples}
\format{
Data frame with 1821 rows and 13 columns (\code{data.table} class).
Composite key: \verb{\{mcid, term\}.}
\describe{
\item{\code{mcid}}{Character. Anonymized student identifier
that connects the four data tables, e.g., "MCID3111142897."}
\item{\code{term}}{Character. Academic year and term the student
attended, encoded \code{YYYYT.}}
\item{\code{cip6}}{Character. The 6-digit CIP code of the program in
which a student is enrolled in this term.}
\item{\code{institution}}{Character. The anonymized name of
the institution the student attended in a given term, e.g.,
"Institution A", "Institution B", etc.}
\item{\code{level}}{Character. Academic level of the student at the
end of this term, e.g., "01 First-Year", "02-Second Year", etc.}
\item{\code{standing}}{Character. Academic standing during the
reported term, e.g., "Good Standing", "Academic Warning", etc.}
\item{\code{coop}}{Character. Cooperative education term, possible
values are "Yes", "No."}
\item{\code{hours_term}}{Numeric. Credit hours earned
in the term.}
\item{\code{hours_term_attempt}}{Numeric. Credit
hours attempted in the term.}
\item{\code{hours_cumul}}{Numeric, cumulative credit hours
earned.}
\item{\code{hours_cumul_attempt}}{Numeric. Cumulative
credit hours attempted.}
\item{\code{gpa_term}}{Numeric. Term grade point average.}
\item{\code{gpa_cumul}}{Numeric. Cumulative grade point average.}
}
}
\usage{
toy_term
}
\description{
A subset of rows from the midfielddata \code{term} table matching the IDs in
\code{toy_student.} A small dataset for use in examples.
}
\seealso{
Other toy-data:
\code{\link{toy_course}},
\code{\link{toy_degree}},
\code{\link{toy_student}}
}
\concept{toy-data}
\keyword{datasets}
