% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_ethnicity_sex.R
\name{unite_ethnicity_sex}
\alias{unite_ethnicity_sex}
\title{Unite ethnicity and sex variables}
\usage{
unite_ethnicity_sex(x)
}
\arguments{
\item{x}{A data frame with at least two character variables \code{ETHNIC} and \code{SEX}.}
}
\value{
A data frame with a new variable \code{PEER_GROUP}. Other variables in the data frame are returned unaffected.
}
\description{
Creates a character variable \code{PEER_GROUP} by uniting the characters in the \code{ETHNIC} and \code{SEX} variables.
}
