% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-graph.R
\name{utilities-graph}
\alias{expon_scale_x_log10}
\alias{midfield_theme}
\title{Graphics utility functions}
\usage{
expon_scale_x_log10()

midfield_theme()
}
\value{
Modifies an existing graph
}
\description{
Provides a set of customized layers for graphs created with for
\code{ggplot2}.
}
\details{
\describe{
\item{\code{expon_scale_x_log10()}}{Applies the
  \code{ggplot2::scale_x_log10()} function, marks the scale with
  powers of ten in exponential form, and displays the logarithmic minor
  grid lines.}
\item{\code{midfield_theme()}}{Applies \code{theme_minimal()} with some
  additional edits.}
}
}
