% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{utils_graph}
\alias{scale_x_log10_expon}
\alias{theme_midfield}
\title{Graphics utility functions}
\usage{
scale_x_log10_expon()

theme_midfield()
}
\value{
Modifies an existing graph.
}
\description{
A set of customized layers for ggplot2 graphs.
}
\details{
\describe{
\item{\code{scale_x_log10_expon()}}{Applies the
  \code{ggplot2::scale_x_log10()} function, marks the scale with
  powers of ten in exponential form, and displays the logarithmic minor
  grid lines.}
\item{\code{theme_midfield()}}{Applies \code{theme_minimal()} with
  additional edits: font family is "sans"; all text is 10 point; all
  lines are pale gray (#D9D9D9) and size 0.4. }
}
}
