% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{vignette_data}
\alias{vignette_data}
\alias{ever}
\title{Vignette data examples}
\format{A data frame}
\source{
MIDFIELD
}
\usage{
ever
}
\description{
Several tidy data sets used in the vignettes.
}
\details{
\describe{

\item{\code{ever}}{A data frame of students ever in a major.}

}
}
\examples{
ever
}
\keyword{datasets}
