% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_fill.R
\name{zero_fill}
\alias{zero_fill}
\title{Group and summarize a count of students}
\usage{
zero_fill(df)
}
\arguments{
\item{df}{A df frame of unique students}
}
\value{
df frame with a count of observations at the intersections of all variables in the \code{df} input df frame.
}
\description{
Counts the number of students grouped by all the variables in \code{df}. The count is assigned the variable name in \code{count_label}. If the variables \code{sex}, \code{race}, or \code{institution} are present, the function fills a zero for a missing count at every intersection of these variables.
}
