% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{act_sat_scale}
\alias{act_sat_scale}
\title{ACT-SAT conversion scale}
\format{
\code{data.table} with 28 rows and 4 columns:
\describe{
\item{\code{act_comp}}{Numerical, ACT composite score.}
\item{\code{sat_lower}}{Numerical, total SAT, lower limit of range
corresponding to the ACT composite score.}
\item{\code{sat_equiv}}{Numerical, total SAT, value to use when
converting ACT score to a single SAT score.}
\item{\code{sat_upper}}{Numerical, total SAT, upper limit of range
corresponding to the ACT composite score.}
}
}
\source{
ACT/SAT Concordance (2018) ACT Education Corp. \url{https://www.act.org/content/dam/act/unsecured/documents/ACT-SAT-Concordance-Tables.pdf}
}
\usage{
act_sat_scale
}
\description{
Data frame for converting between ACT and SAT scores. A range of SAT scores
converts to a single ACT score; an ACT score converts to a single
value equivalent SAT score.
}
\seealso{
Other scales: 
\code{\link{grade_scale}}
}
\concept{scales}
\keyword{datasets}
