% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_completion_status.R
\name{add_completion_status}
\alias{add_completion_status}
\title{Determine program completion status for every student}
\usage{
add_completion_status(dframe, midfield_degree = degree)
}
\arguments{
\item{dframe}{Data frame of student unit record (SUR) observations keyed 
by student ID. Required variables are \code{mcid} and 
\code{timely_term}.}

\item{midfield_degree}{Data frame of SUR degree observations keyed 
by student ID. Default is \code{degree}. Required variables are 
\code{mcid} and \code{term_degree}.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
 \item Rows are not modified.
 \item Grouping structures are not preserved.
 \item Columns listed below are added. \strong{Caution!} An existing column 
 with the same name as one of the added columns is silently overwritten. 
 Other columns are not modified. 
}
Columns added:
\describe{
 \item{\code{term_degree}}{Character. Term in which a program is completed. 
 Encoded YYYYT.}
 \item{\code{completion}}{Logical. TRUE denotes students completing their 
 programs.}
 \item{\code{completion_status}}{Character. Label each observation to 
 indicate program completion status. Possible values are: 
 "timely", indicating program completion no later than the timely 
 completion term; "untimely", indicating program completion after  
 the timely completion term; and "NA" indicating non-completion.}
}
}
\description{
Add columns to a data frame of Student Unit Record (SUR) 
observations that indicate whether a student completed their program, 
and if so, whether their completion was timely. Requires a MIDFIELD 
\code{degree} data frame in the environment.
}
\details{
By "program completion" we mean an undergraduate earning their 
baccalaureate degree. In many studies, students must complete their 
programs in a specified time span, for example 4-, 6-, or 8-years after 
admission. If they do, their completion is timely; if not, their completion 
is untimely and they are grouped with the non-completers when computing 
a metric such as graduation rate. 

Completion status is "timely" for students completing their programs 
no later than their timely completion terms. See also 
\code{add_timely_term()}.
}
\examples{
# Start with an excerpt from the student data set 
dframe <- toy_student[1:10, .(mcid)]

# Timely term column is required to add completion status column
dframe <- add_timely_term(dframe, midfield_term = toy_term)

# Add completion status column
add_completion_status(dframe, midfield_degree = toy_degree)

# Existing completion_status column, if any, is overwritten
dframe[, completion_status := NA_character_]
add_completion_status(dframe, midfield_degree = toy_degree)
}
\seealso{
Other add_*: 
\code{\link{add_data_sufficiency}()},
\code{\link{add_timely_term}()}
}
\concept{add_*}
