% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_completion_status.R
\name{add_completion_status}
\alias{add_completion_status}
\title{Determine program completion status for every student}
\usage{
add_completion_status(dframe, midfield_degree = degree)
}
\arguments{
\item{dframe}{Data frame of student unit record (SUR) observations keyed 
by student ID. Required variables are \code{mcid} and 
\code{timely_term}.}

\item{midfield_degree}{Data frame of SUR degree observations keyed 
by student ID. Default is \code{degree}. Required variables are 
\code{mcid} and \code{term_degree}.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
 \item Rows are not modified.
 \item Grouping structures are not preserved.
 \item Columns listed below are added. \strong{Caution!} An existing column 
 with the same name as one of the added columns is silently overwritten. 
 Other columns are not modified. 
}
Columns added:
\describe{
 \item{\code{term_degree}}{Character. Term in which a program is completed. 
 Encoded YYYYT.}
 \item{\code{completion}}{Logical. TRUE denotes students completing their 
 programs.}
 \item{\code{completion_status}}{Character. Label each observation to 
 indicate program completion status. Possible values are: 
 \code{positive}, indicating programs completed no later than their timely 
 completion term; and \code{negative}, indicating programs never completed 
 as well as programs completed after their timely completion terms.}
}
}
\description{
Add columns to a data frame of Student Unit Record (SUR) 
observations that indicate whether a student completed their program, 
and if so, whether their completion was timely. Requires a MIDFIELD 
\code{degree} data frame in the environment.
}
\details{
By "program completion" we mean an undergraduate earning their 
baccalaureate degree. In many studies, students must complete their 
programs in a specified time span, for example 4-, 6-, or 8-years after 
admission. If they do, their completion is timely; if not, their completion 
is untimely and they are grouped with the non-completers when computing 
a metric such as graduation rate. 

Completion status is "positive" for students completing their programs 
no later than their timely completion terms. See also 
\code{add_timely_term()}.
}
\examples{
# Start with IDs and add institution and timely term
# dframe <- toy_student[1:10, .(mcid)]
# dframe <- add_institution(dframe, midfield_term = toy_term)
# dframe <- add_timely_term(dframe, midfield_term = toy_term)
# 
# 
# # Timely completion column without detail
# add_completion_timely(dframe, midfield_degree = toy_degree)
# 
# 
# # Timely completion column with detail
# add_completion_timely(dframe, midfield_degree = toy_degree, detail = TRUE)
# 
# 
# # If present, existing completion_timely column is overwritten
# DT1 <- add_completion_timely(dframe, midfield_degree = toy_degree)
# DT2 <- add_completion_timely(DT1, midfield_degree = toy_degree)
# all.equal(DT1, DT2)

}
\seealso{
Other add_*: 
\code{\link{add_data_sufficiency}()},
\code{\link{add_institution}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()}
}
\concept{add_*}
