% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_completion_timely.R
\name{add_completion_timely}
\alias{add_completion_timely}
\title{Add a column to evaluate timely completion}
\usage{
add_completion_timely(dframe, midfield_table = NULL, ..., details = NULL)
}
\arguments{
\item{dframe}{data frame with required variables
\code{mcid} and \code{timely_term}}

\item{midfield_table}{MIDFIELD degree data, default \code{midfielddata::degree},
with required variables \code{mcid} and \code{term}}

\item{...}{not used, forces later arguments to be used by name}

\item{details}{logical scalar to add columns reporting information on
which the evaluation is based, default FALSE}
}
\value{
A \code{data.table} with the following properties:
\itemize{
    \item Rows are not modified
    \item Column \code{completion_timely} is added, columns
          \code{completion} and \code{term_degree} are added optionally
    \item Grouping structures are not preserved
}
}
\description{
A logical variable is added to a data frame indicating whether a student
has completed their program in a timely manner.
}
\details{
Program completion is typically considered timely if it occurs within a
given span of years after matriculation. In a persistence metric involving
graduation, students with timely completion should be grouped as graduates.
Students whose completion is not timely should be grouped with the
non-graduates.

The data frame argument must include the \code{timely_term} column
obtained using the \code{add_timely_term()} function. Completion is
considered timely if: 1) the student has completed a program; and 2) the
degree term is no later than the timely completion limit.

If \code{details} is TRUE, additional column(s) that support the finding
are returned as well. Here the extra columns are \code{completion}
indicating by TRUE/FALSE if the student completed their program and
\code{term_degree} from the degree table giving the term in
which the first degree(s), if any, was earned.
}
\examples{
# TBD
}
\seealso{
Other functions: 
\code{\link{add_data_sufficiency}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()},
\code{\link{condition_fye}()},
\code{\link{condition_multiway}()},
\code{\link{filter_by_key}()},
\code{\link{filter_text}()}
}
\concept{functions}
