% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_completion_timely.R
\name{add_completion_timely}
\alias{add_completion_timely}
\title{Add a column to evaluate program completion}
\usage{
add_completion_timely(dframe, midfield_table, ..., details = NULL)
}
\arguments{
\item{dframe}{Data frame with required variables
\code{mcid} and \code{timely_term}.}

\item{midfield_table}{MIDFIELD \code{degree} data table or equivalent with
required variables \code{mcid} and \code{term}.}

\item{...}{Not used, forces later arguments to be used by name.}

\item{details}{Optional flag to add columns reporting information
on which the evaluation is based, default FALSE.}
}
\value{
A \code{data.table} with the following properties:
\itemize{
    \item Rows are not modified.
    \item Column \code{completion_timely} is added with an option to add
          columns \code{completion} and \code{term_degree}.
    \item Grouping structures are not preserved.
}
}
\description{
Add a column of logical values (TRUE/FALSE) to a data frame indicating
whether a student completes their program in a timely manner. Based on
information in the MIDFIELD \code{degree} data table or equivalent.
}
\details{
Program completion is typically considered timely if it occurs within a
specific span of years after admission. In a persistence metric that depends
on program completion (graduation), only students whose program completion
is timely are counted as graduates; students whose program completion
is un-timely (taking longer than the specific span to complete) are counted
as non-graduates.

The input data frame \code{dframe} must include the \code{timely_term}
column obtained using the \code{add_timely_term()} function. Completion is
considered timely if: 1) the student has completed a program; and 2) the
degree term is no later than the estimated timely completion term. The
function itself performs no subsetting.

If \code{details} is TRUE, additional column(s) that support the finding
are returned as well. Here the extra columns are \code{completion}
indicating by TRUE/FALSE if the student completed their program and
\code{term_degree} from the \code{degree} table giving the first term in
which degree(s), if any, are earned.

Existing columns with the same names as the added columns are overwritten.
}
\examples{
# TBD



}
\seealso{
Other add_*: 
\code{\link{add_data_sufficiency}()},
\code{\link{add_institution}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()}
}
\concept{add_*}
