% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_sufficiency.R
\name{add_data_sufficiency}
\alias{add_data_sufficiency}
\title{Add a column to evaluate data sufficiency}
\usage{
add_data_sufficiency(dframe, midfield_term, ..., detail = NULL)
}
\arguments{
\item{dframe}{Data frame with required variables
\code{institution} and \code{timely_term}.}

\item{midfield_term}{MIDFIELD \code{term} data table or equivalent
with required variables \code{institution} and \code{term}.}

\item{...}{Not used, forces later arguments to be used by name.}

\item{detail}{Optional flag to add columns reporting information
on which the evaluation is based, default FALSE.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
    \item Rows are not modified.
    \item Column \code{data_sufficiency} is added with an option to add
          column \code{inst_limit}.
    \item Grouping structures are not preserved.
}
}
\description{
Add a column of logical values (TRUE/FALSE) to a data frame indicating
whether the available data include a sufficient range of years to justify
including a student in an analysis. Obtains the information from the
MIDFIELD \code{term} data table or equivalent.
}
\details{
Program completion is typically considered timely if it occurs within a
specific span of years after admission. Students admitted too near the
last term in the available data are generally excluded from a study because
the data have insufficient range to fairly assess their records.

The input data frame \code{dframe} must include the \code{timely_term}
column obtained using the \code{add_timely_term()} function. Students can be
retained in a study if their estimated timely completion term is no later
than the last term in their institution's data.

If the result in the \code{data_sufficiency} column is TRUE, then then
student should be included in the research. If FALSE, the student should
be excluded before calculating any persistence metric involving program
completion (graduation). The function itself performs no subsetting.

If \code{detail} is TRUE, additional column(s) that support the finding
are returned as well. Here the extra column is \code{inst_limit}, the latest
term reported by the institution in the available data.

Existing columns with the same names as the added columns are overwritten.
}
\examples{
# Start with IDs and add institution and timely term
dframe <- toy_student[1:10, .(mcid)]
dframe <- add_institution(dframe, midfield_term = toy_term)
dframe <- add_timely_term(dframe, midfield_term = toy_term)


# Data sufficiency column
add_data_sufficiency(dframe, midfield_term = toy_term)


# Data sufficiency column with detail
add_data_sufficiency(dframe, midfield_term = toy_term, detail = TRUE)


# If present, existing data_sufficiency column is overwritten
# Using dframe from above,
DT1 <- add_data_sufficiency(dframe, midfield_term = toy_term)
DT2 <- add_data_sufficiency(DT1, midfield_term = toy_term)
all.equal(DT1, DT2)

}
\seealso{
Other add_*: 
\code{\link{add_completion_timely}()},
\code{\link{add_institution}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()}
}
\concept{add_*}
