% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_institution.R
\name{add_institution}
\alias{add_institution}
\title{Add column for institution}
\usage{
add_institution(dframe, midfield_table)
}
\arguments{
\item{dframe}{Data frame with required variable \code{mcid}.}

\item{midfield_table}{MIDFIELD term data table with required variables
\code{mcid}, \code{institutiion}, \code{term}.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
    \item Rows are not modified.
    \item Column \code{institution} is added.
    \item Grouping structures are not preserved.
}
}
\description{
Add variable \code{institution} from the MIDFIELD \code{term} table
to a data frame, using \code{mcid} as the join-by variable.
}
\details{
An existing variable with the same name, if any, is overwritten. If a
student is associated with more than one institution, the institution at
which they completed the most terms is returned.
}
\examples{
# TBD
}
\seealso{
Other functions: 
\code{\link{add_completion_timely}()},
\code{\link{add_data_sufficiency}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()},
\code{\link{condition_fye}()},
\code{\link{condition_multiway}()},
\code{\link{filter_match}()},
\code{\link{filter_search}()}
}
\concept{functions}
