% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_institution.R
\name{add_institution}
\alias{add_institution}
\title{Add a column of institution names}
\usage{
add_institution(dframe, midfield_term)
}
\arguments{
\item{dframe}{Data frame with required variable \code{mcid}.}

\item{midfield_term}{MIDFIELD \code{term} data table or equivalent with
required variables \code{mcid}, \code{institution}, \code{term}.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
    \item Rows are not modified.
    \item Column \code{institution} is added.
    \item Grouping structures are not preserved.
}
}
\description{
Add a column of character values with institution names (or labels) using
student ID as the join-by variable. Obtains the information from the
MIDFIELD \code{term} data table or equivalent. In the MIDFIELD practice
data, the labels are de-identified.
}
\details{
If a student is associated with more than one institution, the institution
at which they completed the most terms is returned. An existing column with
the same name as the added column is overwritten.
}
\examples{
# Extract a column of IDs from student
id <- toy_student[, .(mcid)]


# Add institutions from term
DT1 <- add_institution(id, midfield_term = toy_term)
head(DT1)
nrow(DT1)


# Will overwrite institution column if present
DT2 <- add_institution(DT1, midfield_term = toy_term)
head(DT2)
nrow(DT2)

}
\seealso{
Other add_*: 
\code{\link{add_data_sufficiency}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_completion}()},
\code{\link{add_timely_term}()}
}
\concept{add_*}
