% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_race_sex.R
\name{add_race_sex}
\alias{add_race_sex}
\title{Add columns for race/ethnicity and sex}
\usage{
add_race_sex(dframe, ..., record = NULL)
}
\arguments{
\item{dframe}{data frame}

\item{...}{not used, forces later arguments to be used by name}

\item{record}{data frame of student attributes, default midfieldstudents}
}
\value{
Data frame with the following properties:
\itemize{
    \item Rows are not modified
    \item Columns \code{race} and \code{sex} are added
    \item Data frame attributes \code{tbl} or \code{data.table}
          are preserved
    \item Grouping structures are not preserved
}
}
\description{
Add variables \code{race} and \code{sex} from \code{record} to a data
frame, using \code{id} is the join-by variable.
}
\details{
Existing variables with the same names, if any, are overwritten.
}
\examples{
# TBD

}
\seealso{
Other functions: 
\code{\link{add_timely_limit}()},
\code{\link{eval_fair_record}()},
\code{\link{fair_record_length}()},
\code{\link{filter_text}()},
\code{\link{prepare_fye_mi}()},
\code{\link{prepare_multiway}()},
\code{\link{subset_feasible}()},
\code{\link{subset_matriculant}()}
}
\concept{functions}
