% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_race_sex.R
\name{add_race_sex}
\alias{add_race_sex}
\title{Add two columns for student race/ethnicity and sex}
\usage{
add_race_sex(dframe, midfield_table)
}
\arguments{
\item{dframe}{Data frame with required variable \code{mcid}.}

\item{midfield_table}{MIDFIELD \code{student} data table or equivalent
with required variables \code{mcid}, \code{race}, and \code{sex}.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
    \item Rows are not modified.
    \item Columns \code{race} and \code{sex} are added.
    \item Grouping structures are not preserved.
}
}
\description{
Add two columns of character values of students' self-reported
race/ethnicity and sex using student ID as the join-by variable.
Based on information in the MIDFIELD \code{student} data table.
}
\details{
Existing columns with the same names as the added columns are overwritten.
}
\examples{


# Add race and sex to a data frame of graduates
dframe <- toy_degree[1:5, c("mcid", "cip6")]
add_race_sex(dframe = dframe, midfield_table = toy_student)


# Add race and sex to a data frame from the term table
dframe <- toy_term[21:26, c("mcid", "institution", "level")]
add_race_sex(dframe = dframe, midfield_table = toy_student)


}
\seealso{
Other add_*: 
\code{\link{add_completion_timely}()},
\code{\link{add_data_sufficiency}()},
\code{\link{add_institution}()},
\code{\link{add_timely_term}()}
}
\concept{add_*}
