% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_race_sex.R
\name{add_race_sex}
\alias{add_race_sex}
\title{Add columns for race/ethnicity and sex}
\usage{
add_race_sex(dframe, ..., midfield_table = NULL)
}
\arguments{
\item{dframe}{data frame with required variable \code{mcid}}

\item{...}{not used, forces later arguments to be used by name}

\item{midfield_table}{MIDFIELD \code{student} table, with required
variables \code{mcid}, \code{race}, and \code{sex}}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
    \item Rows are not modified
    \item Columns \code{race} and \code{sex} are added
    \item Grouping structures are not preserved
}
}
\description{
Add variables \code{race} and \code{sex} from \code{student} to a
data frame, using \code{mcid} as the join-by variable.
}
\details{
Existing variables with the same names, if any, are overwritten.
}
\examples{
# TBD
}
\seealso{
Other functions: 
\code{\link{add_completion_timely}()},
\code{\link{add_data_sufficiency}()},
\code{\link{add_timely_term}()},
\code{\link{filter_by_key}()},
\code{\link{filter_text}()},
\code{\link{prepare_fye}()},
\code{\link{prepare_multiway}()}
}
\concept{functions}
