% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_term_cluster.R
\name{add_term_cluster}
\alias{add_term_cluster}
\title{Identify rows of post-baccalaureate terms}
\usage{
add_term_cluster(dframe, midfield_degree = degree)
}
\arguments{
\item{dframe}{Working data frame of student-level records
to which a term-status column is to be added. Required
variables are \code{mcid} and a single term variable.}

\item{midfield_degree}{MIDFIELD \code{degree} data table or
equivalent with required variables \code{mcid} and \code{term_degree.}}
}
\value{
A data frame with the following properties: rows
are preserved; columns \code{first_degree_term} and \code{term_status} are added
or overwritten; all other columns are preserved.
Grouping structures are not necessarily preserved.
The added columns are:
\describe{
\item{\code{first_degree_term}}{Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} --- or NA if no degree recorded.}
\item{\code{term_cluster}}{Character. Possible values are "pre-degree",
"first-degree", and "post-first-degree".}
}

An attempt is made to return data frames of the same
class as the input, so a data.table input should produce a data.table
output, a tibble input should produce a tibble output, etc. Grouped
tibbles, however, are returned as data.frames. To help ensure a tibble
output, try ungrouping a tibble before using it as an input argument.
}
\description{
Provides a means of excluding post-baccalaureate terms by adding a
column of term-status labels identifying pre- and post-baccalaureate
terms.
}
\details{
In a typical analysis, one is interested in a student's progress up to
and including the term in which they earn their first degree or degrees.
Any terms later than the first baccalaureate can usually be excluded from
study.

The input \code{dframe} must have a term variable with one of three possible
names: \code{term} (from the term data table), \code{term_course} (from the course
data table), or \code{term_degree} (from the degree data table).

\code{add_term_cluster()} determines a student's first degree term (if any)
from \code{midfield_degree}, adds that column to \code{dframe}, and adds a second
column of term-status labels identifying pre- and post-baccalaureate terms.
}
\examples{

# Examples TBD
x <- 1
}
