% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_term_cluster.R
\name{add_term_cluster}
\alias{add_term_cluster}
\title{Identify term clusters}
\usage{
add_term_cluster(dframe, midfield_rec = degree)
}
\arguments{
\item{dframe}{Student-records data frame to which term-cluster columns
are to be added. Required variables are \code{mcid} and a single term
variable: \code{term} (when working with the term table),
\code{term_course} (course table), or \code{term_degree} (degree table).}

\item{midfield_rec}{MIDFIELD \code{degree} data table or equivalent with
required variables \code{mcid} and \code{term_degree}.}
}
\value{
Data frame \code{dframe} with added columns. Output has the
following properties:
\itemize{
\item Rows are not modified.
\item Columns added or updated/overwritten:
\itemize{
\item \code{first_degree_term.}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA}.
Joined from \code{midfield_rec$term_degree}.
\item \code{term_cluster.}    Character, indicating that a term belongs
to one of three clusters: terms that are prior to ("pre-degree"),
equal to ("first-degree"), or subsequent to ("post-first-degree")
the student’s first degree term.
}
\item Columns not listed above are not modified.
\item Data frame attributes (except groups) are preserved.
}
}
\description{
Add columns to a student-records data frame indicating an observation
should be included (or excluded) because the term is clustered with
pre-degree (or post-first-degree) terms.
}
\details{
In a typical analysis, one is interested in a student's progress up to
and including the term in which they earn their first degree or degrees.
Any terms later than the first baccalaureate can usually be excluded from
study.
}
\examples{

# Examples TBD
x <- 1
}
\seealso{
Other add_*:
\code{\link[=add_timely_term]{add_timely_term()}},
\code{\link[=completion_status]{completion_status()}},
\code{\link[=data_sufficiency]{data_sufficiency()}}
}
\concept{add_*}
