% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_timely_term.R
\name{add_timely_term}
\alias{add_timely_term}
\title{Calculate a timely completion term for every student}
\usage{
add_timely_term(
  dframe,
  midfield_term = term,
  ...,
  details = NULL,
  span = NULL,
  sched_span = NULL
)
}
\arguments{
\item{dframe}{Data frame of student unit record (SUR) observations keyed 
by student ID. Required variable is \code{mcid}.}

\item{midfield_term}{MIDFIELD term data frame keyed by student ID.
 Default is \code{term}. Required variables are \code{mcid}, 
\code{institution}, \code{term}, and \code{level}.}

\item{...}{Not used, forces later arguments to be used by name.}

\item{details}{Optional logical value. TRUE returns the additional 
variables used in determining the results. Default is FALSE.}

\item{span}{Optional numeric scalar, number of years to define timely
completion. Values that are 100\%, 150\%, and 200\% of the "scheduled
span" (\code{sched_span}) are commonly used. Default 6 years.}

\item{sched_span}{Optional numeric scalar, the number of years an
institution officially schedules for completing a program. Default
4 years.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
    \item Rows are not modified.
    \item Columns \code{institution} and \code{timely_term} are added; 
    additional columns are added via the \code{details} argument.  
    \item Grouping structures are not preserved.
}
}
\description{
Add a column of values to a data frame of Student Unit Record (SUR) 
observations indicating the latest term by which a student can complete  
their program and have it considered timely. An institution column is 
returned as well.
}
\details{
The basic heuristic starts with \code{span} number of years for each student
(default 6 years). The span for students admitted at a higher level than
first year are reduced by one year for each full year the student is
assumed to have completed. For example, a student admitted at the 
second-year level is assumed to have completed one year of a program, 
so their span is reduced by one year.

The adjusted span of years is added to their starting term; the result is
the timely completion term reported in the \code{timely_term} column.

The optional \code{details} argument returns the additional variables used 
in determining the results: a student's initial term and level and the 
adjusted span in years.
}
\section{Caution}{

Existing columns with the same names as the added columns are silently 
overwritten.
}

\examples{
# Using the toy data sets
# DT <- toy_student[1:10, .(mcid)]
# add_timely_term(DT, midfield_term = toy_term)
# 
# 
# # Add details on which the timely term is based
# add_timely_term(DT, midfield_term = toy_term, detail = TRUE)
# 
# 
# # Define timely completion as 200\% of scheduled span (8 years)
# add_timely_term(DT, midfield_term = toy_term, span = 8)
# 
# 
# # Optional arguments (after ...) must be named
# add_timely_term(DT, toy_term, detail = TRUE, span = 6)

}
\seealso{
Other add_*: 
\code{\link{add_data_sufficiency}()},
\code{\link{add_institution}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_completion}()}
}
\concept{add_*}
