% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_timely_term.R
\name{add_timely_term}
\alias{add_timely_term}
\title{Add column for timely completion term}
\usage{
add_timely_term(dframe, midfield_table, ..., details = NULL, span = NULL)
}
\arguments{
\item{dframe}{Data frame with required variable \code{mcid}.}

\item{midfield_table}{MIDFIELD term data table with required variables
\code{mcid}, \code{term}, and \code{level}.}

\item{...}{Not used, forces later arguments to be used by name.}

\item{details}{Logical scalar to add columns reporting information on
which the timely completion term is based, default FALSE.}

\item{span}{Numeric scalar, number of years to define timely completion,
default 6 years.}
}
\value{
A \code{data.table}  with the following properties:
\itemize{
    \item Rows are not modified.
    \item Column \code{timely_term} is added, columns \code{term_i},
          \code{level_i}, and \code{adj_span} are added optionally.
    \item Grouping structures are not preserved.
}
}
\description{
Given information about a student's academic path, determine the latest
term for which program completion could be considered timely.
}
\details{
The basic heuristic starts with \code{span} number of years for each student
(default 6 years). The span for students admitted at a higher level than
first year are reduced by one year for each full year the student is
assumed to have completed.

For example, a student admitted at the second-year level is assumed
to have completed one year of a program, so their span is reduced by one
year. Similarly, spans are reduced by two years for students admitted at
the 3rd-year level and by three years for students admitted at the
fourth-year level.

The adjusted span of years is added to their starting term; the result is
the timely completion reported in the \code{timely_term} column added to the
data frame.

If \code{details} is TRUE, additional column(s) that support the finding
are returned as well. Here the extra columns are the student's initial
(admission) term \code{term_i}, initial level \code{level_i}, and
the adjusted span \code{adj_span}.
}
\examples{
# TBD
}
\seealso{
Other functions: 
\code{\link{add_completion_timely}()},
\code{\link{add_data_sufficiency}()},
\code{\link{add_institution}()},
\code{\link{add_race_sex}()},
\code{\link{condition_fye}()},
\code{\link{condition_multiway}()},
\code{\link{filter_match}()},
\code{\link{filter_text}()}
}
\concept{functions}
