% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline_mcid}
\alias{baseline_mcid}
\title{Baseline ID bloc to start a typical analysis}
\format{
\code{data.table} with 76875 rows and 1 column:
\describe{
\item{\code{mcid}}{Character. Anonymized student identifier
that connects the four data tables, e.g., "MCID3111142897."}
}
}
\usage{
baseline_mcid
}
\description{
Data frame of IDs after processing the practice data for data sufficiency and
degree seeking. Provides a convenient bloc to start many of the analysis
illustrated in the package articles.
}
\seealso{
Other case-study-data:
\code{\link{study_observations}},
\code{\link{study_programs}},
\code{\link{study_results}}
}
\concept{case-study-data}
\keyword{datasets}
