% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio_sci}
\alias{bio_sci}
\title{CIP codes for biological and biomedical sciences}
\format{A character vector.}
\source{
CIP 2010, US National Center for Education Statistics
  (\url{https://nces.ed.gov/ipeds/cipcode}).
}
\usage{
bio_sci
}
\description{
Classification of Instructional Programs (CIP) 6-digit codes for biological
and biomedical science programs, series 26.
}
\examples{
head(bio_sci, n = 10L)

print(cip_filter(bio_sci))
}
\keyword{datasets}
