% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_fye}
\alias{case_fye}
\title{Imputed starting majors for FYE students}
\format{A data frame with columns:
\describe{
 \item{id}{Unique, anonymized MIDFIELD student identifier.}
 \item{cip6}{Imputed starting major (6-digit CIP code) of FYE students. }
}}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata} package
}
\usage{
case_fye
}
\description{
Imputed starting majors for 6000 first-year engineering (FYE) students in
the midfielddata package.
}
\details{
Some US institutions have first year engineering (FYE) programs---typically
a common first year curriculum that is a prerequisite for declaring an
engineering major. For some persistence metrics, we have to determine what
we call a "starting major" for FYE students---the predicted engineering
program such as Civil, Electrical, or Mechanical, the student would have
declared had they not been required to enroll in FYE.
}
\keyword{datasets}
