% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_program_group}
\alias{case_program_group}
\title{CIP case data for four programs}
\format{A tidy data frame (tibble) with 1546 observations and 6 variables.
All variables are characters. An observation is a unique program.
\describe{
  \item{cip2}{An instructional program's 2-digit code,  representing "the
  most general groupings of related programs" (US National Center for
  Education Statistics).}
  \item{cip2name}{Name of a program at the 2-digit level.}
  \item{cip4}{An instructional program's 4-digit code, representing
  "intermediate groupings of programs that have comparable content and
  objectives".}
  \item{cip4name}{Name of a program at the 4-digit level.}
  \item{cip6}{An instructional program's 6-digit code, representing
  "specific instructional programs".}
  \item{cip6name}{Name of a program at the 6-digit level}
  \item{program}{A program name.}
}}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata} package
}
\usage{
case_program_group
}
\description{
An example of CIP codes and names, a subset of \code{cip} data plus an added \code{program} variable for four engineering programs.
}
\details{
The CIP codes and names plus a user-selected program name for Civil, Electrical, Industrial, and Mechanical Engineering. The "Selecting programs" vignette describes how these data were collected. The data frame is used be several other vignettes as a starting point in selecting programs to study.
}
\keyword{datasets}
