% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_stickiness}
\alias{case_stickiness}
\title{Stickiness case data}
\format{
A data frame (tibble) with 32 observations and 6 variables.

\describe{
 \item{program}{Instructional programs selected for study.}
 \item{race}{Student race and sex from \code{midfieldstudents} data.}
 \item{sex}{Student sex from \code{midfieldstudents} data.}
 \item{ever}{Number of students ever enrolled in a program. Grouped and summarized from \code{midfieldterms} data.}
 \item{grad}{Number of students graduating from the program. Grouped and summarized from \code{midfielddegrees} data.}
 \item{stick}{Program stickiness: the ratio of the number of students graduating from a program to the number ever enrolled in the program.}
}
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata} package
}
\usage{
case_stickiness
}
\description{
An data frame of stickiness data grouped by program and student race/ethnicity and sex.
}
\details{
For Civil, Electrical, Industrial, and Mechanical Engineering, the number of students ever enrolled, the number of graduates, and the resulting stickiness.
}
\seealso{
The "Compute stickiness" vignette.
}
\keyword{datasets}
