% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_stickiness}
\alias{case_stickiness}
\title{Stickiness case data}
\format{A data frame with columns:
\describe{
 \item{program}{Instructional programs selected for study.}
 \item{race}{Student race from \code{midfieldstudents} data.}
 \item{sex}{Student sex from \code{midfieldstudents} data.}
 \item{ever}{Number of students ever enrolled in a program. Grouped and
 summarized from \code{midfieldterms} data.}
 \item{grad}{Number of students graduating from the program. Grouped and
 summarized from \code{midfielddegrees} data.}
 \item{stick}{Program stickiness: the ratio of the number of students
 graduating from a program to the number ever enrolled in the program.}
}}
\source{
A pre-processed data frame constructed following the steps described in the \code{stickiness} vignette.
}
\usage{
case_stickiness
}
\description{
An example of stickiness data
}
\examples{
\dontrun{
 case_stickiness
}
}
\keyword{datasets}
