% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catch_error}
\alias{catch_error}
\title{Error handling}
\usage{
catch_error(f)
}
\arguments{
\item{f}{Function with arguments expecting an error}
}
\value{
Does not return anything. The side effect is to output to the terminal.
}
\description{
A wrapper on \code{base::tryCatch()} for previewing an error message, if any.
}
\examples{
# Example data frames
sel_ids <- toy_student[14:18, (mcid)]

s <- toy_student[mcid \%chin\% sel_ids, .(mcid, sex)]
t <- toy_term[mcid \%chin\% sel_ids, .(mcid, term)]
d <- toy_degree[mcid \%chin\% sel_ids, .(mcid, term_degree)]

# No error
catch_error(post_bacc_terms(t, d))

# Error, no term variable 
catch_error(post_bacc_terms(s, d))

# Error, missing dframe argument
catch_error(post_bacc_terms())

# Error, missing degree argument
catch_error(post_bacc_terms(t))
}
