% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_error.R
\name{catch_error}
\alias{catch_error}
\title{Error handling}
\usage{
catch_error(f)
}
\arguments{
\item{f}{Function with arguments expecting an error}
}
\value{
Does not return anything. The side effect is to output to the terminal.
}
\description{
A wrapper on \code{base::tryCatch()} for previewing an error message, if any.
}
\examples{
# Example data frames
s <- toy_student[, .(mcid)]
t <- toy_term[, .(mcid, term)]
d <- toy_degree[, .(mcid, term_degree)]

# No error
catch_error(add_post_bacc(t, d))

# Error, no term variable 
catch_error(add_post_bacc(s, d))

# Error, missing dframe argument
catch_error(add_post_bacc())

# Error, missing degree argument
catch_error(add_post_bacc(t))

}
