% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip6_select.R
\name{cip6_select}
\alias{cip6_select}
\title{Select the 6-digit CIP code and name and add a program label}
\usage{
cip6_select(
  data,
  program = NULL,
  ...,
  cip6 = "cip6",
  cip6name = "cip6name",
  cip4name = "cip4name",
  cip2name = "cip2name"
)
}
\arguments{
\item{data}{Data frame of CIP codes and names with the same structure as midfieldr \code{cip}.}

\item{program}{Character variable. There are four options:

If program = NULL, the default, the 4-digit CIP program names are assigned to the new column.

If program = "named_series", then the series name is assigned to the new column.

If one of three strings ("cip2name", "cip4name", or "cip6name"), then the 2-digit, 4-digit, or 6-digit CIP program names are assigned to the new column.

Any other string fills the new column.}

\item{...}{Not used for values, forces later arguments to bind by name.}

\item{cip6}{Optional argument, the column name in quotes of the 6-digit CIP code variable in \code{data}. Default is "cip6".}

\item{cip6name}{Optional argument, the column name in quotes of the 6-digit CIP program name variable in \code{data}. Default is "cip6name".}

\item{cip4name}{Optional argument, the column name in quotes of the 4-digit CIP program name variable in \code{data}. Default is "cip4name".}

\item{cip2name}{Optional argument, the column name in quotes of the 2-digit CIP program name variable in \code{data}. Default is "cip2name".}
}
\value{
A data frame with the 6-digit code and name columns from the input data frame with an added \code{program} variable.
}
\description{
From a CIP data frame, select the columns with the 6-digit codes and names and add a new character variable named \code{program}.
}
\details{
Assigning a custom label to a program or a group of programs provides the option of grouping and summarizing CIP data in ways not afforded by the default program names in the 2010 CIP data.
}
\examples{
# Extract the Philosophy and Religion programs and label
x <- cip_filter(cip, series = "^38")
y <- cip6_select(x, program = "Phil-Relig")

# With no program argument, the 6-digit CIP names are assigned
y <- cip6_select(x)

# Argument options include
y <- cip6_select(x, program = "cip6name")
y <- cip6_select(x, program = "cip4name")
y <- cip6_select(x, program = "cip2name")

}
\seealso{
\code{\link[midfieldr]{cip_filter}} for exploring CIP codes.
}
