% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip_filter.R
\name{cip_filter}
\alias{cip_filter}
\title{Filter CIP data}
\usage{
cip_filter(data = NULL, series = NULL)
}
\arguments{
\item{data}{A data frame of character variables (or coercible to one).
If NULL, \code{data} is the midfieldr \code{cip} dataset.}

\item{series}{A character vector (or coercible to one) of search terms. The
conditions used to filter the data.}
}
\value{
A subset of \code{data} returned as a data frame (tibble) of
character variables.
}
\description{
Filter a data frame of Classification of Instructional Programs (CIP) codes
and return the rows that match conditions.
}
\details{
\code{cip_filter()} is essentially a search tool to filter the data frame
given by the \code{data} argument and return all rows containing the terms
specified in the \code{series} argument.

Variables are not altered.

Typical search terms include words and phrases describing academic programs
or all or part of 2-, 4-, or 6-digit CIP codes. midfieldr provides several
collections of predefined search terms for groups of programs such as
Biological Sciences, Engineering, Physical Sciences, etc.
}
\examples{
cip_filter()
cip_filter(series = "History")
cip_filter(series = c("American", "History"))
cip_filter(series = "History") \%>\%
  cip_filter(series = "American")
cip_filter(series = "^14")
cip_filter(series = "^14") \%>\%
  cip_filter(series = "Civil")

}
