% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip_filter.R
\name{cip_filter}
\alias{cip_filter}
\title{Filter CIP data}
\usage{
cip_filter(data = NULL, series = NULL)
}
\arguments{
\item{data}{A data frame of character variables (or coercible to one).
Required variables are \code{CIP2}, \code{CIP4}, and \code{CIP6}. If NULL,
\code{data} is the midfieldr \code{cip} dataset.}

\item{series}{The conditions used to filter the data. A character vector
(or coercible to one) of any combination of 2, 4, or 6-digit CIP codes.
If NULL, \code{data} is returned unchanged.}
}
\value{
A data frame: \code{data} filtered by the conditions in \code{series}.
}
\description{
Filter a data frame of Classification of Instructional Programs (CIP) codes
to return the rows that match conditions.
}
\details{
The purpose of \code{cip_filter()} is to obtain the CIP codes of specific
programs and use the results to obtain the records of students in those
programs.
}
\examples{
cip_filter(series = c("490101", "490205"))
cip_filter(series = c("^4901", "^4902"))
cip_filter(series = c("^49", "^99"))
cip_filter(series = c("^54", "^4901", "490205"))
cip_filter(series = seq(540102, 540108))

}
