% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip_filter.R
\name{cip_filter}
\alias{cip_filter}
\title{Filter by CIP data}
\usage{
cip_filter(series = NULL, ..., reference = NULL)
}
\arguments{
\item{series}{atomic character vector of search terms to filter by}

\item{...}{not used for values, forces later arguments to bind by name}

\item{reference}{a reference data frame from which 2-digit, 4-digit, and 6-digit CIP codes and program names are obtained, default \code{cip}}
}
\value{
data frame, subset of \code{reference}
}
\description{
Filter a data frame of Classification of Instructional Programs (CIP) codes and return the rows that match conditions.
}
\details{
The \code{series} argument is an atomic character vector of search terms  used to filter the reference CIP data set. Typical search terms include words and phrases describing academic programs or their CIP codes.

Several collections of named series are provided to facilitate searching for groups of programs such as Biological Sciences, Engineering, Physical Sciences, etc. For the named series help page, run \code{?cip_series}.

The function returns a subset of \code{reference} with all rows containing any of the search terms. Variables are not altered.
}
\examples{
cip_filter()
cip_filter(series = "History")
cip_filter(series = c("American", "History"))
cip_filter(series = "History") \%>\%
  cip_filter(series = "American", reference = .)
cip_filter(series = "^14")
cip_filter(series = "^14") \%>\%
  cip_filter(series = "Civil", reference = .)
}
