% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip_filter.R
\name{cip_filter}
\alias{cip_filter}
\title{Filter CIP data}
\usage{
cip_filter(data = NULL, keep_any = NULL, ..., drop_any = NULL)
}
\arguments{
\item{data}{Data frame or tibble of CIP codes, default \code{midfieldr::cip}.}

\item{keep_any}{Character vector used to filter \code{data}, keeping all rows in which any matches or partial matches are detected.}

\item{...}{Not used for values, forces later optional arguments to bind by name}

\item{drop_any}{Character vector used to filter \code{data}, dropping all rows in which any matches or partial matches are detected.}
}
\value{
Data frame
}
\description{
Filter a data frame of Classification of Instructional Programs (CIP) codes and return the rows that match conditions.
}
\details{
The \code{keep_any} argument is an atomic character vector of search terms  used to filter the CIP data frame. Typical search terms include words and phrases describing academic programs or their CIP codes. The optional \code{drop_any} argument is similar except it drops rows instead of keeping them.

Several named series are provided as possible values for \code{keep_any} to facilitate searching for groups of programs such as Biological Sciences, Engineering, Physical Sciences, etc. For the named series help page, run \code{? cip_series}.

The function returns a subset of \code{data}. All columns are retained.
}
\examples{
cip_filter(cip, "History")
cip_filter(cip, "History", drop_any = c("^04", "^13", "^50"))
cip_filter(cip, "History") \%>\%
  cip_filter("American")
cip_filter(cip, cip_engr) \%>\%
  cip_filter("Civil")
}
