% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip_label.R
\name{cip_label}
\alias{cip_label}
\title{Add a column of program labels to a CIP data frame}
\usage{
cip_label(data, program = NULL)
}
\arguments{
\item{data}{A filtered subset of the midfieldr \code{cip} data frame.}

\item{program}{An optional character variable. There are four options:

If NULL, the default, the 6-digit CIP program names are assigned to the new column.

If NULL, but one of the named \code{cip_series} is used to create the \code{data} argument, then the series name is assigned to the new column.

If one of three strings ("cip2name", "cip4name", or "cip6name"), then the 2-digit, 4-digit, or 6-digit CIP program names are assigned to the new column.

Any other string is assigned to the new column.}
}
\value{
The input data frame with an added \code{program} variable.
}
\description{
Adds a new character variable named \code{program} to a CIP data frame.
}
\examples{
# Extract the Philosophy and Religion programs
x <- cip_filter(series = "^38")
y <- cip_label(data = x, program = "test label")

# The function is pipe-compatible
y <- x \%>\% cip_label("test label")

# With no program argument, the 6-digit CIP names are assigned
y <- x \%>\% cip_label()

# Argument options include
y <- x \%>\% cip_label(program = "cip6name")
y <- x \%>\% cip_label(program = "cip4name")
y <- x \%>\% cip_label(program = "cip2name")

# When a 4-digit level name is desired
y <- cip_filter(series = "^1410") \%>\% cip_label("cip4name")

# But if the result is too long, it can be manually reassigned
y <- cip_filter(series = "^1410") \%>\% cip_label("Electrical Engineering")
}
