% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cip_series}
\alias{cip_series}
\alias{cip_bio_sci}
\alias{cip_engr}
\alias{cip_math_stat}
\alias{cip_other_stem}
\alias{cip_phys_sci}
\alias{cip_stem}
\title{Named series of 6-digit CIP codes}
\format{
An atomic character vector of 6-digit CIP codes.

\describe{
\item{\code{cip_bio_sci}}{Biological and biomedical science programs, series 26.}
\item{\code{cip_engr}}{Engineering programs, series 14.}
\item{\code{cip_math_stat}}{Mathematics and statistics programs, series 27.}
\item{\code{cip_phys_sci}}{Physical science programs, series 40.}
\item{\code{cip_other_stem}}{Other STEM programs, not individually named.}
\item{\code{cip_stem}}{All Science, Technology, Engineering, and Mathematics (STEM) programs.}
}

An object of class \code{character} of length 101.

An object of class \code{character} of length 56.

An object of class \code{character} of length 17.

An object of class \code{character} of length 218.

An object of class \code{character} of length 43.

An object of class \code{character} of length 435.
}
\source{
NSF STEM programs, \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}

CIP 2010, US National Center for Education Statistics, \url{https://nces.ed.gov/ipeds/cipcode}.
}
\usage{
cip_bio_sci

cip_engr

cip_math_stat

cip_other_stem

cip_phys_sci

cip_stem
}
\description{
Atomic character vectors of 6-digit CIP codes for specified groups of programs.
}
\examples{
head(cip_math_stat, n = 10L)
head(cip_phys_sci, n = 10L)

cip_filter(data = cip, keep_any = cip_math_stat)
cip_filter(data = cip, keep_any = cip_phys_sci)
}
\keyword{datasets}
