% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion_status.R
\name{completion_status}
\alias{completion_status}
\title{Determine completion status}
\usage{
completion_status(dframe, midf_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \verb{\{mcid, timely_term\}.}}

\item{midf_table}{\code{degree} data frame with required
variables \verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added and all unique columns are
preserved. If required to prevent overwriting, new column names are
suffixed (\verb{.1, .2,} etc.). Redundant columns, differing by suffix only,
are dropped. The new variables are:
\itemize{
\item \code{bacc_term}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA.}
Joined from the \code{term_degree} variable in \code{midf_table.}
\item \code{completion}   Character. Completion status, possible values of
"timely", "late", and "NA".
}
}
}
\description{
Determine the \emph{completion status} for each student in a data frame and
add columns that support the findings.
}
\details{
If a population has been filtered for data sufficiency, then determining
every student's \emph{completion status} is feasible. Completing an academic
program in a timely manner means that a student completes the requirements
for a degree within a set time span, typically 4, 6, or 8 years after
admission depending on the definition adopted in a particular study. The
term at the end of that span is the \emph{timely completion term.}

If the student's degree term is no later than their timely completion term,
then their completion status is "timely"; if later, their status is "late".
For students with no degree, completion status is NA.
}
\examples{
# Assign toy data sets
student <- toy_student
term <- toy_term
degree <- toy_degree

# Start with a selected population
x <- student[c(9:11, 21:30, 344:345), .(mcid)]
x

# Add the required columns from timely_term()
x <- timely_term(x, midf_table = term)
x <- x[, .(mcid, timely_term)]
x

# Add completion status columns
x <- completion_status(x, midf_table = degree)
x

# No change if new columns match existing columns
y = completion_status(x, midf_table = degree)

# If new column should match existing but does not,
# new column with suffix .1, .2, etc., is added. 
# Indicates an error has occurred somewhere. 
y$bacc_term[1] <- "19893"
z = completion_status(y, midf_table = degree)
z

# Typical application retains "timely" rows only
x[completion == "timely"]
}
