% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion_status.R
\name{completion_status}
\alias{completion_status}
\title{Determine completion status}
\usage{
completion_status(dframe, midf_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \verb{\{mcid, timely_term\}.}}

\item{midf_table}{\code{degree} data frame with required
variables \verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item Columns with names different from the new columns (named
below) are not modified; columns with matching names are replaced.
The new columns added are:
\itemize{
\item \code{completion_term}   Equal to \code{term_degree} from \code{midf_table.}
\item \code{completion_status}   Character. Possible values of "timely",
"late" and "NA".
}
}
}
\description{
Determine the \emph{completion status} for each student in a data frame and
add columns that support the findings.
}
\details{
If a population has been filtered for data sufficiency, then determining
every student's \emph{completion status} is feasible. Completing an academic
program in a timely manner means that a student completes the requirements
for a degree within a set time span, typically 4, 6, or 8 years after
admission depending on the definition adopted in a particular study. The
term at the end of that span is the \emph{timely completion term.}

If the student's degree term is no later than their timely completion term,
then their completion status is "timely"; if later, their status is "late".
For students with no degree, completion status is NA.
}
\examples{
# Assign toy data sets
student <- toy_student
term <- toy_term
degree <- toy_degree

# Start with a selected population
x <- student[c(9:11, 21:30, 344:345), .(mcid)]
x

# Add the required columns from timely_term()
x <- timely_term(x, midf_table = term)
x <- x[, .(mcid, timely_term)]
x

# Add completion status columns
x <- completion_status(x, midf_table = degree)
x

# If you repeat, the new columns are overwritten
completion_status(x, midf_table = degree)

# Typical application retains "timely" rows only
x[completion_status == "timely"]
}
