% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion_status.R
\name{completion_status}
\alias{completion_status}
\title{Determine completion status}
\usage{
completion_status(dframe, midfield_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble). Required variables: \verb{\{mcid, timely_term\}}.}

\item{midfield_table}{MIDFIELD  \emph{\strong{degree}}  dataset (data frame or data frame extension). Required variables:
\verb{\{mcid, term_degree\}}.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved.
\item Rows are filtered for unique \code{mcid} values.
\item Columns \verb{\{mcid, timely_term\}} are retained. All other columns are
dropped and the following columns are added:
\itemize{
\item \code{term_degree.}   Character. Term in which the first degree(s) are
completed, encoded \code{YYYYT}. Joined from \code{midfield_table.}
\item \code{completion_status.}   Character. Possible values are "timely"
for students completing a degree no later than their timely
completion terms; "late" for students completing their program
after their timely completion term; and "NA" for non-completers.
}
\item Groups and keys are not preserved.
}
}
\description{
To a data frame keyed by student ID, add a column indicating if a
student completed their program, and if so, whether their completion
was timely or late. Columns of supporting information are
also added. Columns not related to the task are dropped.
}
\details{
In many studies, students must complete their programs in a specified time
span to be considered "timely", for example 4, 6, or 8 years after
admission. By "completion" we mean an
undergraduate earning their first baccalaureate degree (or degrees, for
students earning more than one degree in the same term).

The goal of determining timely completion is to refine a population, that
is, obtain a data frame of IDs that satisfy our constraints. Thus
\code{completion_status()} yields a column of completion status values and
columns of supporting information keyed by ID. All other columns in
\code{dframe} (if any) are dropped.

The supporting information in the output is provided so that the user
can review the findings. After review, we usually delete all columns
except the IDs, yielding the refined population that was our goal.
}
\examples{
term <- toy_term
degree <- toy_degree

# Start with a small population 
x <- toy_student[21:36, .(mcid)]
x

# Timely term column is required
x <- timely_term(x, term)
x

# Add completion status column, columns not used are dropped
x <- completion_status(x, degree)
x

# Existing completion status column (if any) is replaced
x[, completion_status := NA_character_][]
completion_status(x, degree)
}
