% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion_status.R
\name{completion_status}
\alias{completion_status}
\title{Determine completion status}
\usage{
completion_status(dframe, midfield_rec = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble). Required variables: \verb{\{mcid, timely_term\}}.}

\item{midfield_rec}{MIDFIELD records \emph{degree} data frame or data frame extension. Required variables:
\verb{\{mcid, term_degree\}}.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are not preserved.
\item Rows are filtered for unique \code{mcid} values.
\item Columns \verb{\{mcid, timely_term\}} are retained (all other columns
are dropped). New columns added:
\itemize{
\item \code{term_degree.}   Character. Term in which the first degree(s) are
completed, encoded \code{YYYYT}. Joined from \code{midfield_rec.}
\item \code{completion_status.}   Character. Possible values are "timely"
for students completing a degree no later than their timely
completion terms; "late" for students completing their program
after their timely completion term; and "NA" for non-completers.
}
}
}
\description{
To a data frame keyed by student ID, add a column indicating if a
student completed their program, and if so, whether their completion
was timely or late. Columns of supporting information are
also added.
}
\details{
In many studies, students must complete their programs in a specified time
span to be considered "timely", for example 4, 6, or 8 years after
admission. By "completion" we mean an
undergraduate earning their first baccalaureate degree (or degrees, for
students earning more than one degree in the same term).

The goal of determining timely completion is to refine a population, that
is, obtain a data frame of IDs that satisfy our constraints. Thus
\code{completion_status()} yields a column of completion status values and
columns of supporting information keyed by ID. All other columns in
\code{dframe} (if any) are dropped.

The supporting information in the output is provided so that the user
can review the findings. After review, we usually delete all columns
except the IDs, yielding the refined population that was our goal.
}
\examples{
# Start with an excerpt from the student data set 
dframe <- toy_student[1:10, .(mcid)]

# Timely term column is required to add completion status column
dframe <- timely_term(dframe, toy_term)

# Add completion status column
completion_status(dframe, toy_degree)

# Existing completion_status column, if any, is overwritten
dframe[, completion_status := NA_character_][]
completion_status(dframe, toy_degree)
}
\seealso{
Other add_*:
\code{\link[=add_term_cluster]{add_term_cluster()}},
\code{\link[=data_sufficiency]{data_sufficiency()}},
\code{\link[=timely_term]{timely_term()}}
}
\concept{add_*}
