% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_multiway.R
\name{condition_multiway}
\alias{condition_multiway}
\title{Condition multiway data for graphing}
\usage{
condition_multiway(
  dframe,
  categ_col,
  quant_col,
  ...,
  detail = NULL,
  order_by = NULL,
  param_col = NULL
)
}
\arguments{
\item{dframe}{Data frame with at least three columns: two categorical
variables (columns can be character or factor) and one quantitative
response variable (column must be numeric).}

\item{categ_col}{Character vector with the names (in quotes) of the two
categorical columns.}

\item{quant_col}{Character scalar with the name (in quotes) of the
quantitative response column.}

\item{...}{Not used, forces later arguments to be used by name.}

\item{detail}{Optional flag to retain the columns created to implement the
ordering scheme, default FALSE. When TRUE, the extra columns are
useful for graphing vertical reference lines---indicating, for
instance, the panel median, mean, or total count---to visualize the
quantity by which the panels are ordered.}

\item{order_by}{Optional character scalar (in quotes) assigning the
       method for ordering the levels of the categorical variables.
       The following values are possible:
       \itemize{

       \item{"median"} {(default) Orders by the median of values in the
       quantitative column grouped by category.}

       \item{"mean"} {Orders by the mean of values in the quantitative
       column grouped by category.}

       \item{"sum"} {Orders by the sum of values in the quantitative
       column grouped by category. Useful when the quantitative variable
       is a count or frequency.}

       \item{"percent"} {Orders by ratios computed by category. Used when
       the quantitative response variable is the ratio (in percent) of two
       columns of integer counts (frequencies). The counts are summed by
       category to obtain grouped percentages used to order the levels of
       the categorical variables. Requires two parameter columns be
       identified in \code{param_col} and assumes that the parameter with
       the larger column sum is the denominator of the ratio.}

       \item{"alphabet"} {Orders the levels of the categorical variables
       alphabetically. Rarely useful for perceiving effects, but can be
       useful for value look up.}

       }}

\item{param_col}{Optional character vector with the names (in quotes)
of the columns used as parameters in the \code{order_by} method.
Currently supports \code{"percent"} method only---expects two column
names (in any order) of the integer count columns used to construct
the percentage response column.}
}
\value{
A \code{data.table} with the following properties:
\itemize{
  \item Rows are not modified.
  \item The quantitative column is not modified.
  \item The two categorical columns are factors with levels ordered by
        the method selected. Optionally, the additional columns (if any)
        created to implement the ordering scheme can be retained.
  \item Grouping structures are not preserved.
}
}
\description{
Transform a data frame such that two categorical variables are factors
with levels ordered for display in a Cleveland "multiway dot plot," where
the ordering of the panels and rows are crucial to the perception of
effects. In multiway data---as defined by
\insertCite{Cleveland:1993;textual}{midfieldr}---there is a single
quantitative value (or response) for every combination of levels
of two categorical variables. Typically the quantitative column is used to
order the levels of the two categorical columns.
}
\details{
In the multiway dot plot, there are panels, the individual dot plots
of the display, and there are levels, the rows of each panel. One category
is encoded by the panels; the other by the rows. All panels have the same
quantitative scale on the x-axis and the same organization of category
levels on the y-axis. Panels and rows are ordered so that ordering scheme
increases in "graph order", that is, increases from left to right and from
bottom to top.

Note that "multiway" in our context refers to the data structure and graph
design defined by Cleveland, not to the methods of analysis described by
\insertCite{Kroonenberg:2008;textual}{midfieldr}.
}
\references{
\insertAllCited{}
}
\seealso{
Other condition_*: 
\code{\link{condition_fye}()}
}
\concept{condition_*}
