% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_multiway.R
\name{condition_multiway}
\alias{condition_multiway}
\title{Condition multiway data for graphing}
\usage{
condition_multiway(dframe, ..., details = NULL)
}
\arguments{
\item{dframe}{Data frame of multiway data.}

\item{...}{Not used, forces later arguments to be used by name.}

\item{details}{Logical scalar to add columns reporting the medians on which
the order of the levels is based, default FALSE.}
}
\value{
A \code{data.table} with the following properties:
\itemize{
  \item Rows are not modified.
  \item Quantitative column is not modified.
  \item Categorical columns are factors with levels ordered by
        median quantitative values.
  \item Grouping structures are not preserved.
}
}
\description{
Transform a data frame such that two categorical variables are factors
with levels ordered by medians of the quantitative variable. Rows and
columns of the data frame are otherwise unaffected.
}
\details{
In multiway data, there is a single quantitative value (or response)
for every combination of levels of two categorical variables.
\code{condition_multiway()} converts the columns of categorical variables
to factors and orders the factor levels by increasing medians of the
quantitative response variable.

The \code{dframe} argument, therefore, must have three columns only. One
column is a quantitative variable (type numeric) and two columns are
categorical variables (type character or factor).

Note that "multiway" in our context refers to the data structure and graph
design defined by \insertCite{Cleveland:1993;textual}{midfieldr}, not to
the methods of analysis described by
\insertCite{Kroonenberg:2008;textual}{midfieldr}.
}
\examples{
catg1 <- rep(c("urban", "rural", "suburb", "village"), each = 2)
catg2 <- rep(c("men", "women"), times = 4)
value <- c(0.22, 0.14, 0.43, 0.58, 0.81, 0.46, 0.15, 0.20)

# structure before
dframe <- data.frame(catg1, catg2, value)
str(dframe)

# structure after
mw <- condition_multiway(dframe)
str(mw)
}
\references{
\insertAllCited{}
}
\seealso{
Other functions: 
\code{\link{add_completion_timely}()},
\code{\link{add_data_sufficiency}()},
\code{\link{add_institution}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()},
\code{\link{condition_fye}()},
\code{\link{filter_match}()},
\code{\link{filter_text}()}
}
\concept{functions}
