% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sufficiency.R
\name{data_sufficiency}
\alias{data_sufficiency}
\title{Determine data sufficiency}
\usage{
data_sufficiency(dframe, midf_table = term)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables
\verb{\{mcid, entry_term, timely_term\}.}}

\item{midf_table}{\code{term} data frame with required variables
\verb{\{mcid, term, institution\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added unless they duplicate existing
variables---redundant columns, whether new or existing, are dropped
(see Details). The new variables are:
\itemize{
\item \code{data_range}   Character. Institution data range, encoded
\verb{YYYYT-YYYYT,} indicating the institution's first and last term in the
database. Extracted from \code{midf_table.}
\item \code{sufficiency}   Character. Data sufficiency results.
Possible values are "satisfied", "fail-lower," and "fail-upper."
}
}
}
\description{
Determine institutional \emph{data sufficiency} for each student in a data frame
and add columns that support the findings.
}
\details{
In most studies, the population must satisfy the \emph{data sufficiency}
criterion, developed as follows:
\itemize{
\item Program \emph{completion} means satisfying the requirements for a first
baccalaureate degree.
\item Completion \emph{status} is "timely" if accomplished within a set time
span, typically 4, 6, or 8 years after admission depending on the
definition one adopts. The \emph{timely-completion term} is the term at the
end of that span.
\item The \emph{data sufficiency} test identifies students whose actual admission
term and projected timely completion term both lie within their
institution's data range. These are the students for whom completion
status---timely or otherwise---can be positively asserted, and are
therefore the only students included a population.
}

To apply this criterion, our heuristic labels a row (keyed by student ID)
"exclude-upper" when a student's timely completion term exceeds the upper
limit of their institution's data range;  "exclude-lower" when their initial
term matches the non-summer, lower limit of the data range; and "include"
otherwise. The rationale for these specific filters is explained in our
data sufficiency article (see references).

When midfieldr functions add columns to a data frame, such as  \verb{sufficiency,}  they are dropped if they duplicate an existing variable. When an added variable has the same name as an existing variable but different values, the new variable name acquires a suffix, e.g., \code{sufficiency.1.} These details are managed by \verb{select_unique_cols().} See its help page for examples.

An added variable with a suffix indicates one of two possibilities:
\enumerate{
\item The existing variable has inadvertently been named the same as the new variable. The two variables represent different information entirely and the existing variable should probably be re-named before running this function to add the new variable.
\item The two variables represent the same information but their values disagree. Such differences are not expected---these added variables typically depend on fixed quantities, e.g., a student's admission term or an institution's data range, and should not change during an analysis. In such a case, the user should investigate the possibility of an error having been introduced at some point.
}
}
\examples{
# Assign toy data sets
student <- toy_student
term <- toy_term

# Start with a selected population
x <- student[c(9:11, 21:30, 344:345), .(mcid)]
x

# Add the required columns from timely_term()
x <- timely_term(x, midf_table = term)
x <- x[, .(mcid, entry_term, timely_term)]
x

# Add data sufficiency columns
x <- data_sufficiency(x, midf_table = term)
x

# No change if columns to be added are redundant
data_sufficiency(x, midf_table = term)

# Filter to retain "satisfied" rows only
x[sufficiency == "satisfied"]
}
\references{
R. Layton, R. Long, M. Ohland, M. Orr, and S. Lord (2026), "Data sufficiency,"  \url{https://midfieldr.github.io/midfieldr/articles/art-020-data-sufficiency.html}
}
