% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sufficiency.R
\name{data_sufficiency}
\alias{data_sufficiency}
\title{Determine data sufficiency}
\usage{
data_sufficiency(dframe, midfield_table = term)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables
\verb{\{mcid, term_i, timely_term\}.}}

\item{midfield_table}{\code{term} data frame with required variables
\verb{\{mcid, term, institution\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item Columns with names different from the new columns (named
below) are not modified; columns with matching names are replaced.
The new columns added are:
\itemize{
\item \code{institution}   Character. Name of the institution at which a
student is enrolled in a term.
\item \code{lower_limit}   Character. Initial term of an institution's
data range, encoded \code{YYYYT}. Extracted from \code{midfield_table.}
\item \code{upper_limit}   Character. Final term of an institution's
data range, encoded \code{YYYYT}. Extracted from \code{midfield_table.}
\item \code{data_sufficiency}   Character. Possible values are "include",
"exclude-lower," and "exclude-upper."
}
}
}
\description{
Determine \emph{data sufficiency} for each student in a data frame and add
columns that support the findings.
}
\details{
\emph{Data sufficiency} is a criterion for including or excluding a student
record based on the feasibility of determining their completion status given
the range of data available from their institution. If determining
completion status is feasible, the student record is included in the
study population; if not, they must be excluded to avoid biased counts
of completers and non-completers. Such biases occur at the upper and lower
bounds of an institution's data range.

To apply this criterion, our heuristic labels a row
"exclude-upper" when a student's timely completion term exceeds the upper
limit of their institution's data range;  "exclude-lower" when their initial
term matches the lowest non-summer limit of the data range; and "include"
otherwise. The rationale for these specific filters is
explained in our data sufficiency article (see references). In most
studies, the population must satisfy the data sufficiency requirement.
}
\examples{
term <- toy_term

# Start with a selected population.
x <- toy_student[c(9:15, 342:344), .(mcid, sex)]
x

# Add the required columns from timely_term().
x <- timely_term(x, midfield_table = term)
x <- x[, .(mcid, sex, term_i, timely_term)]
x

# Add data sufficiency columns. Unrelated columns (sex) are unaffected.
x <- data_sufficiency(x, midfield_table = term)
x

# Repeat. New columns silently replace existing columns of the same name.
y <- data_sufficiency(x, midfield_table = term)
y
}
\references{
Richard Layton, Russell Long, Matthew Ohland, Marisa Orr, and Susan Lord (2026) Data sufficiency, https://midfieldr.github.io/midfieldr/articles/art-020-data-sufficiency.html
}
