% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sufficiency.R
\name{data_sufficiency}
\alias{data_sufficiency}
\title{Determine data sufficiency}
\usage{
data_sufficiency(dframe, midfield_table = term)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables
\verb{\{mcid, term_i, timely_term\}.} The latter two variables are
provided by \verb{timely_term().}}

\item{midfield_table}{\code{term} data frame with required variables
\verb{\{mcid, term, institution\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved.
\item Rows are not modified except duplicated rows are removed.
Row order is preserved.
\item Variables \verb{\{mcid, term_i, timely_term\}} are retained. All other
columns (if any) are dropped and the following variables are added:
\itemize{
\item \code{institution.}   Character. Institution in which the student is
enrolled in the given term. Extracted from \code{midfield_table.} The
limits given in the next two columns are specific to the institution.
\item \code{lower_limit.}   Character. Initial term of an institution's
data range, encoded \code{YYYYT}. Extracted from \code{midfield_table.}
Compared to \code{term_i} to determine the lower-limit exclusion.
\item \code{upper_limit.}   Character. Final term of an institution's
data range, encoded \code{YYYYT}. Extracted from \code{midfield_table.}
Compared to \code{timely_term} to determine upper-limit exclusion.
\item \code{data_sufficiency.}   Character. Possible values are "include",
if the data are sufficient; and "exclude-lower" or "exclude-upper"
if not, indicating at which boundary of the data range the ambiguity
occurs.
}
\item Groups and keys are not preserved.
}
}
\description{
For each student in a data frame, determine whether or not their record
lies sufficiently within their institution's data range to unambiguously
assess their completion status and if so include them in the study
population. Label  each row with this \emph{data sufficiency} result (include
or exclude) and add columns that support the findings.
}
\details{
\emph{Timely completion} means completing a program no later than a specified
interval---typical values are 4, 6, or 8 years after admission. The
\emph{data sufficiency} criterion states that student records must be limited
to those for which available data from an institution are sufficient to
assess timely completion without biased counts of completers or
non-completers. Such biases occur at the lower and upper bounds of an
institution's data range. Affected students must be identified and
excluded to prevent false summary counts.

In our heuristic, the criteria is implemented via two filters. Rows are
labeled for exclusion when: 1) a student ID is extant in the non-summer
lower limit of an institution's data range; or 2) a student ID has a
timely completion term that exceeds the upper limit of the institution's
data range. The results are documented in the output.
}
\examples{
term <- toy_term

# Start with a small population 
x <- toy_student[c(9:15, 342:344), .(mcid)]
x

# Timely term column is required
x <- timely_term(x, term)
x

# Add data sufficiency column, columns not used are dropped
x <- data_sufficiency(x, term)
x

# Existing data sufficiency column (if any) is replaced
x[, data_sufficiency := NA_character_][]
data_sufficiency(x, term)
}
