% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sufficiency.R
\name{data_sufficiency}
\alias{data_sufficiency}
\title{Determine data sufficiency}
\usage{
data_sufficiency(dframe, midf_table = term)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables
\verb{\{mcid, term_i, timely_term\}.}}

\item{midf_table}{\code{term} data frame with required variables
\verb{\{mcid, term, institution\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item Columns with names different from the new columns (named
below) are not modified; columns with matching names are replaced.
The new columns added are:
\itemize{
\item \code{institution}   Character. Name of the institution at which a
student is enrolled in a term.
\item \code{lower_limit}   Character. Initial term of an institution's
data range, encoded \code{YYYYT}. Extracted from \code{midf_table.}
\item \code{upper_limit}   Character. Final term of an institution's
data range, encoded \code{YYYYT}. Extracted from \code{midf_table.}
\item \code{data_sufficiency}   Character. Possible values are "include",
"exclude-lower," and "exclude-upper."
}
}
}
\description{
Determine institutional \emph{data sufficiency} for each student in a data frame
and add columns that support the findings.
}
\details{
In most studies, the population must satisfy the \emph{data sufficiency}
criterion, developed as follows:
\itemize{
\item Program \emph{completion} means satisfying the requirements for a first
baccalaureate degree.
\item Completion \emph{status} is "timely" if accomplished within a set time
span, typically 4, 6, or 8 years after admission depending on the
definition one adopts. The \emph{timely-completion term} is the term at the
end of that span.
\item The \emph{data sufficiency} test identifies students whose actual admission
term and projected timely completion term both lie within their
institution's data range. These are the students for whom completion
status---timely or otherwise---can be positively asserted, and are
therefore the only students included a population.
}

To apply this criterion, our heuristic labels a row (keyed by student ID)
"exclude-upper" when a student's timely completion term exceeds the upper
limit of their institution's data range;  "exclude-lower" when their initial
term matches the non-summer, lower limit of the data range; and "include"
otherwise. The rationale for these specific filters is explained in our
data sufficiency article (see references).
}
\examples{
term <- toy_term

# Start with a selected population.
x <- toy_student[c(9:15, 342:344), .(mcid, sex)]
x

# Add the required columns from timely_term().
x <- timely_term(x, midf_table = term)
x <- x[, .(mcid, sex, term_i, timely_term)]
x

# Add data sufficiency columns. Unrelated columns (sex) are unaffected.
x <- data_sufficiency(x, midf_table = term)
x

# Repeat. New columns silently replace existing columns of the same name.
y <- data_sufficiency(x, midf_table = term)
y
}
\references{
R. Layton, R. Long, M. Ohland, M. Orr, and S. Lord (2026), "Data sufficiency,"  \url{https://midfieldr.github.io/midfieldr/articles/art-020-data-sufficiency.html}
}
