% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{engr}
\alias{engr}
\title{CIP codes for programs in engineering}
\format{A character vector.}
\source{
NSF STEM programs: \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}.
}
\usage{
engr
}
\description{
All 6-digit codes from the 2010 CIP 14 series.
}
\examples{
\dontrun{
 engr
}
}
\keyword{datasets}
