% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ever_filter.R
\name{ever_filter}
\alias{ever_filter}
\title{Filter term data for enrolled students}
\usage{
ever_filter(series, ..., reference = NULL, id = "id", cip6 = "cip6",
  term = "term")
}
\arguments{
\item{series}{atomic character vector of 6-digit CIP codes specifying the
programs to filter by}

\item{...}{not used for values, forces later arguments to bind by name}

\item{reference}{a reference data frame from which student IDs, academic
terms, and CIP codes are obtained, default \code{midfieldterms}}

\item{id}{character column name of the ID variable in \code{reference}}

\item{cip6}{character column name of the CIP code variable
in \code{reference}}

\item{term}{character column name of term variable in \code{reference}}
}
\value{
Data frame with character variables for student ID and program CIP
}
\description{
Filter academic term data to find all students ever enrolled in specified
programs.
}
\details{
To use this function, the \code{midfielddata} package must be installed to
provide \code{midfieldterms}, the default reference data set.

The \code{series} argument is an atomic character vector of 6-digit CIP
codes used to filter the reference data set and extract student IDs
and terms.

The function returns a subset of \code{reference} with every unique
combination of student ID and program CIP, representing every student
ever enrolled in the specified programs. Only the student ID and program
CIP are returned; other variables in \code{reference} are quietly dropped.

Optional arguments. An alternate reference data set can be assigned via
the \code{reference} argument. The alternate must have variables for
student ID, 6-digit CIP, and terms with the same structure as in
\code{midfieldterms}. Use the  \code{id}, \code{cip6}, and \code{term}
arguments to reassign the variables names, if necessary, to match the
variable names in the alternate.
}
\examples{
ever_filter(series = "540104")

}
\seealso{
\code{\link[midfieldr]{cip_filter}} for obtaining 6-digit CIP codes
}
