% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_cip}
\alias{exa_cip}
\title{Example CIP excerpt}
\format{
A data frame (tibble) with 12 observations and 6 variables.
The variables are characters.
\describe{
  \item{cip6}{6-digit program code.}
  \item{cip6name}{6-digit program name.}
  \item{cip4}{4-digit program code.}
  \item{cip4name}{4-digit program name.}
  \item{cip2}{2-digit program code.}
  \item{cip2name}{2-digit program name.}
}
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}
 package
}
\usage{
exa_cip
}
\description{
Excerpt of CIP data for a case study.
}
\details{
An excerpt from the \code{cip} data set for Civil, Electrical,
Industrial, and Mechanical Engineering. The data are an intermediate
result in a case study developed in the midfieldr vignettes.
}
\examples{
exa_cip
cip6_select(data = exa_cip, program = "cip4name")
}
\seealso{
\code{cip} and the "Explore CIP" vignette.

Other example_data: 
\code{\link{exa_imputed_fye}},
\code{\link{exa_program_group}},
\code{\link{exa_stickiness}}
}
\concept{example_data}
\keyword{datasets}
