% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_group}
\alias{exa_group}
\title{Example program group}
\format{
A data frame (tibble) with 12 observations and 3 variables.
All variables are characters. An observation is a unique program.
\describe{
  \item{cip6}{6-digit program code.}
  \item{cip6name}{6-digit program name.}
  \item{program}{A short program name assigned for grouping,
  summarizing, and joining.}
}
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}
 package
}
\usage{
exa_group
}
\description{
Six-digit CIP codes and names plus a custom program name for
Civil, Electrical, Industrial, and Mechanical Engineering. Used for
grouping, summarizing, and joining in a case study in the vignettes.
}
\examples{
x <- unique(exa_group[["cip6"]])
y <- unique(exa_group[["program"]])
}
\seealso{
The "Gather program data" vignette.

Other example_data: 
\code{\link{exa_cip}},
\code{\link{exa_id}},
\code{\link{exa_stickiness}}
}
\concept{example_data}
\keyword{datasets}
