% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_id}
\alias{exa_id}
\alias{exa_ever}
\alias{exa_grad}
\title{Example student IDs}
\format{
Character vector of student IDs.
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}
 package
}
\usage{
exa_ever

exa_grad
}
\description{
Students IDs from midfielddata for Civil, Electrical, Industrial,
and Mechanical Engineering. Used in a case study in the vignettes.
}
\details{
\describe{
\item{\code{exa_ever}}{IDs of students ever enrolled in the example
programs.}
\item{\code{exa_grad}}{IDs of students graduating from the example
programs.}
}
}
\examples{
library(midfielddata)
midfieldstudents[midfieldstudents[["id"]] \%in\% exa_ever, ]
midfieldstudents[midfieldstudents[["id"]] \%in\% exa_grad, ]
}
\seealso{
Other example_data: 
\code{\link{exa_cip}},
\code{\link{exa_group}},
\code{\link{exa_stickiness}}
}
\concept{example_data}
\keyword{datasets}
