% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_stickiness}
\alias{exa_stickiness}
\title{Example stickiness data}
\format{
\code{data.frame} with 32 rows and 6 columns keyed by
program, race/ethnicity, and sex. The variables are:
\describe{
 \item{program}{instructional programs selected for study}
 \item{race}{student race/ethnicity}
 \item{sex}{student sex}
 \item{ever}{number of students ever enrolled in a program and completion
 is feasible}
 \item{grad}{number of students graduating from the program}
 \item{stick}{program stickiness, a persistence metric}
}
}
\usage{
exa_stickiness
}
\description{
Stickiness metric results for a case study of Civil, Electrical,
Industrial, and Mechanical Engineering programs in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}. Results
are grouped by program, race/ethnicity, and sex.
}
\examples{
# prepare data for multiway graph
exa_stickiness$race_sex <- paste(exa_stickiness$race, exa_stickiness$sex)
columns_we_want <- c("program", "race_sex", "stick")
pre_mw <- exa_stickiness[, columns_we_want, drop = FALSE]

}
\seealso{
Other example_data: 
\code{\link{exa_cip}},
\code{\link{exa_group}},
\code{\link{exa_id}}
}
\concept{example_data}
\keyword{datasets}
