% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exa_stickiness}
\alias{exa_stickiness}
\title{Example stickiness data}
\format{
A data frame (tibble) with 32 observations and 6 variables.
\describe{
 \item{program}{Instructional programs selected for study. Character
 variable.}
 \item{race}{Student race and sex from \code{midfieldstudents} data.
  Character variable.}
 \item{sex}{Student sex from \code{midfieldstudents} data.  Character
 variable.}
 \item{ever}{Number of students ever enrolled in a program. Grouped
 and summarized from \code{midfieldterms} data.}
 \item{grad}{Number of students graduating from the program. Grouped
 and summarized from \code{midfielddegrees} data.}
 \item{stick}{Program stickiness: the ratio of the number of
 students graduating from a program to the number ever enrolled in the
 program.}
}
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}
 package
}
\usage{
exa_stickiness
}
\description{
Stickiness metric results for Civil, Electrical, Industrial, and
Mechanical Engineering. Data are grouped by program, student
race/ethnicity, and sex. Used for graphing the results in a case study
in the vignettes.
}
\examples{
exa_stickiness
}
\seealso{
The "Compute stickiness" vignette.

Other example_data: 
\code{\link{exa_cip}},
\code{\link{exa_group}},
\code{\link{exa_id}}
}
\concept{example_data}
\keyword{datasets}
