% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cip.R
\name{filter_cip}
\alias{filter_cip}
\title{Filter CIP data}
\usage{
filter_cip(series = NULL, data = NULL)
}
\arguments{
\item{series}{The conditions used to filter the data. A character vector (or coercible to one) of 2, 4, or 6-digit CIP codes.}

\item{data}{A data frame of CIP codes and program names at the 2, 4, and 6-digit levels: \code{CIP2}, \code{PRGM2}, \code{CIP4}, \code{PRGM4}, \code{CIP6}, and \code{PRGM6}. The default is the midfieldr \code{cip} dataset. All variables are characters.}
}
\value{
A data frame: \code{data} filtered by the conditions in \code{series}.
}
\description{
Filter a data frame of Classification of Instructional Programs (CIP) codes to return the rows that match conditions.
}
\details{
In the midfieldr datasets (\code{student}, \code{course}, \code{term},   \code{degree}) a 6-digit CIP code identifies the program in which a student is enrolled at matriculation, in a term, and at graduation. The purpose of \code{filter_cip()} is to obtain the 6-digit codes of the programs one wishes to study.
}
\examples{
filter_cip(series = c("490101", "490205"))
filter_cip(series = c("4901", "4902"))
filter_cip(series = c("49", "99"))
filter_cip(series = c("54", "4901", "490205"))
filter_cip(series = seq(540102, 540108))

}
