% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_match.R
\name{filter_match}
\alias{filter_match}
\title{Subset rows by matching values in shared key columns}
\usage{
filter_match(dframe, match_to, by_col, ..., select = NULL)
}
\arguments{
\item{dframe}{Data frame to be subset and returned. Must contain the key
column named in \code{by_col}.}

\item{match_to}{Data frame with key column values to be matched to. The
only column used is the required key column named in \code{by_col}.}

\item{by_col}{Character scalar, name of the key column. Values in the
key column must be character strings.}

\item{...}{Not used, force later arguments to be used by name.}

\item{select}{Optional character vector of \code{dframe} column names
to retain,
default all columns.}
}
\value{
A \code{data.table} with the following properties:
\itemize{
    \item Rows with matching values.
    \item All columns or those specified by \code{select}.
    \item Grouping structures are not preserved.
}
}
\description{
Subset rows of one data frame by values of a key column that match values
of a similar key column in a second data frame. Columns are not subset
unless selected in an optional argument.
}
\details{
Two data frames are input. The \code{match_to} data frame is subset to
retain its key column only.  The result is merged with \code{dframe} in an
inner-join, returning the rows of \code{dframe} with values matching the key
values in \code{match_to}.

Column subsetting occurs after the inner join, so the
key column does not have to be included in the vector of column names in
\code{select}.
}
\examples{
# Start with a toy sample of student (built-in data set)
str(toy_student)


# Filter by matching IDs of graduates
x_student <- filter_match(toy_student,
                          match_to = toy_degree,
                          by_col   = "mcid")
# Note change in number of rows
str(x_student)


# Same filter and select 3 columns only
x_student <- filter_match(toy_student,
                          match_to = toy_degree,
                          by_col   = "mcid",
                          select   = c("mcid", "race", "sex"))
str(x_student)


# Filter term data for engineering program codes (start with "14")
engr_term <- toy_term[grepl("^14", cip6), ]
str(engr_term)


# Filter student by matching IDs of engineering students
x_student <- filter_match(toy_student,
                   match_to = engr_term,
                   by_col   = "mcid",
                   select   = c("mcid", "institution", "transfer", "sex"))
str(x_student)


# The 'by_col' column does not have to be included in the `select` columns
x_student <- filter_match(toy_student,
                          match_to = engr_term,
                          by_col   = "mcid",
                          select   = c("institution", "transfer", "sex"))
str(x_student)


}
\seealso{
Other filter_*: 
\code{\link{filter_search}()}
}
\concept{filter_*}
