% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_programs.R
\name{filter_programs}
\alias{filter_programs}
\title{Choose rows of CIP data}
\usage{
filter_programs(dframe, pattern, ..., negate = FALSE)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble). Expected variables (or subset thereof):
\verb{\{cip6name, cip6, cip4name, cip4, cip2name, cip2\}.}}

\item{pattern}{Character vector of search strings, including regular
expressions.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{negate}{Logical (default FALSE). If TRUE, inverts the
resulting Boolean vector.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are not preserved.
\item Rows are a subset of the input and appear in the same order.
\item Columns are not modified.
}
}
\description{
Subset a CIP data frame, retaining rows that match or partially match
any string in a vector of character strings.
}
\details{
Each element of the \code{pattern} vector is matched row-wise to every
value in \code{dframe} using \verb{grepl().} Row values are coerced to character
strings if possible. If \code{negate = FALSE} (default), a match retains
the full row; if \verb{negate = TRUE,} a match removes the full row.
}
\examples{
# Subset using keywords
filter_programs(cip, pattern = "history")

# Subset using codes
filter_programs(cip, pattern = "^54")

# Multiple passes to narrow the results
first_pass <- filter_programs(cip, "math")[, .(cip6name, cip6)]
first_pass

second_pass <- filter_programs(first_pass, c("bio", "educ"), negate = TRUE)
second_pass

third_pass <- filter_programs(second_pass, c("^27", "^30"))
third_pass
 
}
