% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_text.R
\name{filter_text}
\alias{filter_text}
\title{Subset rows by text}
\usage{
filter_text(
  data,
  keep_text = NULL,
  ...,
  drop_text = NULL,
  keep_col = NULL,
  unique_row = NULL
)
}
\arguments{
\item{data}{data frame to be subset}

\item{keep_text}{character vector of search patterns for retaining rows}

\item{...}{not used, force later arguments to be used by name}

\item{drop_text}{character vector of search patterns for dropping rows}

\item{keep_col}{character vector of column names, default all columns}

\item{unique_row}{logical, remove duplicate rows, default TRUE}
}
\value{
Data frame with the following properties:
\itemize{
    \item Rows matching elements of \code{keep_text}
    \item Columns specified by \code{keep_col}
    \item Data frame extensions such as \code{tbl} or \code{data.table}
    are preserved
    \item Grouping structures are not preserved
}
}
\description{
Subset a data frame, retaining rows that match or partially match
character strings. All columns retained unless specified by
\code{keep_col}. The function uses \code{grepl()}, therefore
non-character columns that can be coerced to character are also searched
for matches.
}
\examples{
# filter using keywords
filter_text(cip, keep_text = "engineering")

# drop_text argument must be named
filter_text(cip, "civil engineering", drop_text = "technology")

# filter using numerical codes
filter_text(cip, keep_text = c("050125", "160501"))

# filter using regular expressions
filter_text(cip, keep_text = "^54")
filter_text(cip, keep_text = c("^1407", "^1408"))

# select columns
filter_text(cip,
            keep_text = "^54",
            keep_col = c("cip4", "cip4name"))

\dontrun{
# unsuccessful terms produce a message
filter_text(cip, c("050125", "111111", "160501", "Bogus", "^55"))
}
}
\seealso{
Other functions: 
\code{\link{prepare_fye_mi}()},
\code{\link{prepare_multiway}()},
\code{\link{subset_feasible}()},
\code{\link{subset_ipeds}()}
}
\concept{functions}
