% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_text.R
\name{filter_text}
\alias{filter_text}
\title{Subset rows by text}
\usage{
filter_text(dframe, keep_text = NULL, ..., drop_text = NULL, select = NULL)
}
\arguments{
\item{dframe}{data frame to be searched}

\item{keep_text}{character vector of search text for retaining rows}

\item{...}{not used, force later arguments to be used by name}

\item{drop_text}{character vector of search text for dropping rows}

\item{select}{character vector of column names to search and return,
default all columns}
}
\value{
A \code{data.table} with the following properties:
\itemize{
    \item Unique rows matching elements of \code{keep_text}
    \item All columns or those specified by \code{select}
    \item Grouping structures are not preserved
}
}
\description{
Subset a data frame, retain unique rows that match or partially match
character strings. All columns retained unless specified by
\code{select}. Uses \code{grepl()}, therefore non-character columns
that can be coerced to character are also searched for matches.
}
\details{
To improve the speed of the search, columns are subset by the values in
the\code{select} argument before the text search starts.
}
\examples{
# subset using keywords
filter_text(cip, keep_text = "engineering")

# drop_text argument, when used, must be named
filter_text(cip, "civil engineering", drop_text = "technology")

# subset using numerical codes
filter_text(cip, keep_text = c("050125", "160501"))

# subset using regular expressions
filter_text(cip, keep_text = "^54")
filter_text(cip, keep_text = c("^1407", "^1408"))

# select columns
filter_text(cip,
            keep_text = "^54",
            select = c("cip6", "cip4name"))

\dontrun{
# unsuccessful terms produce a message
filter_text(cip, c("050125", "111111", "160501", "Bogus", "^55"))
}
}
\seealso{
Other functions: 
\code{\link{add_completion_timely}()},
\code{\link{add_data_sufficiency}()},
\code{\link{add_race_sex}()},
\code{\link{add_timely_term}()},
\code{\link{condition_fye}()},
\code{\link{condition_multiway}()},
\code{\link{filter_by_key}()}
}
\concept{functions}
