% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_ever.R
\name{gather_ever}
\alias{gather_ever}
\title{Gather students ever enrolled in a program}
\usage{
gather_ever(term, programs)
}
\arguments{
\item{term}{A data frame of term data from \pkg{midfieldterm}.}

\item{programs}{A data frame with the 6-digit CIP codes (\code{CIP6}) and
custom label (\code{PGRM}) for the programs being studied.}
}
\value{
A data frame with variables  MID, BEGINYEAR, TYEAR, TTERM, and
INSTITUTION from the original \code{term} data frame plus new variables:

\tabular{ll}{
\code{TCIPN} \tab The "key" variable for gathering the variable names
\code{TCIP}, \code{TCIP2}, and \code{TCIP3}.\cr
\code{CIP6} \tab The 6-digit code of the program in which the student is
enrolled.\cr
\code{ETHNIC} \tab Student race/ethnicty as defined in
\pkg{midfieldstudent}\cr
\code{SEX} \tab Student sex as defined in \pkg{midfieldstudent}\cr
}

Each observation in the resulting tidy data frame is a unique program per
student. Thus the student identifier (\code{MID}) for students who change
majors will appear in more than one row, once each per unique program. If
extant, values \code{TCIP2} and \code{TCIP3} indicate that a student is
enrolled in more than one program in that term.
}
\description{
Filters the \code{term} data frame from the \pkg{midfieldterm} package to
return only those students ever enrolled in the programs being studied.
Accounts for students enrolled in more than one program in a term. Adds
student demographics from \pkg{midfieldstudent}.
}
