% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_ever.R
\name{gather_ever}
\alias{gather_ever}
\title{Gather students ever enrolled in a program}
\usage{
gather_ever(cip)
}
\arguments{
\item{cip}{A data frame created using \code{filter_cip()}.}
}
\value{
A data frame of unique students
}
\description{
Gather students ever enrolled in a program using specified CIP program codes. Used to compute the stickiness metric.
}
