% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_ever.R
\name{gather_ever}
\alias{gather_ever}
\title{Gather ever-enrolled students}
\usage{
gather_ever(.data = NULL, series = NULL)
}
\arguments{
\item{.data}{Data frame of academic term data with character variables \code{id}, \code{cip6}, and \code{term}}

\item{series}{Atomic character vector of 6-digit CIP codes used to filter the data}
}
\value{
Data frame with character variables \code{id} and \code{cip6}
}
\description{
Filter academic term data for students ever enrolled in a set of programs.
}
\details{
To use this function, the \code{midfielddata} package must be installed to provide the default \code{.data} argument, \code{midfieldterms}.

The input data frame must include the three character variables: \code{id}, \code{cip6}, and \code{term}. The default data include every term for every student, so an ID will be associated with the same program in multiple terms. Students who change majors will have their ID associated with different programs in different terms. Students enrolled in more than one major will have their ID associated with multiple programs in the same term. All unique combinations of ID and program are retained.

The \code{series} argument is an atomic character vector of 6-digit CIP codes used to filter the academic term data frame.

The function returns a subset of the input data frame with every unique combination of student ID and program CIP, representing every student ever enrolled in the set of programs being studied. Only \code{id} and \code{cip6} are returned; other variables in the input data frame are dropped.
}
\examples{
ever <- gather_ever(series = "540104")
head(ever)

}
\seealso{
\code{\link[midfieldr]{cip_filter}} for obtaining 6-digit CIP codes
}
