% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_ever.R
\name{gather_ever}
\alias{gather_ever}
\title{Gather students ever enrolled in a program}
\usage{
gather_ever(program_group)
}
\arguments{
\item{program_group}{A data frame that includes 6-digit CIP codes
(\code{CIP6}) of the programs being studied.}
}
\value{
A data frame extracted from the \code{midfieldterms} dataset
with variables \code{ID}, \code{institution}, \code{CIP6}.

Each observation in the resulting tidy data frame is a unique program per
student. Thus the student identifier (\code{ID}) for students who change
majors will appear in more than one row, once each per unique program.
}
\description{
Filters a data frame to
return only those students ever enrolled in the programs being studied.
Accounts for students enrolled in more than one program in a term.
}
