% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_grad.R
\name{gather_grad}
\alias{gather_grad}
\title{Gather program graduates}
\usage{
gather_grad(data, group)
}
\arguments{
\item{data}{A data frame of term data from \pkg{midfielddegrees}}

\item{group}{A data frame with 6-digit CIP codes (\code{CIP6}) for the programs being studied}
}
\value{
A data frame. Each observation is a unique degree per unique
student. Returns students earning their first degree (including multiple
degrees in the same term), but ignores any degrees earned in subsequent
terms.
}
\description{
Filters a data frame to return those students earning degrees from the
programs being studied.
}
