% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_grad.R
\name{gather_grad}
\alias{gather_grad}
\title{Gather graduates}
\usage{
gather_grad(.data = NULL, series = NULL)
}
\arguments{
\item{.data}{Data frame of academic degree data with character variables \code{id}, \code{cip6}, and \code{term_degree}}

\item{series}{Atomic character vector of 6-digit CIP codes used to filter the data}
}
\value{
Data frame with character variables \code{id} and \code{cip6}
}
\description{
Filter academic degree data for students graduating from a set of programs.
}
\details{
To use this function, the \code{midfielddata} package must be installed to provide the default \code{.data} argument, \code{midfielddegrees}.

The input data frame must include the three character variables: \code{id}, \code{cip6}, and \code{term_degree}. The default data include every degree a student earned, but only the student's first degree (or multiple degrees if earned in the same term) are retained.

The \code{series} argument is an atomic character vector of 6-digit CIP codes used to filter the academic degree data frame.

The function returns a subset of the input data frame with every unique combination of student ID and CIP of the program(s) in which they earned their first degree(s). Only \code{id} and \code{cip6} are returned; other variables in the input data frame are dropped.
}
\examples{
grad <- gather_grad(series = "540104")
head(grad)

}
\seealso{
\code{\link[midfieldr]{cip_filter}} for obtaining 6-digit CIP codes
}
