% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_start.R
\name{gather_start}
\alias{gather_start}
\title{Gather students starting in programs}
\usage{
gather_start(.data, cip_group, transfer = FALSE)
}
\arguments{
\item{.data}{A data frame of student matriculation information with two required character variables: student \code{id} and the \code{cip6} code for their starting major. Default is \code{midfieldstudents}.}

\item{cip_group}{A data frame with two required character variables:  \code{cip6} or \code{start} (6-digit CIP codes) and \code{program} (program labels).}

\item{transfer}{Logical to include (TRUE) or exclude (FALSE) transfer students. Default is FALSE.}
}
\value{
A data frame with variables \code{id} (unique MIDFIELD student ID), \code{start} (6-digit CIP code), and \code{program} (a program label).
}
\description{
Filters a data frame of student matriculation information (ID and starting major) to keep those starting in specified programs.
}
\details{
\code{gather_start()} uses \code{dplyr::inner_join()} to combine the rows and columns of its two arguments, both data frames. All variables in the two arguments except \code{id}, \code{cip6} or \code{start} and \code{program} are quietly dropped.

\code{gather_start()} excludes transfer students unless you set \code{transfer = TRUE}. If you are using the result to compute graduation rates per the IPEDS definition, transfer students should be omitted.
}
\examples{
cip_group <- cip_filter(cip, series = "^54")
cip_group <- cip_label(cip_group, program = "cip2name")
starters <- gather_start(cip_group = cip_group)
head(starters)

}
