% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_filter.R
\name{grad_filter}
\alias{grad_filter}
\title{Filter degree data for graduating students}
\usage{
grad_filter(
  data = NULL,
  filter_by = NULL,
  ...,
  id = "id",
  cip6 = "cip6",
  term = "term_degree"
)
}
\arguments{
\item{data}{Data frame of student IDs, academic terms, and CIP codes, default \code{midfielddegrees}.}

\item{filter_by}{Atomic character vector of 6-digit CIP codes specifying the programs to filter by.}

\item{...}{Not used for values, forces later arguments to bind by name}

\item{id}{Optional argument, the quoted column name of the student ID variable in \code{data}. Default is "id".}

\item{cip6}{Optional argument, the quoted column name of the 6-digit CIP code variable in \code{data}. Default is "cip6".}

\item{term}{Optional argument, the quoted column name of the term variable in \code{data}. Default is "term".}
}
\value{
Data frame with character variables for student ID and program CIP
}
\description{
Filter academic degree data to find all students graduating from specified programs.
}
\details{
The \code{midfielddata} package must be installed to provide the \code{midfielddegrees} data frame as the default value for the \code{data} argument. Optionally, a data frame having the same structure as \code{midfielddegrees} may be used as the value for the \code{data} argument.

The \code{filter_by} argument is an atomic character vector of 6-digit CIP codes used to filter \code{data} and extract student IDs and terms.

The function returns a subset of \code{data} with every unique combination of student ID and CIP of the program(s) in which they earned their first degree(s). Only the student ID and program CIP are returned; other variables in \code{data} are quietly dropped.

Optional arguments. Student ID variables in the midfielddata data sets are named "id". If your data frames use a different name, you can either 1) rename your variables to "id", or 2) use the optional \code{id} argument to pass the alternate variable name to the function. The same is true for the \code{cip6} and \code{term} variables.
}
\examples{
library("midfielddata")
(grad_filter(filter_by = "540104"))

}
\seealso{
\code{\link[midfieldr]{cip6_select}} for selecting 6-digit CIP codes and naming programs.
}
