% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_filter.R
\name{grad_filter}
\alias{grad_filter}
\title{Gather graduates}
\usage{
grad_filter(series, ..., reference = NULL, id = "id", cip6 = "cip6",
  term = "term_degree")
}
\arguments{
\item{series}{atomic character vector specifying the 6-digit CIP codes of
the programs to filter by}

\item{...}{not used for values, forces later arguments to bind by name}

\item{reference}{a reference data frame from which graduating student IDs,
CIP codes, and terms are obtained, default \code{midfielddegrees}}

\item{id}{character column name of the ID variable in \code{reference}}

\item{cip6}{character column name of the CIP code variable in \code{reference}}

\item{term}{character column name of term variable in \code{reference}}
}
\value{
Data frame with character variables for student ID and program CIP
}
\description{
Filter academic term data to find all students graduating from specified
programs.
}
\details{
To use this function, the \code{midfielddata} package must be installed to
provide \code{midfielddegrees}, the default reference data set.

The \code{series} argument is an atomic character vector of 6-digit CIP
codes used to filter the reference data set and extract student IDs and terms.

The function returns a subset of \code{reference} with every unique
combination of student ID and CIP of the program(s) in which they earned
their first degree(s). Only the student ID and program CIP are returned;
other variables in \code{reference} are quietly dropped.

Optional arguments. An alternate reference data set can be assigned
via the \code{reference} argument. The alternate must have variables for
student ID, 6-digit CIP, and terms in which degrees were earned with the
same structure as in \code{midfielddegrees}. Use the  \code{id}, \code{cip6},
and \code{term} arguments to reassign the variables names, if necessary, to
match the variable names in the alternate.
}
\examples{
grad <- grad_filter(series = "540104")
grad

}
\seealso{
\code{\link[midfieldr]{cip_filter}} for obtaining 6-digit CIP codes
}
