% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{graph_functions}
\alias{expon_scale_x_log10}
\alias{midfield_theme}
\title{Graph functions}
\usage{
expon_scale_x_log10()

midfield_theme()
}
\value{
Modifies an existing graph
}
\description{
Provides a set of customized layers for graphs created with for \code{ggplot2}.
}
\details{
\describe{

\item{\code{expon_scale_x_log10()}}{Applies the usual \code{scale_x_log10()} function with the scale marked with powers of ten in exponential form and the logarithmic minor grid lines shown.}

\item{\code{midfield_theme()}}{Applies \code{theme_minimal()} with some additional edits.}
}
}
