% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_undergrad.R
\name{is_undergrad}
\alias{is_undergrad}
\title{Categorize qualification level by term}
\usage{
is_undergrad(dframe, midf_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \code{{mcid}} and one of
\verb{\{term, term_course, term_degree\}.}}

\item{midf_table}{\code{degree} data frame with required variables
\verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added and all unique columns are
preserved. If required to prevent overwriting, new column names are
suffixed (\verb{.1, .2,} etc.). Duplicate columns, differing by suffix only,
are dropped. The new variables are:
\itemize{
\item \code{bacc_term}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA}.
Joined from the \code{term_degree} variable in \code{midf_table.}
\item \code{term_focus}    Character. Indicating a term contributes to study
before or after a student's first baccalaureate.
Possible values are "undergrad" and "post-bacc."
}
}
}
\description{
Categorize the qualification level towards which a student is working in
each term. Two levels are used: “undergrad” for terms before a student's
first degree and “post-bacc” (post-baccalaureate) for terms after the first
degree. Added columns support the findings. Post-baccalaureate terms are
typically excluded from the \verb{term, course,} and \code{degree} data tables.
}
\examples{
# select min required columns
term <- toy_term[, .(mcid, term)]
course <- toy_course[, .(mcid, term_course)]
degree <- toy_degree[, .(mcid, term_degree)]

# labeling terms by group: undergrad & grad
term <- is_undergrad(term, midf_table = degree)
course <- is_undergrad(course, midf_table = degree)
degree <- is_undergrad(degree, midf_table = degree)

# results
term[order(-term_focus)]
term[, .N, by = "term_focus"][order(-N)]

course[order(-term_focus)]
course[, .N, by = "term_focus"][order(-N)]

degree[order(-term_focus)]
degree[, .N, by = "term_focus"][order(-N)]
}
