% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_demographics.R
\name{join_demographics}
\alias{join_demographics}
\title{Join student demographics to a data frame}
\usage{
join_demographics(df)
}
\arguments{
\item{df}{A data frame of student attributes that includes the variable \code{MID}.}
}
\value{
The incoming data frame plus two new columns \code{ETHNIC} and \code{SEX}. Other columns are unaffected.
}
\description{
Add variables \code{ETHNIC} and \code{SEX} from the \code{student} dataset to a data frame. `dplyr::left_join()` is the joining function and \code{MID} is the join-by variable.
}
