% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_demographics.R
\name{join_demographics}
\alias{join_demographics}
\title{Join student demographics to a data frame}
\usage{
join_demographics(df)
}
\arguments{
\item{df}{A data frame of student attributes that includes the variable \code{ID}.}
}
\value{
The incoming data frame plus two new columns \code{sex} and \code{race}. Other columns are unaffected.
}
\description{
Add variables \code{race} and \code{sex} from the \code{midfieldstudents} dataset to a data frame. `dplyr::left_join()` is the joining function and \code{ID} is the join-by variable.
}
