% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kable2html}
\alias{kable2html}
\title{Vignette table}
\usage{
kable2html(x, font_size = NULL)
}
\arguments{
\item{x}{data frame}

\item{font_size}{(optional) in points, 11 pt default}
}
\description{
Prints a data frame as an HTML table in a vignette. Uses knitr::kable
and kableExtra::kable_styling to adjust font size. Function is exported
to simplify use in vignettes.
}
\keyword{internal}
