% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{math_stat}
\alias{math_stat}
\title{CIP codes for mathematics and statistics}
\format{A character vector.}
\source{
CIP 2010, US National Center for Education Statistics,
  \url{https://nces.ed.gov/ipeds/cipcode}.
}
\usage{
math_stat
}
\description{
Classification of Instructional Programs (CIP) 6-digit codes for mathematics
and statistics programs, series 27.
}
\examples{
head(math_stat, n = 10L)

print(cip_filter(math_stat))
}
\keyword{datasets}
