% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midfield_fye}
\alias{midfield_fye}
\title{FYE students and imputed starting majors}
\format{
A data frame (tibble) with 5992 observations and 2 variables.
All variables are characters.

\describe{
   \item{id}{unique, anonymized MIDFIELD student identifier}
   \item{cip6}{imputed starting major (6-digit CIP code) of FYE students}
}
}
\source{
\href{https://midfieldr.github.io/midfielddata/}{\pkg{midfielddata}
 package}
}
\usage{
midfield_fye
}
\description{
Imputed starting majors for nearly 6000 first-year engineering (FYE)
students in the \pkg{midfielddata} package.
}
\details{
Some US institutions have first year engineering (FYE) programs---typically
a common first year curriculum that is a prerequisite for declaring
an engineering major. For some persistence metrics, we have to determine
what we call a "starting major" for FYE students---the predicted
engineering program such as Civil, Electrical, or Mechanical, the
student would have declared had they not been required to enroll in FYE.
}
\seealso{
The "Impute FYE starting majors" vignette.

Other cip_data: 
\code{\link{cip}}
}
\concept{cip_data}
\keyword{datasets}
