% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion_status.R, R/data_sufficiency.R,
%   R/filter_programs.R, R/midfieldr-deprecated.R, R/select_basic_cols.R,
%   R/timely_term.R
\name{add_completion_status}
\alias{add_completion_status}
\alias{add_data_sufficiency}
\alias{filter_cip}
\alias{midfieldr-deprecated}
\alias{select_required}
\alias{add_timely_term}
\title{midfieldr deprecated functions}
\usage{
add_completion_status(dframe, midfield_degree = degree)

add_data_sufficiency(dframe, midfield_term = term)

filter_cip(keep_text = NULL, drop_text = NULL, cip = NULL, select = NULL)

select_required(midfield_x, select_add = NULL)

add_timely_term(
  dframe,
  midfield_term = term,
  ...,
  sched_span = NULL,
  span = NULL
)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble).}

\item{midfield_degree}{Data frame or data frame extension of a MIDFIELD \emph{degree} table.}

\item{midfield_term}{Data frame or data frame extension of a MIDFIELD \emph{term} table.}

\item{keep_text}{Deprecated \code{filter_cip()}. Character vector of search text to keep.}

\item{drop_text}{Deprecated \code{filter_cip()}. Character vector of search text to drop.}

\item{cip}{Deprecated \code{filter_cip()}. Data frame of programs to be searched.}

\item{select}{Deprecated \code{filter_cip()}. Character vector of column names to select.}

\item{midfield_x}{Deprecated \code{select_required()}. Data frame from which
columns are selected.}

\item{select_add}{Deprecated \code{select_required()}. Character vector of
col_patterns to search \code{dframe} column names.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{sched_span}{Integer scalar}

\item{span}{Integer scalar}
}
\description{
These functions were deprecated in midfieldr 1.0.4.
}
\details{
\describe{

\item{\code{add_completion_status()}}{is deprecated in favor of
\code{\link[=completion_status]{completion_status()}}. Update midfieldr file names and argument names,
dropping columns not used by the function, and preserving data frame
class.}

\item{\code{add_data_sufficiency()}}{is deprecated in favor of
\code{\link[=data_sufficiency]{data_sufficiency()}}. Update midfieldr file names and argument names,
dropping columns not used by the function, and preserving data frame
class.}

\item{\code{add_timely_term()}}{is deprecated in favor of
\code{\link[=timely_term]{timely_term()}}. Update midfieldr file names and argument names,
dropping columns not used by the function, and preserving data frame
class.}

\item{\code{filter_cip()}}{is deprecated in favor of
\code{\link[=filter_programs]{filter_programs()}}. The new function is similar but with the CIP
data frame as the first argument, enabling chained functions like those
encountered using dplyr and friends.}

\item{\code{select_required()}}{is deprecated in favor of
\code{\link[=select_basic_cols]{select_basic_cols()}}. The new functionality is similar but with
exact matching to the default column names plus preserving data
frame class.}

}
}
