% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr-deprecated.R
\name{midfieldr-deprecated}
\alias{midfieldr-deprecated}
\title{Deprecated functions in midfieldr}
\description{
These functions still work but will be removed (defunct) in a future version.
}
\details{
\itemize{
\item \code{\link{select_required}} is replaced by \code{\link{select_basic_cols}}.
}
}
