% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cip_rows.R, R/midfieldr-deprecated.R,
%   R/select_record_cols.R
\name{filter_cip}
\alias{filter_cip}
\alias{midfieldr-deprecated}
\alias{select_required}
\title{midfieldr deprecated functions}
\usage{
filter_cip(keep_text = NULL, drop_text = NULL, cip = NULL, select = NULL)

select_required(midfield_x, select_add = NULL)
}
\arguments{
\item{keep_text}{Deprecated \code{filter_cip()}. Character vector of search text to keep.}

\item{drop_text}{Deprecated \code{filter_cip()}. Character vector of search text to drop.}

\item{cip}{Deprecated \code{filter_cip()}. Data frame of programs to be searched.}

\item{select}{Deprecated \code{filter_cip()}. Character vector of column names to select.}

\item{midfield_x}{Deprecated \code{select_required()}. Data frame from which
columns are selected.}

\item{select_add}{Deprecated \code{select_required()}. Character vector of
col_patterns to search \code{dframe} column names.}
}
\description{
These functions were deprecated in midfieldr 1.0.4.
}
\details{
\describe{

\item{\code{filter_cip()}}{is deprecated in favor of \code{\link[=filter_cip_rows]{filter_cip_rows()}}. The
new function is similar but with the CIP data frame as the first argument,
enabling chained functions like those encountered using dplyr and friends.}

\item{\code{select_required()}}{is deprecated in favor of \code{\link[=select_record_cols]{select_record_cols()}}.
The new functionality is similar but with exact matching to the default
column names plus preserving data frame class.}

}
}
