% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion_status.R, R/filter_programs.R,
%   R/midfieldr-deprecated.R, R/select_records.R
\name{add_completion_status}
\alias{add_completion_status}
\alias{filter_cip}
\alias{midfieldr-deprecated}
\alias{select_required}
\title{midfieldr deprecated functions}
\usage{
add_completion_status(dframe, midfield_rec = degree)

filter_cip(keep_text = NULL, drop_text = NULL, cip = NULL, select = NULL)

select_required(midfield_x, select_add = NULL)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble). Required variables: \verb{\{mcid, timely_term\}}.}

\item{midfield_rec}{MIDFIELD records \emph{degree} data frame or data frame extension. Required variables:
\verb{\{mcid, term_degree\}}.}

\item{keep_text}{Deprecated \code{filter_cip()}. Character vector of search text to keep.}

\item{drop_text}{Deprecated \code{filter_cip()}. Character vector of search text to drop.}

\item{cip}{Deprecated \code{filter_cip()}. Data frame of programs to be searched.}

\item{select}{Deprecated \code{filter_cip()}. Character vector of column names to select.}

\item{midfield_x}{Deprecated \code{select_required()}. Data frame from which
columns are selected.}

\item{select_add}{Deprecated \code{select_required()}. Character vector of
col_patterns to search \code{dframe} column names.}
}
\description{
These functions were deprecated in midfieldr 1.0.4.
}
\details{
\describe{

\item{\code{add_completion_status()}}{is deprecated in favor of
\code{\link[=completion_status]{completion_status()}}. Name change plus preserves data frame class.}

\item{\code{filter_cip()}}{is deprecated in favor of
\code{\link[=filter_programs]{filter_programs()}}. The new function is similar but with the CIP
data frame as the first argument, enabling chained functions like those
encountered using dplyr and friends.}

\item{\code{select_required()}}{is deprecated in favor of
\code{\link[=select_records]{select_records()}}. The new functionality is similar but with
exact matching to the default column names plus preserving data
frame class.}

}
}
