% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr-deprecated.R, R/same_content.R
\name{midfieldr-deprecated}
\alias{midfieldr-deprecated}
\alias{same_content}
\title{Deprecated functions in package midfieldr}
\usage{
same_content(x, y)
}
\description{
The functions listed below are deprecated. Alternative functions with
similar functionality are mentioned. Help pages for deprecated functions
are available at \code{help("<function>-deprecated")}.
}
\section{\code{same_content}}{

For \code{same_content()}, use \code{check_equiv_frames()}
}

\keyword{internal}
