% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr-package.R
\docType{package}
\name{midfieldr-package}
\alias{midfieldr-package}
\title{Tools for Studying MIDFIELD Student Unit Record Data in R}
\description{
The goal of midfieldr is to provide tools for working with MIDFIELD data,
a resource of longitudinal, de-identified, individual student unit records
from multiple US institutions.
}
\details{
\href{https://engineering.purdue.edu/MIDFIELD}{MIDFIELD} (as of May 2021)
contains individual student record data for 1.7 M unique students at 33
US institutions. MIDFIELD is large enough to permit disaggregation by
multiple characteristics simultaneously, enabling researchers to examine
student characteristics (race/ethnicity, sex, prior achievement) and
curricular pathways (including coursework and major) by institution and
over time.

A proportionate stratified sample of these data (for practice) is
available in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}, an R
data package with longitudinal student unit-records for 98,000
undergraduates at 12 institutions from  1987--2016 organized in four data
tables:
\itemize{
 \item {\href{https://midfieldr.github.io/midfielddata/reference/student.html}{student}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/course.html}{course}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/term.html}{term}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/degree.html}{degree}}
}

The tools in midfieldr work equally well with the practice data in
midfielddata and the research data in MIDFIELD.

This work is supported by a grant from the US National Science
Foundation (EEC 1545667).
}
\concept{package}
