% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr-package.R
\docType{package}
\name{midfieldr-package}
\alias{midfieldr-package}
\title{Tools for Studying MIDFIELD Student Unit Record Data in R}
\description{
The goal of midfieldr is to provide tools for working with MIDFIELD
data, a resource for research on undergraduate students'
trajectories through the university that includes longitudinal,
anonymized, whole population data for multiple US institutions.
}
\section{Background}{

\href{https://engineering.purdue.edu/MIDFIELD}{MIDFIELD} (as of May 2021)
contains individual student record data for 1.7 M unique students at 33
US institutions. MIDFIELD is large enough to permit disaggregation by
multiple characteristics simultaneously, enabling researchers to examine
student characteristics (race/ethnicity, sex, prior achievement) and
curricular pathways (including coursework and major) by institution and
over time.

A proportionate stratified sample of these data (for practice) is
available in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}, an R
data package with longitudinal student unit-records for 98,000
undergraduates at 12 institutions from  1987--2016 organized in four data
tables:
\itemize{
 \item {\href{https://midfieldr.github.io/midfielddata/reference/student.html}{student}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/course.html}{course}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/term.html}{term}}
 \item {\href{https://midfieldr.github.io/midfielddata/reference/degree.html}{degree}}
}

The tools in midfieldr work equally well with the practice data in
midfielddata and the research data in MIDFIELD.
}

\section{Data}{

midfielddata provides a proportionate stratified random sample
of the MIDFIELD research data. The sampling strata are institution, cip4
(the first four digits of the 6-digit CIP code), transfer status,
race/ethnicity, and sex. Contains four data tables: \code{student},
\code{course}, \code{term}, and \code{degree} keyed by an anonymized
student ID.

Complete MIDFIELD data suitable for student-records research are
available to researchers under the following conditions:
\enumerate{
  \item{Your institutional IRB has granted approval for your project
  to study students using MIDFIELD. At most institutions, the
  use of MIDFIELD data for research is  in the IRB "Exempt" category,
  but institutional practices vary.}
  \item{Each researcher using the data signs a letter of
  confidentiality describing  the guidelines for how the data may be
  reported.}
}
The research data and practice data have the same structure with
the same variable names, though some research variables are
omitted from the practice data. Thus R scripts written for the practice
data should generally work with the research data.
}

\references{
\insertAllCited{}
}
\concept{package}
