% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr-package.R
\docType{package}
\name{midfieldr-package}
\alias{midfieldr-package}
\title{\code{midfieldr}: Tools for Student Records Research}
\description{
For investigating undergraduate student records. Designed to work with data
from the MIDFIELD project, a partnership of post-secondary US institutions.
MIDFIELD contains registrar's data for 1.7M undergraduates in all majors at
19 institutions from 1987--2019.
}
\details{
Functions in midfieldr are designed to work with the MIDFIELD research data
(available to MIDFIELD partners) as well as with the
publicly-available practice data in the midfielddata package.

Operations such subsetting, joining, grouping, and summarizing
are performed using one's preferred tools in base R or more advanced systems
such as dplyr or data.table. The vignettes use base R and data.table
for data carpentry and ggplot2 for graphs.
}
\concept{package}
