% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr-package.R
\docType{package}
\name{midfieldr-package}
\alias{midfieldr-package}
\title{\code{midfieldr}: Tools for Student Records Research}
\description{
The *Multiple-Institution Database for Investigating Engineering
Longitudinal Development* (MIDFIELD) is a partnership of US higher
education institutions with engineering programs. MIDFIELD contains
registrar's data for 1.7M undergraduates in all majors at 19 institutions
from 1987--2019 \insertCite{Ohland+Long:2016}{midfieldr}.
}
\details{
The software environment comprises two R packages:
\itemize{
  \item{\pkg{midfieldr} An R package providing functions specialized for
  manipulating MIDFIELD data to examine the intersectionality of
  race/ethnicity, sex, and discipline in persistence metrics such as
  stickiness (retention by a discipline) and graduation rate.}
  \item{\pkg{midfielddata} An R  package with practice data for users to
  learn about student record analysis using R. However, these data are
  not suitable for drawing inferences about student performance, i.e.,
  not for research.}
}

\pkg{midfielddata} provides a proportionate stratified random sample
of the MIDFIELD research data. The sampling strata are institution, cip4
(the first four digits of the 6-digit CIP code), transfer status,
race/ethnicity, and sex. Contains four data tables: \code{student},
\code{course}, \code{term}, and \code{degree} keyed by an anonymized
student ID.

Complete MIDFIELD data suitable for student-records research are
available to researchers under the following conditions:
\enumerate{
  \item{Your institutional IRB has granted approval for your project
  to study students using MIDFIELD. At most institutions, the
  use of MIDFIELD data for research is  in the IRB "Exempt" category,
  but institutional practices vary.}
  \item{Your institutional IRB has granted approval for your project
  to study students using MIDFIELD. At most institutions, the
  use of MIDFIELD data for research is  in the IRB "Exempt" category,
  but institutional practices vary.}
}
}
\references{
\insertAllCited{}
}
\concept{package}
