% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr-package.R
\docType{package}
\name{midfieldr-package}
\alias{midfieldr-package}
\title{Tools and Methods for Working with MIDFIELD Data in R}
\source{
Data provided by MIDFIELD: \url{https://midfield.online/}.
}
\description{
Provides tools and demonstrates methods for working with individual undergraduate student-level records (registrar's data) in R. Tools include filters for program codes, data sufficiency, and timely completion. Methods include gathering blocs of records, computing quantitative metrics such as graduation rate, and creating charts to visualize comparisons. midfieldr is designed to work with data from the MIDFIELD research database, a sample of which is available in the midfielddata data package.
}
\details{
\href{https://midfield.online/}{MIDFIELD}.   A database of anonymized student-level records for approximately 1.7M undergraduates at nineteen US institutions from 1987--2018, of which midfielddata  provides a sample. This research database is currently accessible to MIDFIELD partner institutions only.

\href{https://midfieldr.github.io/midfielddata/}{midfielddata}.   An R data package that supplies anonymized student-level records for 98,000 undergraduates at three US institutions from 1988--2018. A sample of the MIDFIELD database, midfielddata provides practice data for the tools and methods in the midfieldr package.

This work is supported by the US National Science Foundation through grant numbers 1545667 and 2142087.
}
\concept{package}
