% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr-package.R
\docType{package}
\name{midfieldr-package}
\alias{midfieldr}
\alias{midfieldr-package}
\title{midfieldr: Tools and Methods for Working with MIDFIELD Data in 'R'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools and demonstrates methods for working with individual undergraduate student-level records (registrar's data) in 'R'. Tools include filters for program codes, data sufficiency, and timely completion. Methods include gathering blocs of records, computing quantitative metrics such as graduation rate, and creating charts to visualize comparisons. 'midfieldr' interacts with practice data provided in 'midfielddata', an R data package available at \url{https://midfieldr.github.io/midfielddata/}. 'midfieldr' also interacts with the full MIDFIELD database for users who have access. This work is supported by the US National Science Foundation through grant numbers 1545667 and 2142087.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://midfieldr.github.io/midfieldr/}
  \item Report bugs at \url{https://github.com/MIDFIELDR/midfieldr/issues}
}

}
\author{
\strong{Maintainer}: Richard Layton \email{graphdoctor@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Russell Long \email{ralong@purdue.edu} [copyright holder, data manager]
  \item Matthew Ohland \email{ohland@purdue.edu} [copyright holder]
  \item Marisa Orr \email{marisak@clemson.edu} [copyright holder]
  \item Susan Lord \email{slord@sandiego.edu} [copyright holder]
}

}
\keyword{internal}
