% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr.R
\docType{package}
\name{midfieldr}
\alias{midfieldr}
\alias{midfieldr-package}
\title{Undergraduate Student Records for Longitudinal Research}
\source{
Data provided by the MIDFIELD project:
\url{https://engineering.purdue.edu/MIDFIELD}
}
\description{
A package for investigating a data sample from the MIDFIELD database.
The data comprise student records from registrars at participating US
universities, including demographic, term, course, and degree information
for 97,640 undergraduate students from 1987 to 2016.
}
\details{
In our examples illustrating how to use \pkg{midfieldr} tools and datasets,
the findings are generally conditioned by instructional program, student sex,
and student race/ethnicity.  Sex and race/ethnicity categories are limited to
those reported by the US institutions contributing to the MIDFIELD database.

Student record data is provided in four data packages:
\pkg{midfieldstudents}, \pkg{midfieldcourses}, \pkg{midfieldterms}, and
\pkg{midfielddegrees}.
}
\section{Data}{

\describe{

  \item{\code{cip}}{A dataset of names and codes of academic programs. Each of the 1544 observations is a program described by a 6-digit Classification of Instructional Programs (CIP) code defined by the US Department of Education in 2010.}

  \item{\code{midfieldstudents}}{In the \href{https://github.com/MIDFIELDR/midfieldstudents}{\pkg{midfieldstudents}} data package.
 A tidy data frame with 97,640 observations and 15 demographic variables. Each observation is a unique student. Occupies 19 Mb of memory.}

  \item{\code{midfieldcourses}}{In the \href{https://github.com/MIDFIELDR/midfieldcourses}{\pkg{midfieldcourses}}  data package. A tidy data frame with 3.5 M observations and 12 academic course variables. Each observation is one course in one term for one student. Occupies 348 Mb of memory.}

  \item{\code{midfieldterms}}{In the \href{https://github.com/MIDFIELDR/midfieldterms}{\pkg{midfieldterms}}  data package.
  A tidy data frame with 729,014 observations and 13 academic term variables. Each observation is one term for one student. Occupies 82 Mb of memory.}

  \item{\code{midfielddegrees}}{In the \href{https://github.com/MIDFIELDR/midfielddegrees}{\pkg{midfielddegrees}} data package.
  A tidy data frame with 97,640 observations and 5 graduation variables. Each observation is a unique student. Of these students, 47,499 earned their degrees. Occupies 10 Mb of memory.}

}
}

\section{Functions}{

This package provides functions to access, manipulate, and graph the student
record data.

\describe{

  \item{\code{cip_filter()}}{Filter a data frame of Classification of
  Instructional Programs (CIP) codes to return the rows that match conditions.}

  \item{\code{join_demographics()}}{Add variables \code{ETHNIC} and
  \code{SEX} from the \code{student} dataset to a data frame.}

}
}

