% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr.R
\docType{package}
\name{midfieldr}
\alias{midfieldr}
\alias{midfieldr-package}
\title{Undergraduate Student Records For Longitudinal Research
\Sexpr[results=hide]{load("data/data_bits.rda")}}
\source{
Data provided by the MIDFIELD project: \url{https://engineering.purdue.edu/MIDFIELD}
}
\description{
A package for investigating a sample from the MIDFIELD database. The data  comprise student records from registrars at participating US universities, including demographic, term, course, and degree information for \Sexpr{prettyNum(round(data_bits[["obs_student"]], 0), big.mark = ",")} undergraduate students from \Sexpr{data_bits[["year_min"]]} to \Sexpr{data_bits[["year_max"]]}.
}
\details{
In our examples illustrating how to use \pkg{midfieldr} tools and datasets, the findings are generally disaggregated or conditioned by instructional program and student sex and race/ethnicity.  Sex and race/ethnicity categories are limited to those reported by the US institutions contributing to the MIDFIELD database.

Datasets are provided with \pkg{midfieldr} plus four data packages: \pkg{midfieldstudents}, \pkg{midfieldcourses}, \pkg{midfieldterms}, and \pkg{midfielddegrees}.
}
\section{Data}{

\describe{

  \item{\code{cip}}{\href{https://github.com/MIDFIELDR/midfieldr}{midfieldr} package. A dataset of program codes and names of academic fields of study. Each of the 1552 observations is one program at the CIP 6-digit level.}

  \item{\code{midfieldstudents}}{\href{https://github.com/MIDFIELDR/midfieldstudents}{midfieldstudents} package. Demographic data for \Sexpr{prettyNum(round(data_bits[["obs_student"]], 0), big.mark = ",")} students. Each observation is a unique student.}

  \item{\code{midfieldcourses}}{\href{https://github.com/MIDFIELDR/midfieldcourses}{midfieldcourses} package. Academic course data for \Sexpr{prettyNum(round(data_bits[["obs_student"]], 0), big.mark = ",")}. Each observation is one course for one student, yielding \Sexpr{round(data_bits[["obs_course"]]/1e+6, 1)} M observations.}

  \item{\code{midfieldterms}}{\href{https://github.com/MIDFIELDR/midfieldterms}{midfieldterms} package. Academic term data for \Sexpr{prettyNum(round(data_bits[["obs_student"]], 0), big.mark = ",")} students. Each observation is one term for one student, yielding \Sexpr{prettyNum(round(data_bits[["obs_term"]], 0), big.mark = ",")} observations.}

  \item{\code{midfielddegrees}}{\href{https://github.com/MIDFIELDR/midfielddegrees}{midfielddegrees} package. Graduation data for  \Sexpr{prettyNum(round(data_bits[["obs_degree"]], 0), big.mark = ",")} students receiving degrees. Each observation is a unique student. }

}
}

\section{Functions}{

midfieldr provides functions to access, manipulate, and graph the student record data.

\describe{

  \item{\code{cip_filter()}}{Filter a data frame of Classification of Instructional Programs (CIP) codes to return the rows that match conditions.}

  \item{\code{join_demographics()}}{Add variables \code{ETHNIC} and \code{SEX} from the \code{student} dataset to a data frame.}

  }
}

