% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midfieldr.R
\docType{package}
\name{midfieldr}
\alias{midfieldr}
\alias{midfieldr-package}
\title{Undergraduate Student Records For Longitudinal Research}
\description{
A package for investigating a sample from the MIDFIELD database, a dataset compiled from registrar's student records from participating US universities. The sample data include demographic, term, course, and degree information for 165,000 undergraduate students from 1990 to 2016.
}
\details{
The Multi-Institution Database for Investigating Engineering Longitudinal Development (MIDFIELD) was started with longitudinal studies of engineering students in mind (the "E" in MIDFIELD). However, the data comprise complete registrar's student records---any academic program at those institutions can be studied.
}
\section{Data}{

Sample data from the MIDFIELD database are provided in 3 additional packages. The datasets are:

\describe{

  \item{\code{student}}{midfieldstudent package. Demographic data for 165,000 students. Each observation is a unique student.}

  \item{\code{term}}{midfieldterm package. Academic term data for 165,000 students. Each observation is one term for one student, yielding 1.1 M observations.}

  \item{\code{course}}{midfieldcourse package. Academic course data for 165,000 students. Each observation is one course for one student, yielding 5.4 M observations.}

  \item{\code{degree}}{midfieldstudent package. Graduation data for 90,000 students receiving degrees. Each observation is a unique student. }

  \item{\code{cip}}{midfieldr package. A dataset of program codes and names of academic fields of study. Each of the 1552 observations is one program at the CIP 6-digit level. \code{?cip} for more information.}

}
}

