% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_session}
\alias{my_session}
\title{Creates my own shorter version of session_info for vignettes
from the sessioninfo package}
\usage{
my_session(package_string)
}
\description{
Creates my own shorter version of session_info for vignettes
from the sessioninfo package
}
