% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_series.R
\docType{data}
\name{named_series}
\alias{named_series}
\alias{bio_sci}
\alias{engr}
\alias{math_stat}
\alias{other_stem}
\alias{phys_sci}
\alias{stem}
\title{Named series of CIP  codes}
\format{A character vector of 6-digit CIP codes}
\usage{
bio_sci

engr

math_stat

other_stem

phys_sci

stem
}
\description{
Collections of 6-digit CIP codes for groups of programs.
}
\details{
The groups of programs include Engineering, Physical Sciences,

Using the US National Science Foundation's definition of Science, Technology, Engineering, and Mathematics (STEM) programs, we have collected CIP codes and program names for STEM, Mathematics and Statistics, Engineering, Physical Sciences, etc.

A named series is used as a series argument in \code{filter_cip()}.
}
\examples{
engr
filter_cip(engr)
}
\keyword{datasets}
