% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_multiway.R
\name{order_multiway}
\alias{order_multiway}
\title{Order multiway categories}
\usage{
order_multiway(
  dframe,
  quantity,
  categories,
  ...,
  method = NULL,
  ratio_of = NULL
)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with the following required variables: two
independent categorical variables, one quantitative response
variable, and, if \code{method = percent}, its dividend and divisor
variables.}

\item{quantity}{Character. Name of the single multiway quantitative
variable.}

\item{categories}{Character. Vector of names of the two multiway
categorical variables, in any order.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{method}{Character. Method of ordering the levels of the categories;
possible values are “median” (default) or “percent”. The median
method determines medians of the quantitative column grouped by
category. The percent method sums dividends and divisors by category
and calculates their quotients by category.}

\item{ratio_of}{Character. Vector of column names of the dividend
and the divisor that produced the quantitative variable. Names
must be in order, as in \verb{c(dividend, divisor).} Required
when \code{method = "percent,"} ignored otherwise.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Row order is preserved. Duplicated rows are removed.
\item Column specified by \code{quantity} is converted to type double.
Columns specified by \code{categories} are converted to factors and ordered.
\item New columns are added unless they duplicate existing
variables---redundant columns, whether new or existing, are dropped
(see Details). The new columns, with capital letters denoting
placeholders, have the form:
\itemize{
\item \code{CATEGORY_1_LABEL}
\item \code{CATEGORY_2_LABEL}
}

The \code{CATEGORY} placeholder is replaced with the column names from
\code{categories.} The \code{LABEL} placeholder depends on the method. For
example, suppose \verb{categories = c("program", "people").}For the \code{median}
method, the new column names would be \code{program_median}
and \code{people_median.} For the \code{percent} method, the new column names would
be \code{program_metric} and \verb{people_metric,} indicating that the metric had
been calculated for each category independently.
}
}
\description{
Condition data for Cleveland multiway charts. Two independent categorical
variables are converted to factors with their levels ordered by the
single quantitative response variable.
}
\details{
Multiway data comprise a single quantitative value (or response) for
every combination of levels of two categorical variables. The ordering of
the rows and panels, based on the response quantity, is crucial to the
perception of effects (Cleveland, 1993).

Multiway data comprise three variables: a categorical variable of
\eqn{\small m} levels; a second independent categorical variable of
\eqn{\small n} levels; and a quantitative variable (or \emph{response}) of
length \eqn{\small m \times n} that cross-classifies the categories,
that is, there is a value of the response for each combination of levels
of the two categorical variables. If a response value is missing, it is
assumed that a response for every combination is at least feasible.

In a multiway dot plot, one category is encoded by the panels, the second
category is encoded by the rows of each panel, and the quantitative variable
is encoded along identical horizontal scales.

\emph{Redundant columns.} To prevent overwriting, the name of an added
variable such as  \code{program_median}  that matches that of an existing
variable is made unique by adding a suffix, e.g., \code{program_median.1.}
An added variable that otherwise duplicates the existing variable is
redundant and dropped. If not, the presence of the suffixed variable
indicates a potential error. The variables added by
midfieldr functions depend on fixed quantities, e.g., a student's
admission term or an institution's data range, and do not change
during an analysis.
}
\examples{
# Reconfigure built-in data set
DT <- study_results[program == "EE" | program == "ME"]
DT <- DT[race \%chin\% c("Asian", "Black", "Hispanic", "White")]
DT[, people := paste(race, sex)]
DT[, c("race", "sex") := NULL]
data.table::setnames(DT, 
                     old = c("program", "graduates", "ever_enrolled", "stickiness"), 
                     new = c("prgm", "grad", "ever", "stk"))
data.table::setcolorder(DT, c("prgm", "people", "grad", "ever", "stk"))
DT[]

# Factor levels ordered by median
DT1 <- data.table::copy(DT)
DT1 <- DT1[, c("ever", "grad") := NULL]
mw1 <- order_multiway(DT1, 
                      quantity = "stk", 
                      categories = c("prgm", "people"))
data.table::setorderv(mw1, c("prgm_median", "people_median"))
mw1

# Levels in increasing order
levels(mw1$prgm)
levels(mw1$people)

# No change if added columns duplicate existing
mw1a <- order_multiway(mw1, 
                       quantity = "stk", 
                       categories = c("prgm", "people"))
check_equiv_frames(mw1, mw1a)

# Ordering using percent method
mw2 <- order_multiway(DT, 
                      quantity = "stk", 
                      categories = c("prgm", "people"), 
                      method = "percent", 
                      ratio_of = c("grad", "ever"))
data.table::setorderv(mw2, c("prgm_metric", "people_metric"))

# The two ratio_of variables `ever` and `grad` are retained
mw2

# Levels in same increasing order as shown above
levels(mw2$prgm)
levels(mw2$people)

# Order of factor levels depends on the method. Here, for example, 
# program levels are the same for median and percent methods, 
all.equal(levels(mw1$prgm), levels(mw2$prgm))

# but people levels do not have the same order. 
all.equal(levels(mw1$people), levels(mw2$people))
levels(mw1$people)
levels(mw2$people)
}
\references{
Cleveland WS (1993). \emph{Visualizing Data}. Hobart Press, Summit, NJ.
}
