% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{other_stem}
\alias{other_stem}
\title{CIP codes for other STEM}
\format{A character vector.}
\source{
NSF STEM programs: \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}.
}
\usage{
other_stem
}
\description{
Classification of Instructional Programs (CIP) 6-digit codes for
all Science, Technology, Engineering, and Mathematics (STEM) programs not
included in the midfieldr named series: \code{bio_sci}, \code{engr},
\code{math_stat}, and \code{phys_sci}.
}
\examples{
head(other_stem, n = 10L)

print(cip_filter(series = other_stem))
}
\keyword{datasets}
