% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phys_sci}
\alias{phys_sci}
\title{CIP codes for programs in physical science}
\format{A character vector.}
\source{
NSF STEM programs: \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}.
}
\usage{
phys_sci
}
\description{
All 6-digit codes from the 2010 CIP 40 series.
}
\examples{
\dontrun{
 phys_sci
}
}
\keyword{datasets}
