% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_bacc_terms.R
\name{post_bacc_terms}
\alias{post_bacc_terms}
\title{Identify post-baccalaureate terms}
\usage{
post_bacc_terms(dframe, midfield_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble). Required variables: \code{{mcid}} and one of
\verb{\{term, term_course, term_degree\}.}}

\item{midfield_table}{MIDFIELD  \emph{\strong{degree}}  dataset (data frame or data frame extension). Required variables:
\verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved.
\item Rows are not modified.
\item Columns are not modified except a new column replaces an existing
column if it has the same name. The new columns are:
\itemize{
\item \code{first_degree_term.}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA}.
Joined from the \code{term_degree} variable in \code{midfield_table.}
\item \code{term_cluster.}    Character, indicating that a term belongs
to one of three clusters: terms that are prior to ("pre-degree"),
equal to ("first-degree"), or subsequent to ("post-first-degree")
the student’s first degree term.
}
\item Groups and keys are not preserved.
}
}
\description{
To a data frame keyed by student ID and containing an academic term
variable, add a column that clusters terms with respect to a student's
first degree term. Post-baccalaureate terms are typically excluded from
the term, course, and degree data tables.
}
\details{
In a typical analysis, one is interested in a student's progress up to
and including the term in which they earn their first degree or degrees.
Any terms later than the first baccalaureate can usually be excluded from
study.
}
\examples{
# reduce number of columns
term <- select_basic_cols(toy_term)
degree <- select_basic_cols(toy_degree)

# identify term-clusters in a 'term' table
x <- post_bacc_terms(term, degree)
x[, .N, by = "term_cluster"][order(-N)]
x

# identify term-clusters in a 'degree' table
x <- post_bacc_terms(degree, degree)
x[, .N, by = "term_cluster"][order(-N)]
x

# post-first-degree terms are usually dropped
x[term_cluster != "post-first-degree"]
}
