% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_bacc_terms.R
\name{post_bacc_terms}
\alias{post_bacc_terms}
\title{Identify post-baccalaureate terms}
\usage{
post_bacc_terms(dframe, midfield_rec = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble). Required variables: \code{{mcid}} and one of
\verb{\{term, term_course, term_degree\}.}}

\item{midfield_rec}{MIDFIELD records \emph{degree} data frame or data frame extension. Required variables:
\verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are not preserved.
\item Rows are not modified.
\item Columns are not modified except new columns overwrite old columns of
the same name. New columns:
\itemize{
\item \code{first_degree_term.}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA}.
Joined from \code{midfield_rec$term_degree}.
\item \code{term_cluster.}    Character, indicating that a term belongs
to one of three clusters: terms that are prior to ("pre-degree"),
equal to ("first-degree"), or subsequent to ("post-first-degree")
the student’s first degree term.
}
}
}
\description{
To a data frame keyed by student ID and containing an academic term
variable, add a column that clusters terms with respect to a student's
first degree term. Post-baccalaureate terms are typically excluded from
the term, course, and degree data tables.
}
\details{
In a typical analysis, one is interested in a student's progress up to
and including the term in which they earn their first degree or degrees.
Any terms later than the first baccalaureate can usually be excluded from
study.
}
\examples{

# Examples TBD
x <- 1
}
