% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fye_mice.R
\name{prep_fye_mice}
\alias{prep_fye_mice}
\title{Prepare FYE data for imputation}
\usage{
prep_fye_mice(m_student, m_term, fye_cip = NULL)
}
\arguments{
\item{m_student}{Data frame or data frame extension (e.g., data.table or tibble) with required character variables
\verb{\{mcid, race, sex\}.} Typically based on one's original,
unfiltered \code{student} source data without regard to data sufficiency.}

\item{m_term}{Data frame or data frame extension (e.g., data.table or tibble) with required character variables
\verb{\{mcid, term, cip6, institution\}.} Typically based on one's original,
unfiltered \code{term} source data without regard to data sufficiency.}

\item{fye_cip}{Data frame or data frame extension (e.g., data.table or tibble) with required character variables
\verb{\{institution, fye_cip6\}.} Default has one institution
value (\code{"Institution J"}) and one CIP code value (\code{"140102"}),
compatible with midfielddata practice data and the midfieldr "toy"
datasets.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Rows: One row for every degree-seeking FYE student.
\item Columns: Conditioned for later use as an input to the mice R
package for multiple imputation as follows:
\itemize{
\item \code{mcid}   Character, anonymized student identifier.
\item \code{race}   Factor, race/ethnicity from the \code{student} input data
frame. An imputation predictor variable.
\item \code{sex}   Factor, sex from the \code{student} input data
frame. An imputation predictor variable.
\item \code{institution}   Factor, anonymized institution name from the
\code{term} data frame. An imputation predictor variable.
\item \code{proxy}   Factor, 6-digit CIP code of a student's known,
first degree-granting engineering program or NA representing missing
values to be imputed.
}
}
}
\description{
Constructs a data frame of students ever enrolled in First-Year Engineering
(FYE) programs at their institutions based on information in the MIDFIELD
(or equivalent) \code{student} and \code{term} data tables. Conditions the data for
use as an input to the mice R package for multiple imputation. Sets up
three variables as predictors  (institution, race/ethnicity, and sex) and
one variable to be imputed (program CIP code) keyed by student ID.
}
\section{Background}{
  At some US institutions, engineering students are
required to complete a First-Year Engineering (FYE) program as a prerequisite
for declaring an engineering major. FYE students who then transition to a
degree-granting engineering program are  typically counted as "starters" in
that program for purposes of calculating, for example, graduation rates.

This approach invariably undercounts starters and over-estimates associated
metrics because it fails to account for FYE students who change majors
(never enrolling in an engineering major) or drop out of the database
altogether. Had the FYE program not been required, they would have enrolled
in their preferred engineering program---the count of starters would have
increased and a metric such as graduation rate would have decreased.

To include FYE students when a count of starters is needed, we estimate an
"FYE proxy", that is, the 6-digit CIP codes of the degree-granting
engineering programs that FYE students would have declared had they not been
required to enroll in FYE. The purpose of \code{prep_fye_mice()} is to prepare
the data for imputing the unknown CIP codes.
}

\section{Method}{
 The function extracts all terms for all FYE students and
identifies all engineering programs in which they were ever enrolled. A
\code{proxy} variable is added with one of the following values:
\enumerate{
\item{If a student record includes at least one, non-FYE, degree-granting
engineering program, the CIP code of the first such program is
returned as the student's FYE proxy.}
\item{If not, the proxy is NA and is treated as a missing value to be
imputed by \code{mice()}.}
}

Notes:
\itemize{
\item Missing values (NA) in the required columns are removed. However, a
value of "unknown" in a predictor column, e.g., race/ethnicity or sex,
is acceptable.
\item After running \code{prep_fye_mice()} but before running \code{mice()}, one can edit
the predictor variables if desired. The institution variable should remain
to ensure that a student's imputed program is available at their
institution.
\item The resulting data frame is ready for use as input for the mice package,
with all variables except \code{mcid} returned as factors.
}
}

\examples{
library(data.table)

# Subset student and term data using selected IDs
IDs <- c("MCID3112319668", "MCID3112214437", "MCID3112328548", 
         "MCID3111447797", "MCID3111566004", "MCID3111697452", 
         "MCID3112268500", "MCID3112320295")
student <- select_basic_cols(toy_student[mcid \%chin\% IDs])
term <- select_basic_cols(toy_term[mcid \%chin\% IDs])

# Obtain results
proxy <- prep_fye_mice(student, term)
proxy

# ---------- Examine details
# Note: the CIP code and name for FYE is 140102 Pre-Engineering

# Join program names to term data for display
term_seq <- cip[term, .(mcid, term, cip6, cip6name), on = "cip6", nomatch = NULL]

# Function to display results for individual students
f <- function(IDs, i) {
    cat(paste("Student", i, "record\n"))
    print(term_seq[mcid == IDs[i]])
    cat("\nprep_fye_mice() results\n")
    print(proxy[mcid == IDs[i]])
}

# Example 1: Non-Engineering -> FYE -> Engineering
# 400501 (Chemistry) -> FYE -> 140701 (Chemical Engng)
# FYE proxy is 140701
f(IDs, 1)

# Example 2: FYE -> Engineering -> Non-Engineering
# FYE -> 140901 (Computer Engng) -> 450601 (Economics)
# FYE proxy is 140901
f(IDs, 2)

# Example 3: FYE -> Engineering
# FYE -> 141001 (Electrical Engng)
# FYE proxy is 141001
f(IDs, 3)

# Example 4: FYE -> Engineering -> Engineering
# FYE -> 141901 (Mechanical Engng) -> 143501 (Industrial Engng)
# FYE proxy is 141901 
f(IDs, 4)

# Example 5: Non-Engineering -> FYE -> Leaves the database
# 240102 (General Studies) -> FYE
# FYE proxy is NA 
f(IDs, 5)

# Example 6: FYE -> Leaves the database
# FYE proxy is NA 
f(IDs, 6)

# Example 7: Non-Engineering -> FYE -> Non-Engineering
# 240102 (General Studies) -> FYE -> 110101 (Computer Science)
# FYE proxy is NA 
f(IDs, 7)

# Example 8: FYE -> Non-Engineering
# FYE -> 230101 (English Literature)
# FYE proxy is NA 
f(IDs, 8)
}
