% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fye_mice.R
\name{prep_fye_mice}
\alias{prep_fye_mice}
\title{Prepare FYE data for imputation}
\usage{
prep_fye_mice(m_student, m_term, fye_cip = NULL)
}
\arguments{
\item{m_student}{Data frame or data frame extension (e.g., data.table or tibble) with required character variables
\verb{\{mcid, race, sex\}.} Typically based on one's original,
unfiltered \code{student} source data without regard to data sufficiency.}

\item{m_term}{Data frame or data frame extension (e.g., data.table or tibble) with required character variables
\verb{\{mcid, term, cip6, institution\}.} Typically based on one's original,
unfiltered \code{term} source data without regard to data sufficiency.}

\item{fye_cip}{Data frame or data frame extension (e.g., data.table or tibble) with required character variables
\verb{\{institution, fye_cip6\}.} Default has one institution
value (\code{"Institution J"}) and one CIP code value (\code{"140102"}),
compatible with midfielddata practice data and the midfieldr "toy"
datasets.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Rows: One row for every degree-seeking FYE student.
\item Columns: Conditioned for later use as an input to the mice R
package for multiple imputation as follows:
\itemize{
\item \code{mcid}   Character, anonymized student identifier.
\item \code{race}   Factor, race/ethnicity from the \code{student} input data
frame. An imputation predictor variable.
\item \code{sex}   Factor, sex from the \code{student} input data
frame. An imputation predictor variable.
\item \code{institution}   Factor, anonymized institution name from the
\code{term} data frame. An imputation predictor variable.
\item \code{proxy}   Factor, 6-digit CIP code of a student's known,
first degree-granting engineering program or NA representing missing
values to be imputed.
}
}
}
\description{
Constructs a data frame of students ever enrolled in First-Year Engineering
(FYE) programs at their institutions based on information in the MIDFIELD
(or equivalent) \code{student} and \code{term} data tables. Conditions the data for
use as an input to the mice R package for multiple imputation. Sets up
three variables as predictors  (institution, race/ethnicity, and sex) and
one variable to be imputed (program CIP code) keyed by student ID.
}
\section{Background}{
  At some US institutions, engineering students are
required to complete a First-Year Engineering (FYE) program as a prerequisite
for declaring an engineering major. FYE students who then transition to a
degree-granting engineering program are  typically counted as "starters" in
that program for purposes of calculating, for example, graduation rates.

This approach invariably undercounts starters and over-estimates associated
metrics because it fails to account for FYE students who change majors
(never enrolling in an engineering major) or drop out of the database
altogether. Had the FYE program not been required, they would have enrolled
in their preferred engineering program---the count of starters would have
increased and a metric such as graduation rate would have decreased.

To include FYE students when a count of starters is needed, we estimate an
"FYE proxy", that is, the 6-digit CIP codes of the degree-granting
engineering programs that FYE students would have declared had they not been
required to enroll in FYE. The purpose of \code{prep_fye_mice()} is to prepare
the data for imputing the unknown CIP codes.
}

\section{Method}{
 The function extracts all terms for all FYE students and
identifies all engineering programs in which they were ever enrolled. A
\code{proxy} variable is added with one of the following values:
\enumerate{
\item{If a student record includes at least one, non-FYE, degree-granting
engineering program, the CIP code of the first such program is
returned as the student's FYE proxy.}
\item{If not, the proxy is NA and is treated as a missing value to be
imputed by \code{mice()}.}
}

Notes:
\itemize{
\item Missing values (NA) in the required columns are removed. However, a
value of "unknown" in a predictor column, e.g., race/ethnicity or sex,
is acceptable.
\item After running \code{prep_fye_mice()} but before running \code{mice()}, one can edit
the predictor variables if desired. The institution variable should remain
to ensure that a student's imputed program is available at their
institution.
\item The resulting data frame is ready for use as input for the mice package,
with all variables except \code{mcid} returned as factors.
}
}

\examples{
library(data.table)

# Student A enrolls in FYE (140102) followed by Mechanical Engng 
# (141901). The FYE proxy is Mechanical Engng (141901), abbreviated "ME".
one_student <- data.table(mcid = "A", race = "unknown", sex = "unknown")
term_excerpt <- data.table(
    mcid = "A", 
    institution = "Institution J",
    term = c("20013", "20021"), 
    program = c("FYE", "ME"),
    cip6 = c("140102", "141901")
)
one_student
term_excerpt
prep_fye_mice(one_student, term_excerpt)

# Student B enrolls in FYE (140102) followed by History (540101). 
# The FYE proxy is NA.
one_student <- data.table(mcid = "B", race = "unknown", sex = "unknown")
term_excerpt <- data.table(
    mcid = "B", 
    institution = "Institution J",
    term = c("20013", "20021"), 
    program = c("FYE", "History"),
    cip6 = c("140102", "540101")
)
one_student
term_excerpt
prep_fye_mice(one_student, term_excerpt)

# Student C enrolls in FYE (140102) but drops out of the database.
# The FYE proxy is NA.
one_student <- data.table(mcid = "C", race = "unknown", sex = "unknown")
term_excerpt <- data.table(
    mcid = "C", 
    institution = "Institution J",
    term = c("20013", "20021"), 
    program = c("FYE", "FYE"),
    cip6 = c("140102", "140102")
)
one_student
term_excerpt
prep_fye_mice(one_student, term_excerpt)

# Student D starts in History (540101), switches to FYE for two terms, 
# followed by ME (141901). The FYE proxy is ME (141901).
one_student <- data.table(mcid = "D", race = "unknown", sex = "unknown")
term_excerpt <- data.table(mcid = "D", 
                           institution = "Institution J",
                           term = c("20013", "20021", "20023", "20031"), 
                           program = c("History", "FYE", "FYE", "ME"),
                           cip6 = c("540101", "140102", "140102", "141901"))
one_student
term_excerpt
prep_fye_mice(one_student, term_excerpt)

# Student E is similar to Student D except they switched out of FYE 
# (140102) to History (540101), then returned to FYE followed by ME
# (141901). The FYE proxy is ME (141901).
one_student <- data.table(mcid = "E", race = "unknown", sex = "unknown")
term_excerpt <- data.table(
    mcid = "E", 
    institution = "Institution J",
    term = c("20011", "20013", "20021", "20023", "20031"), 
    program = c("FYE", "History", "History", "FYE", "ME"),
    cip6 = c("140102", "540101", "540101", "140102", "141901")
)
one_student
term_excerpt
prep_fye_mice(one_student, term_excerpt)

# Student F enrolls in FYE at Institution C that uses an alternate CIP
# code (140101), followed by ME (141901), so we have to specify a 
# non-default FYE CIP code. The FYE proxy is ME (141901).
one_student <- data.table(mcid = "F", race = "unknown", sex = "unknown")
term_excerpt <- data.table(
    mcid = "F", 
    institution = "Univ ABC",
    term = c("20013", "20021"), 
    program = c("FYE", "ME"),
    cip6 = c("140101", "141901")
)
fye_cip <- data.table(institution = "Univ ABC", fye_cip6 = "140101")
one_student
term_excerpt
fye_cip
prep_fye_mice(one_student, term_excerpt, fye_cip)

# Using datasets with multiple students
prep_fye_mice(toy_student, toy_term)[order(proxy)]
}
