% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_level.R
\name{qualification_level}
\alias{qualification_level}
\title{Categorize qualification level by term}
\usage{
qualification_level(dframe, midf_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \code{{mcid}} and one of
\verb{\{term, term_course, term_degree\}.}}

\item{midf_table}{\code{degree} data frame with required variables
\verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added or replace existing columns of the same name (if
any). Other columns are not modified. The following variables are added:
\itemize{
\item \code{bacc}    Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA}.
Joined from the \code{term_degree} variable in \code{midf_table.}
\item \code{qual_level}    Character, indicating that a term belongs
to one of two groups: "undergrad" terms are those leading up to
and including the term in which a student completes their first
degree; and "post-bacc" (post-baccalaureate) for all terms after the
first degree.
}
\item Groups and keys are not preserved.
}
}
\description{
Categorize the qualification level towards which a student is working in
each term. Two levels are used: “undergrad” for terms before a student's
first degree and “post-bacc” (post-baccalaureate) for terms after the first
degree. Added columns support the findings. Post-baccalaureate terms are
typically excluded from the \verb{term, course,} and \code{degree} data tables.
}
\examples{
# select min required columns
term <- toy_term[, .(mcid, term)]
course <- toy_course[, .(mcid, term_course)]
degree <- toy_degree[, .(mcid, term_degree)]

# labeling terms by group: undergrad & grad
term <- qualification_level(term, midf_table = degree)
course <- qualification_level(course, midf_table = degree)
degree <- qualification_level(degree, midf_table = degree)

# results
term[order(-qual_level)]
term[, .N, by = "qual_level"][order(-N)]

course[order(-qual_level)]
course[, .N, by = "qual_level"][order(-N)]

degree[order(-qual_level)]
degree[, .N, by = "qual_level"][order(-N)]
}
