% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_sex_join.R
\name{race_sex_join}
\alias{race_sex_join}
\title{Join student demographics to a data frame}
\usage{
race_sex_join(.data)
}
\arguments{
\item{.data}{A data frame of student attributes that includes the variable
\code{id}}
}
\value{
The incoming data frame plus two new columns \code{race} and
\code{sex}, joined by student ID
}
\description{
Add variables \code{race} and \code{sex} from the \code{midfieldstudents}
dataset to a data frame, using \code{id} as the join-by variable.
}
\details{
If sex or race variables are already present, they are returned as-is (not overwritten).
}
