% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcb.R
\name{rcb}
\alias{rcb}
\title{Hex color codes from RColorBrewer}
\usage{
rcb(pattern)
}
\arguments{
\item{pattern}{Character variable of a custom color name of the form
"level_hue",  with 4 levels (dark, mid, light, pale) and 5 hues (Br, BG, PR,
Gn, Gray or Grey). For example, "dark_Br", "mid_BG", "light_Gn", or
"pale_Gray."}
}
\value{
Hexadecimal color code in character form "#nnnnnn"
}
\description{
Accesses the \code{midfieldr::rcb_colors} dataset and returns a hex color
code.
}
\examples{
rcb("dark_Br")
}
\seealso{
Other graph_helper: 
\code{\link{rcb_colors}},
\code{\link{scale_x_log10_expon}()}
}
\concept{graph_helper}
