% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_ever}
\alias{rep_ever}
\title{Case student IDs}
\format{
character vector
}
\usage{
rep_ever
}
\description{
Representative case study. Character vector of 12,409 student IDs
ever enrolled in Civil, Electrical, Industrial, and Mechanical
Engineering programs (including predicted starting programs for
FYE students) in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}.
}
\seealso{
Other case-study-data: 
\code{\link{rep_cip}},
\code{\link{rep_group}},
\code{\link{rep_stickiness_mw}},
\code{\link{rep_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
