% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_group}
\alias{rep_group}
\title{Case program group}
\format{
\code{data.table} with 12 rows and 2 columns keyed by the
6-digit CIP code. The variables are:
\describe{
  \item{cip6}{6-digit program code}
  \item{program}{program name assigned for grouping, summarizing,
  and joining}
}
}
\usage{
rep_group
}
\description{
Representative case study. Six-digit CIP codes and custom program names
for Civil, Electrical, Industrial, and Mechanical Engineering programs in
\href{https://midfieldr.github.io/midfielddata/}{midfielddata}.
}
\seealso{
Other example data: 
\code{\link{rep_cip}},
\code{\link{rep_ever}},
\code{\link{rep_stickiness_mw}},
\code{\link{rep_stickiness}}
}
\concept{example data}
\keyword{datasets}
