% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rep_start}
\alias{rep_start}
\title{Starting programs data set (representative case study)}
\format{
\code{data.table} with 10135 rows and 2 columns keyed by student
ID. The variables are:
\describe{
   \item{id}{student ID}
   \item{start}{starting program}
}
}
\usage{
rep_start
}
\description{
Starting programs for students in the engineering case study. Includes
degree-granting programs that substitute for first-year-engineering (FYE)
programs when needed for a persistence metric, e.g., graduation rate.
}
\details{
The starting program for a non-FYE student is their matriculation program.
The starting programs for FYE students are a subset of the data set
\code{fye_start} that includes a predicted starting program for every
FYE student in midfielddata.
}
\seealso{
Other example_data: 
\code{\link{rep_cip}},
\code{\link{rep_group}},
\code{\link{rep_id}},
\code{\link{rep_stickiness_mw}},
\code{\link{rep_stickiness}}
}
\concept{example_data}
\keyword{datasets}
