% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_basic_cols.R
\name{select_basic_cols}
\alias{select_basic_cols}
\title{Choose columns of student records}
\usage{
select_basic_cols(dframe)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) equivalent to or derived from one of the MIDFIELD
data tables: \verb{\{student, term, course, degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Rows are not modified.
\item Columns are a subset of the input, appearing in the same order.
}
}
\description{
Subset a MIDFIELD data table to retain the variables required by one or
more midfieldr functions. Variables that constitute the key or composite
key for a table are retained as well. A convenience function to reduce
the number of columns displayed.
}
\details{
Functions in midfieldr with a MIDFIELD dataset argument---such as
\verb{student, term,} etc.---typically require only a few of the columns
available in the table. Depending on which table is input, the following
columns are returned if present:
\itemize{
\item \verb{student: \{mcid, race, sex\}}
\item \verb{term: \{mcid, term, cip6, institution, level\}}
\item \verb{course: \{mcid, term_course, abbrev, number\}}
\item \verb{degree: \{mcid, term_degree, cip6\}}
\item Combination of the above if \code{dframe} contains columns from multiple tables.
}
}
\examples{
# Basic usage
select_basic_cols(toy_student[1:5])
select_basic_cols(toy_term[1:5])
select_basic_cols(toy_course[1:5])
select_basic_cols(toy_degree[1:5])

# If the input is not strictly one of the four MIDFIELD data
# tables, all possible required columns are returned.
x <- toy_student[toy_degree, on = c("mcid")][1:5]
select_basic_cols(x)

# Required columns can only be returned if present, 
# e.g., consider the result for a full table:
select_basic_cols(toy_term)

# Compared to the result for a subset of the same table:
y <- toy_term[, .(mcid, term, cip6, hours_term, gpa_term)]
select_basic_cols(y)
}
