% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_basic_cols.R
\name{select_basic_cols}
\alias{select_basic_cols}
\title{Choose columns of student records}
\usage{
select_basic_cols(dframe, col_pattern = NULL, ..., type = NULL)
}
\arguments{
\item{dframe}{Data frame of \emph{student}, \emph{term}, \emph{course}, or \emph{degree} student-level records.}

\item{col_pattern}{Character vector containing strings or regular
expressions to be matched or partially matched to the column
names of \code{dframe}.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{type}{Character identifying the record type. Possible values are "s",
"t", "c", "d", "a", or NULL (default). See Details.}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved.
\item Rows are not modified.
\item Columns are a subset of the input, appearing in the same order.
\item Groups and keys are not preserved.
}
}
\description{
Subset any of the four MIDFIELD data tables \verb{\{student, term, course, degree\}}
by selecting from that data frame the columns required by other midfieldr
functions.
}
\details{
A convenience function to reduce the dimensions of a MIDFIELD
data table by selecting only those columns required by other midfieldr
functions or that are required to form a composite key. Particularly
useful in interactive sessions when viewing the data tables at various
stages of an analysis.

Several midfieldr functions require input data frames containing
specific variables (column names) such as \code{mcid} or \code{cip6}. In addition,
the MIDFIELD data tables have specific variables that act as keys
or composite keys to the information in that table. If the \code{type} argument
is NULL (default), one of the following codes is assigned to return the
column names indicated (if present):
\itemize{
\item \code{type = "s"} (student) looks for \verb{\{mcid, race, sex\}}
\item \code{type = "t"} (term) looks for \verb{\{mcid, term, cip6, institution, level\}}
\item \code{type = "c"} (course) looks for  \verb{\{mcid, term_course, abbrev, number\}}
\item \code{type = "d"} (degree) looks for  \verb{\{mcid, term_degree, cip6\}}
\item \code{type = "a"} looks for all the above columns
}

Specifying the type \verb{\{s, t, c, d, a\}} manually in the argument overrides
the automatic selection. Additional column names can be included by using
the \code{col_pattern} argument. In all cases, unmatched search strings are
silently ignored.
}
\examples{
# Basic usage
select_basic_cols(toy_student[1:5])
select_basic_cols(toy_term[1:5])
select_basic_cols(toy_course[1:5])
select_basic_cols(toy_degree[1:5])

# With col_pattern for additional columns
DT <- toy_student[141:146]
select_basic_cols(DT, col_pattern = c("transfer", "hours_tranfer"))

# Using regular expressions
these_IDs <- DT$mcid
DT <- toy_term[mcid \%chin\% these_IDs]
select_basic_cols(DT, col_pattern = c("^gpa"))
}
