% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_basic_cols.R
\name{select_basic_cols}
\alias{select_basic_cols}
\title{Select basic columns of student-level records}
\usage{
select_basic_cols(dframe, ..., patternv = NULL)
}
\arguments{
\item{dframe}{Data frame from which columns are selected.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{patternv}{Character vector of additional search terms. Can include
regular expressions.}
}
\value{
A subset of \code{dframe} with the following
properties: rows are preserved; columns are selected that exactly
match the built-in set of column names or that partially match any
search strings in \code{patternv}. Grouping structures are not necessarily
preserved.

An attempt is made to return data frames of the same
class as the input, so a data.table input should produce a data.table
output, a tibble input should produce a tibble output, etc. Grouped
tibbles, however, are returned as data.frames. To help ensure a tibble
output, try ungrouping a tibble before using it as an input argument.
}
\description{
Subset a data frame, selecting columns by matching a vector of character
strings. A convenience function to reduce the dimensions of a MIDFIELD
data table by selecting only those columns required by other midfieldr
functions or that are required to form a composite key. Particularly
useful in interactive sessions when viewing the data tables at various
stages of an analysis.
}
\details{
Several midfieldr functions require that their input data frames contain
specific variables (column names) such as \code{mcid} or \code{cip6}. In addition,
the MIDFIELD data tables have specific variables that act as keys
or composite keys to the information in that table. All such are assembled
in a character vector that comprises the default set of column names
returned by \code{select_basic_cols()}. The default column set is
\code{c(mcid, institution, race, sex, term, term_course, term_degree, cip6, level, abbrev, number)}.

The column names of \code{dframe} are searched for exact matches to the default
set of names. Additional column names can be included by using the
\code{patternv} argument---a vector of search terms for matches or partial
matches to the patterns. Regular expressions are permitted. Column names
and search terms not found are silently ignored.
}
\examples{
# Basic usage
select_basic_cols(toy_student[1:5])
select_basic_cols(toy_course[1:5])
select_basic_cols(toy_term[1:5])
select_basic_cols(toy_degree[1:5])

# With patterns for additional columns
DT <- toy_student[141:146]
select_basic_cols(DT, patternv = c("transfer", "hours_tranfer"))

# Using regular expressions
these_IDs <- DT$mcid
DT <- toy_term[mcid \%chin\% these_IDs]
select_basic_cols(DT, patternv = c("^gpa"))
}
