% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_required.R
\name{select_required}
\alias{select_required}
\title{Select required midfieldr variables}
\usage{
select_required(dframe, ..., select = NULL)
}
\arguments{
\item{dframe}{Data frame from which columns are selected, typically 
\code{student}, \code{term}, or \code{degree}.}

\item{...}{Not used, force later arguments to be used by name.}

\item{select}{Optional character vector of column names to return,
default \code{c("mcid", "institution", "race", "sex", "^term", 
"cip6", "level")}.}
}
\value{
A \code{data.table} with the following properties:
\itemize{
    \item Rows are not modified. 
    \item Columns with names that match or partially match the values 
          in \code{select}.
    \item Grouping structures are not preserved.
}
}
\description{
Subset a data frame, selecting columns by matching or partially matching a 
vector of character strings. A convenience function to reduce the dimensions 
of a MIDFIELD data table at the start of a session by selecting only those 
columns typically required by other midfieldr functions. Particularly useful 
in interactive sessions when viewing the data tables at various stages of  
an analysis.
}
\details{
Several midfieldr functions are designed to operate on one or more of the 
MIDFIELD data tables, usually \code{student}, \code{term}, or \code{degree}. 
This family of functions requires only a small subset of available 
variables, e.g., \code{mcid}, \code{cip6}, or \code{term}. We have 
identified the names of all required variables and collected 
appropriate search strings in the default \code{select} vector. 

The column names of \code{dframe} are searched for matches or partial matches 
in \code{select} using \code{grep()}, thus search terms can include regular 
expressions. Variables with names that match or partially match the search 
terms are returned; all other columns are dropped. Rows are unaffected. 
Search terms not present are silently ignored. 

One could use this function to select columns from a non-MIDFIELD data frame, 
but with no benefit to the user---conventional column selection syntax is 
better suited to that task. Here, we specialize the column selection to 
serve midfieldr functions.
}
\concept{select_*}
