% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_unique_cols.R
\name{select_unique_cols}
\alias{select_unique_cols}
\title{Choose unique columns}
\usage{
select_unique_cols(dframe)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble)}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Rows are preserved.
\item Unique columns preserved. Redundant columns dropped.
}
}
\description{
Subset a data frame to retain unique columns. A variable in a data frame is
dropped if, after splitting its name at a period separator ("."), its name
and values are identical to those of another variable. Primarily used
internally in midfieldr functions that add columns to a data frame.
}
\details{
Several midfieldr functions add columns to a working data frame. The goal
of this function is to prevent overwriting existing columns in the data
frame that happen to have the same name as one of the added columns. At
the same time, if both the name and values of a new column duplicate
an existing column, the new column is redundant and can be dropped.

If the name of an existing column happens to match that of a new variable,
the new variable name is made unique by adding a suffix such as ".1", ".2",
etc. Before the final data frame is returned, all variable names are
temporarily split from their dot suffixes. Columns are dropped if their
values and dot-split name duplicate a previous column. Columns are not
dropped if their split names are unique nor if the row-wise values differ.
}
\examples{
# Construct a sample data frame
mcid <- paste0("mc_", c("01", "02", "03"))
term <- c("19911", "19912", "19913")
dframe <- data.frame(mcid, term)
dframe
 
# No effect if all columns are distinct
select_unique_cols(dframe)

# No effect if names are distinct even if values are the same
x <- dframe
x$case <- x$term
x
select_unique_cols(x)

# Suffix-variable remains if there is no root-variable
x$case <- NULL
x$temp.1 <- c("abc", "def", "ghi")
x
select_unique_cols(x)

# Suffix-variable remains if values different from root-variable
x$temp.1 <- NULL
x$case.1 <- c("abc", "def", "ghi")
x
select_unique_cols(x)

# Suffix-variable dropped if otherwise identical to root-variable
x$case.1 <- NULL
x$term.1 <- x$term
x
select_unique_cols(x)

# Multiple redundant columns are dropped 
x$term.1 <- x$term
x$term.abc <- x$term
x
select_unique_cols(x)

# If a midfieldr function introduces a column with a 
# suffix ".1", ".2", etc., this points to a potential 
# error. The two columns with the same root name 
# are expected to be identical.
x <- dframe
x$term.1 <- c("19922", "19912", "19913")
x
select_unique_cols(x)

# Names with other separators are treated as unique
x$term.1 <- NULL
x$term_1 <- x$term
x
select_unique_cols(x)

# In removing a redundant column, the leftmost is retained
x <- data.frame(mcid.1 = mcid, term, mcid)
x
select_unique_cols(x)
}
