% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_unique_cols.R
\name{select_unique_cols}
\alias{select_unique_cols}
\title{Choose unique columns}
\usage{
select_unique_cols(dframe)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble)}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Rows are preserved.
\item Unique columns preserved. Redundant columns dropped.
}
}
\description{
Subset a data frame to retain unique columns. A variable in a data frame is
dropped if, after splitting its name at a period separator ("."), its name
and values are identical to those of another variable. Primarily used
internally in midfieldr functions that add columns to a data frame.
}
\details{
Several midfieldr functions add columns to a working data frame. The goal
of this function is to prevent overwriting existing columns in the data
frame that happen to have the same name as one of the added columns. At
the same time, if both the name and values of a new column duplicate
an existing column, the new column is redundant and can be dropped.

If the name of an existing column happens to match that of a new variable,
the new variable name is made unique by adding a suffix such as ".1", ".2",
etc. Before the final data frame is returned, all variable names are
temporarily split from their dot suffixes. Columns are dropped if their
values and dot-split name duplicate a previous column. Columns are not
dropped if their split names are unique nor if the row-wise values differ.
}
\examples{
# Construct a sample data frame
mcid <- c("mc_01", "mc_02", "mc_03")
term <- c("19911", "19912", "19913")
x <- data.frame(mcid, term)

# In the following, column names "var" and "var.i" 
# represent identical names except for the suffix ".i"
 
# No effect if columns are unique
x
x <- select_unique_cols(x)
x

# Desired effect: drop new "var.i" if it duplicates existing "var"
x$term.1 <- x$term
x
x <- select_unique_cols(x)
x

# Retain "var.i" when "var" does not exist
x <- data.frame(mcid, term.1 = term)
x
x <- select_unique_cols(x)
x

# The leftmost of the redundant columns is retained
x$term <- term
x
x <- select_unique_cols(x)
x

# Redundancy is checked for the dot-separator only
x <- data.frame(mcid, term, term_1 = term, term.1 = term)
x
x <- select_unique_cols(x)
x

# When a midfieldr function adds a variable, e.g., `bacc_term`, and
# a subsequent operation adds it again, `bacc_term.1`, the second 
# addition is usually expected to be redundant and dropped.
x$term_1 <- NULL
x$bacc_term <- c("19951", "19951", "19951")
x$bacc_term.1 <- c("19951", "19951", "19951")
x
x <- select_unique_cols(x)
x

# However, if the suffixed column remains, then the two columns 
# have different values---likely indicating an error in the 
# operations leading up to this point. 
x$bacc_term.1 <- c("19953", "19951", "19951")
x
x <- select_unique_cols(x)
x
}
