% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_uniq}
\alias{sort_uniq}
\title{Extract unique elements and sort}
\usage{
sort_uniq(
  x,
  ...,
  incomparables = FALSE,
  MARGIN = 1,
  fromLast = FALSE,
  decreasing = FALSE,
  na.last = FALSE
)
}
\arguments{
\item{x}{Vector of values to be sorted with any duplicate
values removed.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{incomparables}{A vector of values. See \code{unique()}.}

\item{MARGIN}{An integer. The array margin to be held fixed. Passed
to \code{unique()}.}

\item{fromLast}{Logical. Indicates if duplication should be considered
from the last. Passed to \code{unique()}.}

\item{decreasing}{Logical. Should the sort be increasing or decreasing?
Passed to \code{sort()}.}

\item{na.last}{Logical. Position of NA values. Passed to \code{sort()}.}
}
\value{
A vector of unique values, sorted.
}
\description{
A strict version of \code{sort()} and \code{unique()} (without ...).
}
\examples{
# Character vector
x <- toy_student$race
sort_uniq(x)

# Numeric vector
x <- toy_term$hours_term_attempt
sort_uniq(x)

}
