% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starter_filter.R
\name{starter_filter}
\alias{starter_filter}
\title{Filter student data for starting students}
\usage{
starter_filter(series, ..., reference = NULL, id = "id",
  cip6 = "cip6")
}
\arguments{
\item{series}{atomic character vector of 6-digit CIP codes specifying the
programs to filter by}

\item{...}{not used for values, forces later arguments to bind by name}

\item{reference}{a reference data frame from which student IDs, academic
terms, and CIP codes are obtained, default \code{midfieldterms}}

\item{id}{character column name of the ID variable in \code{reference}}

\item{cip6}{character column name of the CIP code variable in
\code{reference}}
}
\value{
Data frame with character variables for student ID and CIP of the
starting program
}
\description{
Filter entering student data to find all students starting in specified
programs.
}
\details{
To use this function, the \code{midfielddata} package must be installed
to provide \code{midfieldstudents}, the default reference data set.

The \code{series} argument is an atomic character vector of 6-digit CIP
codes used to filter the reference data set and extract student IDs.

The function returns a subset of \code{reference} with the student ID and
starting major CIP. Other variables in \code{reference} are quietly dropped.

Optional arguments. An alternate reference data set can be assigned via
the \code{reference} argument. The alternate must have variables for student
ID and starting CIP.
}
\examples{
placeholder <- 2 + 3
}
\seealso{
\code{\link[midfieldr]{cip_filter}} for obtaining 6-digit CIP codes
}
