% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stem}
\alias{stem}
\title{CIP codes for all STEM}
\format{A character vector.}
\source{
CIP 2010, US National Center for Education Statistics,
  \url{https://nces.ed.gov/ipeds/cipcode}, and the US National Science
   Foundation (NSF) NSF
  \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}.
}
\usage{
stem
}
\description{
Classification of Instructional Programs (CIP) 6-digit codes for all
Science, Technology, Engineering, and Mathematics (STEM)
programs. Concatenates all the midfieldr named series for STEM programs:
\code{bio_sci}, \code{engr}, \code{math_stat}, \code{other_stem}, and
\code{phys_sci}.
}
\examples{
head(stem, n = 10L)

print(cip_filter(series = stem))
}
\keyword{datasets}
