% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_observations}
\alias{study_observations}
\title{Case-study observations}
\format{
\code{data.table} with 8919 rows and 5 columns.
The variables are:
\describe{
\item{\code{mcid}}{Character. Anonymized student identifier
that connects the four data tables, e.g., "MCID3111142897."}
\item{\code{race}}{Character. Race/ethnicity as self-reported
by the student, e.g., "Asian", "Black", "Hispanic", etc.}
\item{\code{sex}}{Character. Sex as self-reported by the student,
possible values are "Female", "Male", and "Unknown."}
\item{\code{program}}{Character. Academic program label.}
\item{\code{bloc}}{Character. A group of student-level data dealt
with as a unit, e.g., starters, students ever-enrolled,
graduates, etc.}
}
}
\usage{
study_observations
}
\description{
Data table of post-processed observations of students ever enrolled in, and
students graduating from, the four programs of the case study. Keyed by
student ID. Provided for the convenience of vignette users.
}
\details{
Starting with the case-study starting pool of students ever enrolled in the
four programs of the study (Civil, Electrical, Industrial/Systems, and
Mechanical Engineering), we filtered the data for data sufficiency, degree
seeking, program, and timely completion.

A data frame of "ever enrolled" and a data frame of "timely graduates" were
bound using shared column names and are distinguished in the \code{bloc} variable.
This data structure facilitates grouping and summarizing by race, sex,
program, and group.
}
\seealso{
Other case-study-data:
\code{\link{baseline_mcid}},
\code{\link{study_programs}},
\code{\link{study_results}}
}
\concept{case-study-data}
\keyword{datasets}
