% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_programs}
\alias{study_programs}
\title{Case-study program codes and names}
\format{
\code{data.table} with 12 rows and 2 columns keyed by the
6-digit CIP code. The variables are:
\describe{
  \item{cip6}{6-digit program code}
  \item{program}{program name assigned for grouping, summarizing,
  and joining}
}
}
\usage{
study_programs
}
\description{
Data frame of 6-digit CIP codes and custom program names for Civil,
Electrical, Industrial, and Mechanical Engineering programs
from midfielddata. Used in the case study developed in the vignettes.
}
\seealso{
Other case-study-data: 
\code{\link{study_enrollees}},
\code{\link{study_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
