% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_programs}
\alias{study_programs}
\title{Case-study programs}
\format{
\code{data.table} with 12 rows and 2 columns keyed by the
6-digit CIP code. The variables are:
\describe{
  \item{cip6}{Character 6-digit code representing "specific instructional
       programs" (US National Center for Education Statistics)}
  \item{program}{Program name assigned for grouping, summarizing,
  and joining}
}
}
\usage{
study_programs
}
\description{
Data frame of 6-digit CIP codes and custom program names for Civil,
Electrical, Industrial, and Mechanical Engineering programs
from midfielddata. Used in the case study developed in the vignettes.
}
\details{
Some program names have been shortened for convenience when creating
graphs.
}
\seealso{
Other case-study-data: 
\code{\link{study_grad_rate}},
\code{\link{study_stickiness}},
\code{\link{study_students}}
}
\concept{case-study-data}
\keyword{datasets}
