% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_programs}
\alias{study_programs}
\title{Case-study program labels and codes}
\format{
\code{data.table} with 15 rows and 2 columns.
The variables are:
\describe{
\item{\code{cip6}}{Character, 6-digit CIP code of program in which
a student is enrolled in a term.}
\item{\code{program}}{Character, abbreviated labels for four engineering
programs. Values are "CE" (Civil Engineering), "EE" (Electrical
Engineering), "ISE" (Industrial/Systems Engineering), and  "ME" (Mechanical
Engineering).}
}
}
\usage{
study_programs
}
\description{
Data table of program CIP codes and labels of the four programs of the case
study. Keyed by 6-digit CIPs. Provided for the convenience of vignette users.
}
\details{
Starting with the midfieldr \code{cip} data set, we extracted the CIPs of the four
programs of the case study and assigned them a custom label to be used for
grouping and summarizing.
}
\seealso{
Other case-study-data: 
\code{\link{baseline_mcid}},
\code{\link{study_observations}},
\code{\link{study_results}}
}
\concept{case-study-data}
\keyword{datasets}
