% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_stickiness}
\alias{study_stickiness}
\title{Case-study stickiness results}
\format{
\code{data.table} with 33 rows and 6 columns keyed by
program, race/ethnicity, and sex.
\describe{

  \item{program}{Character, program name assigned for grouping and
  summarizing and for display in graphs and tables.}

 \item{race}{Character, race/ethnicity as self-reported by the student,
 e.g., Asian, Black, Hispanic/LatinX, etc.}

 \item{sex}{Character, sex as self-reported by the student, values are
 Female, Male.}

 \item{ever}{Numerical, number of students ever enrolled in the program,
 after accounting for data sufficiency and timely completion.}

 \item{grad}{Numerical, number of students completing a program, after
 accounting for data sufficiency and timely completion.}

 \item{stick}{Numerical, program stickiness, the ratio of the number of
 students completing a program to the number of students ever enrolled
 in the program.}

}
}
\usage{
study_stickiness
}
\description{
An intermediate result provided for the convenience of vignette users. 
Longitudinal stickiness results for Civil, Electrical, Industrial, and
Mechanical Engineering programs from midfielddata. Results are keyed by
program, race/ethnicity, and sex.
}
\details{
Longitudinal stickiness is the ratio of the number of students graduating
from a program to the number of students ever enrolled in the program over
the time span of available data.
}
\seealso{
Other case-study-data: 
\code{\link{study_grad_rate}},
\code{\link{study_program}},
\code{\link{study_stickiness_observ}},
\code{\link{study_student}}
}
\concept{case-study-data}
\keyword{datasets}
