% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_students}
\alias{study_students}
\title{Case-study student IDs and CIP codes}
\format{
\code{data.table} with 10,764 rows and 2 columns keyed by
student ID. The variables are:
\describe{
  \item{id}{student ID}
  \item{cip6}{6-digit program code}
}
}
\usage{
study_students
}
\description{
Data frame of student IDs and program codes of students ever enrolled
in the Civil, Electrical, Industrial, and Mechanical Engineering programs
from midfielddata.
}
\details{
No subsetting for other criteria such as timely completion or fair record length
have been performed. Also excluded are the predicted starting majors of
first-year-engineering (FYE) students.
}
\seealso{
Other case-study-data: 
\code{\link{study_programs}},
\code{\link{study_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
