% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_students}
\alias{study_students}
\title{Case-study students}
\format{
\code{data.table} with 8219 rows and 2 columns keyed by
student ID. The variables are:
\describe{
  \item{mcid}{Character, anonymized student identifier}
  \item{cip6}{Character 6-digit code representing "specific instructional
       programs" (US National Center for Education Statistics)}
}
}
\usage{
study_students
}
\description{
Data frame of student IDs and program codes of students ever enrolled
in the Civil, Electrical, Industrial, and Mechanical Engineering programs
from midfielddata.
}
\details{
Data have been subset for data sufficiency. Does not include predicted
starting programs for FYE students.
}
\seealso{
Other case-study-data: 
\code{\link{study_grad_rate}},
\code{\link{study_programs}},
\code{\link{study_stickiness}}
}
\concept{case-study-data}
\keyword{datasets}
