% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_stickiness.R
\name{tally_stickiness}
\alias{tally_stickiness}
\title{Compute stickiness}
\usage{
tally_stickiness(ever, grad)
}
\arguments{
\item{ever}{A data frame with at least one categorical variable and a
numeric variable named \code{ever}}

\item{grad}{A data frame with at least one categorical variable and a
numeric variable named \code{grad}}
}
\value{
A data frame with all the original columns of the two input
arguments plus a stickiness column.

The two input data frames are joined by the set of variables they have in
common.
}
\description{
Stickiness is the ratio of the number of students graduating from a program
to the number of students ever enrolled in a program.
}
\details{
The two input arguments must have at least one categorical variable in
common for joining.
}
