% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_stickiness.R
\name{tally_stickiness}
\alias{tally_stickiness}
\title{Tally the "stickiness" of a program}
\usage{
tally_stickiness(grad, ever, type = NULL, grad_min = NULL)
}
\arguments{
\item{grad}{A data frame of unique students graduating in a set of programs. Gathered from the \code{degree} dataset using \code{gather_grad()}. The set of CIP codes should be the same as those used to create the \code{ever} data frame. The required variables in the \code{grad} data frame are GCIP, ETHNIC, and SEX.}

\item{ever}{A data frame of unique students ever enrolled in a set of programs. Gathered from the \code{term} dataset using \code{gather_ever()}. The set of CIP codes should be the same as those used to create the \code{grad} data frame. The required variables in the \code{ever} data frame are TCIP, ETHNIC, and SEX.}

\item{type}{An integer (2, 4, or 6) representing the CIP taxonomy level. The default is 4-digits.}

\item{grad_min}{The smallest number of graduates in a peer group consistent with protecting student confidentiality. The default is 10.}
}
\value{
A data frame with variables CIP, ETHNIC, SEX, GRAD (the aggregate number of graduates), EVER (the aggregate number of students ever enrolled), and STICK (the resulting stickiness in percent).

A stickiness of zero indicates that members of a peer group were enrolled in a program at one time but none graduated. A peer group missing from the results indicates that none were ever enrolled in that program.
}
\description{
Stickiness is a persistence metric: the ratio of the number of students graduating in a program to the number of students ever enrolled in the program. This function groups and summarizes numbers of students by peer group (ethnicity and sex) and program (determined by the specified group of CIP codes), and computes the resulting stickiness.
}
\details{
Tallying stickiness entails three steps: 1) group and summarize the number of graduates; 2) group and summarize the number of students ever enrolled in a program; 3) join the two summaries and compute stickiness.
}
