% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_addition.R
\name{term_addition}
\alias{term_addition}
\title{Add semesters to an encoded academic term}
\usage{
term_addition(data = NULL, term_col = NULL, add_col = NULL)
}
\arguments{
\item{data}{data frame}

\item{term_col}{character name of a numeric column of terms
encoded YYYYT}

\item{add_col}{character name of a numeric column of the whole number
of semesters to add to YYYYT}
}
\value{
\code{data.frame} with \code{tbl} or \code{data.table}
extensions preserved
}
\description{
Add a whole number of semesters to an academic term encoded \code{YYYYT}
}
\examples{
# placeholder

}
\seealso{
Other data_carpentry: 
\code{\link{completion_feasible}()},
\code{\link{label_programs}()},
\code{\link{order_multiway}()}
}
\concept{data_carpentry}
