% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timely_term.R
\name{timely_term}
\alias{timely_term}
\title{Estimate timely completion terms}
\usage{
timely_term(dframe, midfield_table = term, ..., sched_span = NULL, span = NULL)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble). Required variable: \code{{mcid}}.}

\item{midfield_table}{Data frame of \emph{term} student-level records.
Required variables: \verb{\{mcid, term, level\}}.}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{sched_span}{Integer scalar (default 4), the number of years an
institution officially schedules for completing a program.}

\item{span}{Integer scalar (default 6), number of years to define timely
completion, typically 4, 6, or 8 years (100\%, 150\%, 200\% respectively
of \code{sched_span}).}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved.
\item Rows are filtered for unique student IDs.
\item Variable \code{{mcid}} is retained. All other columns are dropped and the
following variables are added:
\itemize{
\item \code{term_i}   Initial term of a student's longitudinal record,
encoded \code{YYYYT}. Extracted from \code{midfield_table.}
\item \code{level_i}   Character. Student level (01 Freshman, 02 Sophomore,
etc.) in their initial term. Extracted from \code{midfield_table.}
\item \code{adj_span}   Numeric. Integer span of years for timely
completion adjusted for a student's initial level.
\item \code{timely_term}   Character. Latest term by which
program completion
would be considered timely for every student. Encoded \code{YYYYT.}
}
\item Groups and keys are not preserved.
}
}
\description{
Starting with a data frame of unique student IDs, estimate the term by
which a student's program completion would be considered timely and add
columns to the data frame that support the finding.
}
\details{
The latest term by which program completion would be considered
timely is the \emph{timely completion term,} typically 4, 6, or 8 years after
admission depending on the definition adopted in a particular study. By
"completion" we mean an undergraduate earning their first baccalaureate
degree or degrees.

Our heuristic assigns \code{span} number of years (default 6) to every
student. For students admitted at second-year level or higher, the span is
reduced by one year for each full year the student is assumed to have
completed. The adjusted span is added to their initial term to create the
\code{timely_term} values. These results are documented in the output.

Determining completion status requires output variable \code{{timely_term}};
determining data sufficiency requires output variables
\verb{\{term_i, timely_term\}.}
}
\examples{
term <- toy_term

# Start with a small population 
x <- toy_student[c(51:55, 346:350), .(mcid, sex)]
x

# Add timely term, unrelated variables (sex) are dropped
x <- timely_term(x, midfield_table = term)
x

# Existing column with same name as added column is replaced
x[, adj_span := 0L][]
timely_term(x, midfield_table = term)
}
