% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timely_term.R
\name{timely_term}
\alias{timely_term}
\title{Determine timely completion terms}
\usage{
timely_term(dframe, midfield_table = term, ..., sched_span = NULL, span = NULL)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variable \verb{\{mcid\}.}}

\item{midfield_table}{\code{term} data frame with required variables
\verb{\{mcid, term, level\}.}}

\item{...}{Not used for passing values; forces subsequent arguments to be
referable only by name.}

\item{sched_span}{Integer scalar (default 4), the number of years an
institution officially schedules for completing a program.}

\item{span}{Integer scalar (default 6), number of years to define timely
completion, typically 4, 6, or 8 years (100\%, 150\%, 200\% respectively
of \code{sched_span}).}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys are
not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item Columns with names different from the new columns (named
below) are not modified; columns with matching names are replaced.
The new columns added are:
\itemize{
\item \code{term_i}   Character. Initial term of a student's longitudinal
record, encoded \code{YYYYT}. Extracted from \code{midfield_table.}
\item \code{level_i}   Character. Student level (01 Freshman, 02 Sophomore,
etc.) in their initial term. Extracted from \code{midfield_table.}
\item \code{adj_span}   Numeric. Integer span of years for timely
completion adjusted for a student's initial level.
\item \code{timely_term}   Character. Latest term by which program completion
would be considered timely. Encoded \code{YYYYT.}
}
}
}
\description{
Determine the \emph{timely completion term} for each student in a data frame
and add columns that support the findings.
}
\details{
Completing an academic program in a timely manner means that a student
completes the requirements for a degree within a set time span, typically
4, 6, or 8 years after admission depending on the definition adopted in a
particular study. The term at the end of that span is the
\emph{timely completion term.}

Our heuristic assigns a time span for timely completion to every student
(default is 6 academic years). For students admitted at second-year level
or higher, the span is reduced by one academic year for each full year the
student is assumed to have completed. The adjusted span is added to their
initial term at an institution to create the \code{timely_term} value for each
observation.
}
\examples{
term <- toy_term

# Start with a selected population. 
x <- toy_student[c(51:55, 346:350), .(mcid, sex)]
x

# Add timely term columns. Unrelated columns (sex) are unaffected.
x <- timely_term(x, midfield_table = term)
x

# Repeat. New columns silently replace existing columns of the same name.
y <- timely_term(x, midfield_table = term)
y
}
