% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_course}
\alias{toy_course}
\title{Course data for examples}
\format{
\code{data.table} with 4616 rows and 6 columns keyed by
student ID.
\describe{
 \item{mcid}{Character, anonymized student ID}
 \item{institution}{Character, de-identified institution name, e.g.,
      Institution A, Institution B, etc.}
 \item{term}{character, academic year and term, format YYYYT}
 \item{abbrev}{character, course alpha identifier, e.g. ENGR, MATH,
 ENGL, composite key (see id)}
 \item{number}{character, course numeric identifier, e.g. 101, 3429,
 composite key (see id)}
 \item{grade}{character, course grade, e.g., A+, A, A-, B+, I, NG, etc.}
}
}
\usage{
toy_course
}
\description{
A small subset by row and column of the MIDFIELD course table
for use in package examples.
}
\seealso{
Other toy-data: 
\code{\link{toy_degree}},
\code{\link{toy_student}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
