% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_course}
\alias{toy_course}
\title{Course data for examples}
\format{
\code{data.table} with 4616 rows and 6 columns keyed by student ID.

\describe{

\item{\code{mcid}}{Character, anonymized student ID.}

\item{\code{institution}}{Character, de-identified institution name, e.g.,
Institution A, Institution B, etc.}

\item{\code{term}}{Character, academic year and term, format YYYYT.}

\item{\code{abbrev}}{Character, course alphabetical identifier, e.g. ENGR, MATH,
ENGL.}

\item{\code{number}}{Character, course numeric identifier, e.g. 101, 3429.}

\item{\code{grade}}{Character, course grade, e.g., A+, A, A-, B+, I, NG, etc.}

}
}
\usage{
toy_course
}
\description{
Selected variables modeled on those in the \code{course} practice data for use in
package examples and articles. Sampled from an early version of the practice
data, the toy data are not a current practice data sample.
}
\seealso{
Other toy-data: 
\code{\link{toy_degree}},
\code{\link{toy_student}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
