% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_degree}
\alias{toy_degree}
\title{Small 'degree' dataset for examples}
\format{
Data frame with 193 rows and 4 columns (\code{data.table} class).
Composite key: \verb{\{mcid, term_degree, cip6\}.}
\describe{
\item{\code{mcid}}{Character. Anonymized student identifier
that connects the four data tables, e.g., "MCID3111142897."}
\item{\code{term_degree}}{Character. Academic year and term
in which a student completes their program, encoded \code{YYYYT.}}
\item{\code{cip6}}{Character. The 6-digit CIP code of the program
that the student completes in this term.}
\item{\code{institution}}{Character. The anonymized name of
the institution the student attended in a given term, e.g.,
"Institution A", "Institution B", etc.}
\item{\code{degree}}{Character. Type of degree awarded, e.g.,
"Bachelor of Arts in Geography", "Bachelor of Science in Finance," etc.}
}
}
\usage{
toy_degree
}
\description{
A subset of rows from the midfielddata \code{degree} table that comprises
those students from the \code{toy_student} dataset who complete a program.
A small dataset used in examples.
}
\seealso{
Other toy-data:
\code{\link{toy_course}},
\code{\link{toy_student}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
