% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_degree}
\alias{toy_degree}
\title{Degree data for examples}
\format{
\code{data.table} with 64 rows and 5 columns keyed by
student ID.
\describe{

 \item{mcid}{Character, anonymized student ID.}

 \item{institution}{Character, anonymized institution name, e.g.,
      Institution A, Institution B, etc.}

 \item{term}{Character, academic year and term in which
 a student completes their program, format YYYYT.}

 \item{cip6}{Character, 6-digit CIP code of program in which a student
 earns a degree.}

 \item{degree}{Character, type of degree awarded, e.g., Bachelor's
 Degree, Bachelor of Arts, Bachelor of Science, etc.}

}
}
\usage{
toy_degree
}
\description{
A small subset by row and column of the MIDFIELD degree table
for use in package examples.
}
\seealso{
Other toy-data: 
\code{\link{toy_course}},
\code{\link{toy_student}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
