% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_degree}
\alias{toy_degree}
\title{Degree data for examples}
\format{
\code{data.table} with 65 rows and 4 columns keyed by student ID.

\describe{

\item{\code{mcid}}{Character, anonymized student ID.}

\item{\code{institution}}{Character, anonymized institution name, e.g.,
Institution A, Institution B, etc.}

\item{\code{term_degree}}{Character, academic year and term in which
a student completes their program, format YYYYT.}

\item{\code{cip6}}{Character, 6-digit CIP code of program in which a student
earns a degree.}

}
}
\usage{
toy_degree
}
\description{
Selected variables modeled on those in the \code{degree} practice data for use in
package examples and articles. Sampled from an early version of the practice
data, the toy data are not a current practice data sample.
}
\seealso{
Other toy-data: 
\code{\link{toy_course}},
\code{\link{toy_student}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
