% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_student}
\alias{toy_student}
\title{Student data for examples}
\format{
\code{data.table} with 100 rows and 6 columns keyed by
student ID.
\describe{

 \item{mcid}{Character, de-identified student ID.}

 \item{institution}{Character, de-identified institution name, e.g.,
  Institution A, Institution B, etc.}

 \item{transfer}{Character, stating whether the student is a First-Time in
 College students or a First-Time Transfer student.}

 \item{hours_transfer}{Numeric, transfer hours accepted at the
 institution.}

 \item{race}{Character, race/ethnicity as self-reported by the student,
 e.g., Asian, Black, Latine, etc.}

 \item{sex}{character, sex as self-reported by the student, possible
 values are Female, Male, and Unknown.}

}
}
\usage{
toy_student
}
\description{
A small subset by row and column of the MIDFIELD student table
for use in package examples.
}
\seealso{
Other toy-data: 
\code{\link{toy_course}},
\code{\link{toy_degree}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
