% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_student}
\alias{toy_student}
\title{Student data for examples}
\format{
\code{data.table} with 99 rows and 4 columns keyed by student ID:
\describe{
\item{\code{mcid}}{Character, de-identified student ID.}
\item{\code{institution}}{Character, de-identified institution
name, e.g., Institution A, Institution B, etc.}
\item{\code{race}}{Character, race/ethnicity as self-reported
by the student, e.g., Asian, Black, Latine, etc.}
\item{\code{sex}}{Character, sex as self-reported by the student,
possible values are Female, Male, and Unknown.}
}
}
\usage{
toy_student
}
\description{
Selected variables modeled on those in the \code{student} practice data for use in
package examples and articles. Sampled from an early version of the practice
data, the toy data are not a current practice data sample.
}
\seealso{
Other toy-data: 
\code{\link{toy_course}},
\code{\link{toy_degree}},
\code{\link{toy_term}}
}
\concept{toy-data}
\keyword{datasets}
