% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_term}
\alias{toy_term}
\title{Term data for examples}
\format{
\code{data.table} with 1095 rows and 13 columns keyed by student ID
and term.
\describe{
\item{\code{mcid}}{Character, de-identified student ID. Key column.}
\item{\code{institution}}{Character, de-identified institution
name, e.g., Institution A, Institution B, etc.}
\item{\code{term}}{Character, academic year and term, format YYYYT.
Key column.}
\item{\code{cip6}}{Character, 6-digit CIP code of program in
which a student is enrolled in a term, e.g., \code{090101},
\code{141201}, \code{260901}, \code{420101}, etc.}
\item{\code{level}}{Character, 01 Freshman, 02 Sophomore, etc.
The equivalent values in the current practice data are 01
First-Year, 02-Second Year, etc.}
\item{\code{standing}}{Character, academic standing, e.g.,
\verb{Good Standing}, \verb{Academic Warning}, etc.}
\item{\code{coop}}{Character, cooperative education term, possible
values are \code{Yes}, \code{No}.}
\item{\code{hours_term}}{Numeric, credit hours earned
in the term.}
\item{\code{hours_term_attempt}}{Numeric, credit
hours attempted in the term.}
\item{\code{hours_cumul}}{Numeric, cumulative credit hours
earned.}
\item{\code{hours_cumul_attempt}}{Numeric, cumulative
credit hours attempted.}
\item{\code{gpa_term}}{Numeric, term grade point average.}
\item{\code{gpa_cumul}}{Numeric, cumulative grade point average.}
}
}
\usage{
toy_term
}
\description{
Selected variables modeled on those in the \code{term} practice data for use in
package examples and articles. Sampled from an early version of the practice
data, the toy data are not a current practice data sample.
}
\seealso{
Other toy-data:
\code{\link{toy_course}},
\code{\link{toy_degree}},
\code{\link{toy_student}}
}
\concept{toy-data}
\keyword{datasets}
