% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_term}
\alias{toy_term}
\title{Term data for examples}
\format{
\code{data.table} with 100 rows and 6 columns keyed by
student ID.
\describe{
 \item{mcid}{Character, anonymized student identifier}
 \item{institution}{Character, anonymized institution name, e.g.,
      Institution A, Institution B, etc.}
 \item{term}{character, academic year and term, format YYYYT}
 \item{cip6}{character, 6-digit CIP code of program in which a student
 is enrolled in a term}
 \item{level}{character, 01 Freshman, 02 Sophomore, etc.}
 \item{hours_term}{numeric, credit hours earned in the term}
}
}
\usage{
toy_term
}
\description{
A small subset by row and column of the MIDFIELD term table
for use in package examples.
}
\seealso{
Other toy-data: 
\code{\link{toy_course}},
\code{\link{toy_degree}},
\code{\link{toy_student}}
}
\concept{toy-data}
\keyword{datasets}
