% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undergrad_term_id.R
\name{undergrad_term_id}
\alias{undergrad_term_id}
\title{Identify undergraduate terms}
\usage{
undergrad_term_id(dframe, midf_table = degree)
}
\arguments{
\item{dframe}{Data frame or data frame extension (e.g., data.table or tibble) with required variables \code{{mcid}} and one of the
following: \verb{\{term, term_course, term_degree\}.}}

\item{midf_table}{\code{degree} data frame with required variables
\verb{\{mcid, term_degree\}.}}
}
\value{
Data frame with the following properties:
\itemize{
\item Data frame class is preserved. Groups and keys
are not preserved.
\item Row order is preserved. Rows with \code{NA} values in any
of the required variables are removed. Duplicated rows are removed.
\item New columns are added unless they are redundant (see Details). The new variables are:
\itemize{
\item \code{bacc_term} Character. Term of a student's first
baccalaureate, encoded \code{YYYYT} or, if no degree recorded, \code{NA.}
Joined from the \code{term_degree} variable in \code{midf_table.}
\item \code{term_id} Character. Distinguish post-baccalaureate terms from
undergraduate terms. Possible values are "undergrad" and "post-bacc."
}
}
}
\description{
Distinguishes between undergraduate terms
(to retain) and post-baccalaureate terms (to exclude). Adds columns to the data frame that support the
finding.
}
\details{
Typically used in refining student records to obtain a baseline for further
analysis, thus the principal argument is the \verb{term, course,} or \code{degree}
data table with all columns. Can be applied however to any data frame
containing the required variables.

\emph{Redundant columns.} To prevent overwriting, a variable such as
\code{bacc_term}  is not added to the data frame if it duplicates
an existing variable. The test for redundancy is managed internally
by calling \code{rm_redundant_cols()} before the final data frame is
returned. For documentation, see \code{?rm_redundant_cols}.
}
\examples{
# For illustration only, choose a minimum set of columns
term <- toy_term[, .(mcid, term)]
course <- toy_course[, .(mcid, term_course)]
degree <- toy_degree[, .(mcid, term_degree)]

# Example result
term

# Add term ID columns
x <- undergrad_term_id(term, midf_table = degree)
x[order(-term_id)]

# No change if added columns duplicate existing
y <- undergrad_term_id(x, midf_table = degree)
check_equiv_frames(x, y)

# Filter to retain "undergrad" rows only
x[term_id == "undergrad"]

# Function is applied to all tables containing a term-value
term <- undergrad_term_id(term, midf_table = degree)
course <- undergrad_term_id(course, midf_table = degree)
degree <- undergrad_term_id(degree, midf_table = degree)
}
