% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{utils_graph}
\alias{scale_x_log10_expon}
\alias{theme_midfield}
\title{Graphics utility functions}
\usage{
scale_x_log10_expon(...)

theme_midfield(...)
}
\arguments{
\item{...}{Arguments that pass to the ggplot2 function}
}
\value{
Modifies an existing graph.
}
\description{
A set of customized layers for ggplot2 graphs.
}
\details{
\describe{
\item{\code{scale_x_log10_expon(...)}}{Applies the \code{scale_x_log10(...)}
 function from ggplot2 with the scale marked in exponential form powers of
 ten and with logarithmic minor grid lines.}
\item{\code{theme_midfield(...)}}{Applies \code{theme_minimal(...)} from
 gplot2 with replacements: all text is 10 pt; lines are light gray (#D9D9D9)
  and size 0.4.}
}
}
\seealso{
Other graph_helper: 
\code{\link{rcb_colors}}
}
\concept{graph_helper}
