% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_fill.R
\name{zero_fill}
\alias{zero_fill}
\title{Assign zero to missing combinations}
\usage{
zero_fill(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Specification of columns to expand.}
}
\value{
Data frame with new rows with count = 0 for missing combinations.
}
\description{
A wrapper around \code{tidyr::expand()}. Expands a data frame to include
all combinations of values, including those not found in the data. Replaces
NA in numeric columns with zero.
}
